Struct aws_sdk_rds::client::fluent_builders::ModifyDBInstance

pub struct ModifyDBInstance<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> ModifyDBInstance<C>

pub async fn send(
self
) -> Result<ModifyDbInstanceOutput, SdkError<ModifyDBInstanceError>> where
C: SmithyConnector,

pub fn db_instance_identifier(self, input: impl Into<String>) -> Self

The DB instance identifier. This value is stored as a lowercase string.

Constraints:

Must match the identifier of an existing DBInstance.

pub fn set_db_instance_identifier(self, input: Option<String>) -> Self

pub fn allocated_storage(self, input: i32) -> Self

The new amount of storage (in gibibytes) to allocate for the DB instance.

For MariaDB, MySQL, Oracle, and PostgreSQL, the value supplied must be at least 10% greater than the current value. Values that are not at least 10% greater than the existing value are rounded up so that they are 10% greater than the current value.

For the valid values for allocated storage for each engine, see CreateDBInstance.

pub fn set_allocated_storage(self, input: Option<i32>) -> Self

pub fn db_instance_class(self, input: impl Into<String>) -> Self

The new compute and memory capacity of the DB instance, for example, db.m4.large. Not all DB instance classes are available in all Amazon Web Services Regions, or for all database engines. For the full list of DB instance classes, and availability for your engine, see DB Instance Class in the Amazon RDS User Guide.

If you modify the DB instance class, an outage occurs during the change. The change is applied during the next maintenance window, unless ApplyImmediately is enabled for this request.

Default: Uses existing setting

pub fn set_db_instance_class(self, input: Option<String>) -> Self

pub fn db_subnet_group_name(self, input: impl Into<String>) -> Self

The new DB subnet group for the DB instance. You can use this parameter to move your DB instance to a different VPC. If your DB instance isn't in a VPC, you can also use this parameter to move your DB instance into a VPC. For more information, see Working with a DB instance in a VPC in the Amazon RDS User Guide.

Changing the subnet group causes an outage during the change. The change is applied during the next maintenance window, unless you enable ApplyImmediately.

Constraints: If supplied, must match the name of an existing DBSubnetGroup.

Example: mySubnetGroup

pub fn set_db_subnet_group_name(self, input: Option<String>) -> Self

pub fn db_security_groups(self, inp: impl Into<String>) -> Self

A list of DB security groups to authorize on this DB instance. Changing this setting doesn't result in an outage and the change is asynchronously applied as soon as possible.

Constraints:

If supplied, must match existing DBSecurityGroups.

pub fn set_db_security_groups(self, input: Option<Vec<String>>) -> Self

pub fn vpc_security_group_ids(self, inp: impl Into<String>) -> Self

A list of EC2 VPC security groups to authorize on this DB instance. This change is asynchronously applied as soon as possible.

Amazon Aurora

Not applicable. The associated list of EC2 VPC security groups is managed by the DB cluster. For more information, see ModifyDBCluster.

Constraints:

If supplied, must match existing VpcSecurityGroupIds.

pub fn set_vpc_security_group_ids(self, input: Option<Vec<String>>) -> Self

pub fn apply_immediately(self, input: bool) -> Self

A value that indicates whether the modifications in this request and any pending modifications are asynchronously applied as soon as possible, regardless of the PreferredMaintenanceWindow setting for the DB instance. By default, this parameter is disabled.

If this parameter is disabled, changes to the DB instance are applied during the next maintenance window. Some parameter changes can cause an outage and are applied on the next call to RebootDBInstance, or the next failure reboot. Review the table of parameters in Modifying a DB Instance in the Amazon RDS User Guide. to see the impact of enabling or disabling ApplyImmediately for each modified parameter and to determine when the changes are applied.

pub fn set_apply_immediately(self, input: Option<bool>) -> Self

pub fn master_user_password(self, input: impl Into<String>) -> Self

The new password for the master user. The password can include any printable ASCII character except "/", """, or "@".

Changing this parameter doesn't result in an outage and the change is asynchronously applied as soon as possible. Between the time of the request and the completion of the request, the MasterUserPassword element exists in the PendingModifiedValues element of the operation response.

Amazon Aurora

Not applicable. The password for the master user is managed by the DB cluster. For more information, see ModifyDBCluster.

Default: Uses existing setting

MariaDB

Constraints: Must contain from 8 to 41 characters.

Microsoft SQL Server

Constraints: Must contain from 8 to 128 characters.

MySQL

Constraints: Must contain from 8 to 41 characters.

Oracle

Constraints: Must contain from 8 to 30 characters.

PostgreSQL

Constraints: Must contain from 8 to 128 characters.

Amazon RDS API actions never return the password, so this action provides a way to regain access to a primary instance user if the password is lost. This includes restoring privileges that might have been accidentally revoked.

pub fn set_master_user_password(self, input: Option<String>) -> Self

pub fn db_parameter_group_name(self, input: impl Into<String>) -> Self

The name of the DB parameter group to apply to the DB instance. Changing this setting doesn't result in an outage. The parameter group name itself is changed immediately, but the actual parameter changes are not applied until you reboot the instance without failover. In this case, the DB instance isn't rebooted automatically and the parameter changes isn't applied during the next maintenance window.

Default: Uses existing setting

Constraints: The DB parameter group must be in the same DB parameter group family as this DB instance.

pub fn set_db_parameter_group_name(self, input: Option<String>) -> Self

pub fn backup_retention_period(self, input: i32) -> Self

The number of days to retain automated backups. Setting this parameter to a positive number enables backups. Setting this parameter to 0 disables automated backups.

Enabling and disabling backups can result in a brief I/O suspension that lasts from a few seconds to a few minutes, depending on the size and class of your DB instance.

These changes are applied during the next maintenance window unless the ApplyImmediately parameter is enabled for this request. If you change the parameter from one non-zero value to another non-zero value, the change is asynchronously applied as soon as possible.

Amazon Aurora

Not applicable. The retention period for automated backups is managed by the DB cluster. For more information, see ModifyDBCluster.

Default: Uses existing setting

Constraints:

Must be a value from 0 to 35

Can be specified for a MySQL read replica only if the source is running MySQL 5.6 or later

Can be specified for a PostgreSQL read replica only if the source is running PostgreSQL 9.3.5

Can't be set to 0 if the DB instance is a source to read replicas

pub fn set_backup_retention_period(self, input: Option<i32>) -> Self

pub fn preferred_backup_window(self, input: impl Into<String>) -> Self

The daily time range during which automated backups are created if automated backups are enabled, as determined by the BackupRetentionPeriod parameter. Changing this parameter doesn't result in an outage and the change is asynchronously applied as soon as possible. The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Web Services Region. For more information, see Backup window in the Amazon RDS User Guide.

Amazon Aurora

Not applicable. The daily time range for creating automated backups is managed by the DB cluster. For more information, see ModifyDBCluster.

Constraints:

Must be in the format hh24:mi-hh24:mi

Must be in Universal Time Coordinated (UTC)

Must not conflict with the preferred maintenance window

Must be at least 30 minutes

pub fn set_preferred_backup_window(self, input: Option<String>) -> Self

pub fn preferred_maintenance_window(self, input: impl Into<String>) -> Self

The weekly time range (in UTC) during which system maintenance can occur, which might result in an outage. Changing this parameter doesn't result in an outage, except in the following situation, and the change is asynchronously applied as soon as possible. If there are pending actions that cause a reboot, and the maintenance window is changed to include the current time, then changing this parameter will cause a reboot of the DB instance. If moving this window to the current time, there must be at least 30 minutes between the current time and end of the window to ensure pending changes are applied.

For more information, see Amazon RDS Maintenance Window in the Amazon RDS User Guide.

Default: Uses existing setting

Format: ddd:hh24:mi-ddd:hh24:mi

Valid Days: Mon | Tue | Wed | Thu | Fri | Sat | Sun

Constraints: Must be at least 30 minutes

pub fn set_preferred_maintenance_window(self, input: Option<String>) -> Self

pub fn multi_az(self, input: bool) -> Self

A value that indicates whether the DB instance is a Multi-AZ deployment. Changing this parameter doesn't result in an outage and the change is applied during the next maintenance window unless the ApplyImmediately parameter is enabled for this request.

pub fn set_multi_az(self, input: Option<bool>) -> Self

pub fn engine_version(self, input: impl Into<String>) -> Self

The version number of the database engine to upgrade to. Changing this parameter results in an outage and the change is applied during the next maintenance window unless the ApplyImmediately parameter is enabled for this request.

For major version upgrades, if a nondefault DB parameter group is currently in use, a new DB parameter group in the DB parameter group family for the new engine version must be specified. The new DB parameter group can be the default for that DB parameter group family.

If you specify only a major version, Amazon RDS will update the DB instance to the default minor version if the current minor version is lower. For information about valid engine versions, see CreateDBInstance, or call DescribeDBEngineVersions.

pub fn set_engine_version(self, input: Option<String>) -> Self

pub fn allow_major_version_upgrade(self, input: bool) -> Self

A value that indicates whether major version upgrades are allowed. Changing this parameter doesn't result in an outage and the change is asynchronously applied as soon as possible.

Constraints: Major version upgrades must be allowed when specifying a value for the EngineVersion parameter that is a different major version than the DB instance's current version.

pub fn set_allow_major_version_upgrade(self, input: Option<bool>) -> Self

pub fn auto_minor_version_upgrade(self, input: bool) -> Self

A value that indicates whether minor version upgrades are applied automatically to the DB instance during the maintenance window. Changing this parameter doesn't result in an outage except in the following case and the change is asynchronously applied as soon as possible. An outage results if this parameter is enabled during the maintenance window, and a newer minor version is available, and RDS has enabled auto patching for that engine version.

pub fn set_auto_minor_version_upgrade(self, input: Option<bool>) -> Self

pub fn license_model(self, input: impl Into<String>) -> Self

The license model for the DB instance.

Valid values: license-included | bring-your-own-license | general-public-license

pub fn set_license_model(self, input: Option<String>) -> Self

pub fn iops(self, input: i32) -> Self

The new Provisioned IOPS (I/O operations per second) value for the RDS instance.

Changing this setting doesn't result in an outage and the change is applied during the next maintenance window unless the ApplyImmediately parameter is enabled for this request. If you are migrating from Provisioned IOPS to standard storage, set this value to 0. The DB instance will require a reboot for the change in storage type to take effect.

If you choose to migrate your DB instance from using standard storage to using Provisioned IOPS, or from using Provisioned IOPS to using standard storage, the process can take time. The duration of the migration depends on several factors such as database load, storage size, storage type (standard or Provisioned IOPS), amount of IOPS provisioned (if any), and the number of prior scale storage operations. Typical migration times are under 24 hours, but the process can take up to several days in some cases. During the migration, the DB instance is available for use, but might experience performance degradation. While the migration takes place, nightly backups for the instance are suspended. No other Amazon RDS operations can take place for the instance, including modifying the instance, rebooting the instance, deleting the instance, creating a read replica for the instance, and creating a DB snapshot of the instance.

Constraints: For MariaDB, MySQL, Oracle, and PostgreSQL, the value supplied must be at least 10% greater than the current value. Values that are not at least 10% greater than the existing value are rounded up so that they are 10% greater than the current value.

Default: Uses existing setting

pub fn set_iops(self, input: Option<i32>) -> Self

pub fn option_group_name(self, input: impl Into<String>) -> Self

A value that indicates the DB instance should be associated with the specified option group. Changing this parameter doesn't result in an outage except in the following case and the change is applied during the next maintenance window unless the ApplyImmediately parameter is enabled for this request. If the parameter change results in an option group that enables OEM, this change can cause a brief (sub-second) period during which new connections are rejected but existing connections are not interrupted.

Permanent options, such as the TDE option for Oracle Advanced Security TDE, can't be removed from an option group, and that option group can't be removed from a DB instance once it is associated with a DB instance

pub fn set_option_group_name(self, input: Option<String>) -> Self

pub fn new_db_instance_identifier(self, input: impl Into<String>) -> Self

The new DB instance identifier for the DB instance when renaming a DB instance. When you change the DB instance identifier, an instance reboot occurs immediately if you enable ApplyImmediately, or will occur during the next maintenance window if you disable Apply Immediately. This value is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

The first character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

Example: mydbinstance

pub fn set_new_db_instance_identifier(self, input: Option<String>) -> Self

pub fn storage_type(self, input: impl Into<String>) -> Self

Specifies the storage type to be associated with the DB instance.

If you specify Provisioned IOPS (io1), you must also include a value for the Iops parameter.

If you choose to migrate your DB instance from using standard storage to using Provisioned IOPS, or from using Provisioned IOPS to using standard storage, the process can take time. The duration of the migration depends on several factors such as database load, storage size, storage type (standard or Provisioned IOPS), amount of IOPS provisioned (if any), and the number of prior scale storage operations. Typical migration times are under 24 hours, but the process can take up to several days in some cases. During the migration, the DB instance is available for use, but might experience performance degradation. While the migration takes place, nightly backups for the instance are suspended. No other Amazon RDS operations can take place for the instance, including modifying the instance, rebooting the instance, deleting the instance, creating a read replica for the instance, and creating a DB snapshot of the instance.

Valid values: standard | gp2 | io1

Default: io1 if the Iops parameter is specified, otherwise gp2

pub fn set_storage_type(self, input: Option<String>) -> Self

pub fn tde_credential_arn(self, input: impl Into<String>) -> Self

The ARN from the key store with which to associate the instance for TDE encryption.

pub fn set_tde_credential_arn(self, input: Option<String>) -> Self

pub fn tde_credential_password(self, input: impl Into<String>) -> Self

The password for the given ARN from the key store in order to access the device.

pub fn set_tde_credential_password(self, input: Option<String>) -> Self

pub fn ca_certificate_identifier(self, input: impl Into<String>) -> Self

Indicates the certificate that needs to be associated with the instance.

pub fn set_ca_certificate_identifier(self, input: Option<String>) -> Self

pub fn domain(self, input: impl Into<String>) -> Self

The Active Directory directory ID to move the DB instance to. Specify none to remove the instance from its current domain. The domain must be created prior to this operation. Currently, only MySQL, Microsoft SQL Server, Oracle, and PostgreSQL DB instances can be created in an Active Directory Domain.

For more information, see Kerberos Authentication in the Amazon RDS User Guide.

pub fn set_domain(self, input: Option<String>) -> Self

pub fn copy_tags_to_snapshot(self, input: bool) -> Self

A value that indicates whether to copy all tags from the DB instance to snapshots of the DB instance. By default, tags are not copied.

Amazon Aurora

Not applicable. Copying tags to snapshots is managed by the DB cluster. Setting this value for an Aurora DB instance has no effect on the DB cluster setting. For more information, see ModifyDBCluster.

pub fn set_copy_tags_to_snapshot(self, input: Option<bool>) -> Self

pub fn monitoring_interval(self, input: i32) -> Self

The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the DB instance. To disable collecting Enhanced Monitoring metrics, specify 0. The default is 0.

If MonitoringRoleArn is specified, then you must also set MonitoringInterval to a value other than 0.

Valid Values: 0, 1, 5, 10, 15, 30, 60

pub fn set_monitoring_interval(self, input: Option<i32>) -> Self

pub fn db_port_number(self, input: i32) -> Self

The port number on which the database accepts connections.

The value of the DBPortNumber parameter must not match any of the port values specified for options in the option group for the DB instance.

Your database will restart when you change the DBPortNumber value regardless of the value of the ApplyImmediately parameter.

MySQL

Default: 3306

Valid values: 1150-65535

MariaDB

Default: 3306

Valid values: 1150-65535

PostgreSQL

Default: 5432

Valid values: 1150-65535

Type: Integer

Oracle

Default: 1521

Valid values: 1150-65535

SQL Server

Default: 1433

Valid values: 1150-65535 except 1234, 1434, 3260, 3343, 3389, 47001, and 49152-49156.

Amazon Aurora

Default: 3306

Valid values: 1150-65535

pub fn set_db_port_number(self, input: Option<i32>) -> Self

pub fn publicly_accessible(self, input: bool) -> Self

A value that indicates whether the DB instance is publicly accessible.

When the DB instance is publicly accessible, its DNS endpoint resolves to the private IP address from within the DB instance's VPC, and to the public IP address from outside of the DB instance's VPC. Access to the DB instance is ultimately controlled by the security group it uses, and that public access is not permitted if the security group assigned to the DB instance doesn't permit it.

When the DB instance isn't publicly accessible, it is an internal DB instance with a DNS name that resolves to a private IP address.

PubliclyAccessible only applies to DB instances in a VPC. The DB instance must be part of a public subnet and PubliclyAccessible must be enabled for it to be publicly accessible.

Changes to the PubliclyAccessible parameter are applied immediately regardless of the value of the ApplyImmediately parameter.

pub fn set_publicly_accessible(self, input: Option<bool>) -> Self

pub fn monitoring_role_arn(self, input: impl Into<String>) -> Self

The ARN for the IAM role that permits RDS to send enhanced monitoring metrics to Amazon CloudWatch Logs. For example, arn:aws:iam:123456789012:role/emaccess. For information on creating a monitoring role, go to To create an IAM role for Amazon RDS Enhanced Monitoring in the Amazon RDS User Guide.

If MonitoringInterval is set to a value other than 0, then you must supply a MonitoringRoleArn value.

pub fn set_monitoring_role_arn(self, input: Option<String>) -> Self

pub fn domain_iam_role_name(self, input: impl Into<String>) -> Self

The name of the IAM role to use when making API calls to the Directory Service.

pub fn set_domain_iam_role_name(self, input: Option<String>) -> Self

pub fn promotion_tier(self, input: i32) -> Self

A value that specifies the order in which an Aurora Replica is promoted to the primary instance after a failure of the existing primary instance. For more information, see Fault Tolerance for an Aurora DB Cluster in the Amazon Aurora User Guide.

Default: 1

Valid Values: 0 - 15

pub fn set_promotion_tier(self, input: Option<i32>) -> Self

pub fn enable_iam_database_authentication(self, input: bool) -> Self

A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

This setting doesn't apply to Amazon Aurora. Mapping Amazon Web Services IAM accounts to database accounts is managed by the DB cluster.

For more information about IAM database authentication, see IAM Database Authentication for MySQL and PostgreSQL in the Amazon RDS User Guide.

pub fn set_enable_iam_database_authentication(self, input: Option<bool>) -> Self

pub fn enable_performance_insights(self, input: bool) -> Self

A value that indicates whether to enable Performance Insights for the DB instance.

For more information, see Using Amazon Performance Insights in the Amazon Relational Database Service User Guide.

pub fn set_enable_performance_insights(self, input: Option<bool>) -> Self

pub fn performance_insights_kms_key_id(self, input: impl Into<String>) -> Self

The Amazon Web Services KMS key identifier for encryption of Performance Insights data.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK).

If you do not specify a value for PerformanceInsightsKMSKeyId, then Amazon RDS uses your default CMK. There is a default CMK for your Amazon Web Services account. Your Amazon Web Services account has a different default CMK for each Amazon Web Services Region.

pub fn set_performance_insights_kms_key_id(self, input: Option<String>) -> Self

pub fn performance_insights_retention_period(self, input: i32) -> Self

The amount of time, in days, to retain Performance Insights data. Valid values are 7 or 731 (2 years).

pub fn set_performance_insights_retention_period(
self,
input: Option<i32>
) -> Self

pub fn cloudwatch_logs_export_configuration(
self,
input: CloudwatchLogsExportConfiguration
) -> Self

The configuration setting for the log types to be enabled for export to CloudWatch Logs for a specific DB instance.

A change to the CloudwatchLogsExportConfiguration parameter is always applied to the DB instance immediately. Therefore, the ApplyImmediately parameter has no effect.

pub fn set_cloudwatch_logs_export_configuration(
self,
input: Option<CloudwatchLogsExportConfiguration>
) -> Self

pub fn processor_features(self, inp: impl Into<ProcessorFeature>) -> Self

The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

pub fn set_processor_features(
self,
input: Option<Vec<ProcessorFeature>>
) -> Self

pub fn use_default_processor_features(self, input: bool) -> Self

A value that indicates whether the DB instance class of the DB instance uses its default processor features.

pub fn set_use_default_processor_features(self, input: Option<bool>) -> Self

pub fn deletion_protection(self, input: bool) -> Self

A value that indicates whether the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled. For more information, see Deleting a DB Instance.

pub fn set_deletion_protection(self, input: Option<bool>) -> Self

pub fn max_allocated_storage(self, input: i32) -> Self

The upper limit to which Amazon RDS can automatically scale the storage of the DB instance.

For more information about this setting, including limitations that apply to it, see Managing capacity automatically with Amazon RDS storage autoscaling in the Amazon RDS User Guide.

pub fn set_max_allocated_storage(self, input: Option<i32>) -> Self

pub fn certificate_rotation_restart(self, input: bool) -> Self

A value that indicates whether the DB instance is restarted when you rotate your SSL/TLS certificate.

By default, the DB instance is restarted when you rotate your SSL/TLS certificate. The certificate is not updated until the DB instance is restarted.

Set this parameter only if you are not using SSL/TLS to connect to the DB instance.

If you are using SSL/TLS to connect to the DB instance, follow the appropriate instructions for your DB engine to rotate your SSL/TLS certificate:

For more information about rotating your SSL/TLS certificate for RDS DB engines, see Rotating Your SSL/TLS Certificate. in the Amazon RDS User Guide.

For more information about rotating your SSL/TLS certificate for Aurora DB engines, see Rotating Your SSL/TLS Certificate in the Amazon Aurora User Guide.

pub fn set_certificate_rotation_restart(self, input: Option<bool>) -> Self

pub fn replica_mode(self, input: ReplicaMode) -> Self

A value that sets the open mode of a replica database to either mounted or read-only.

Currently, this parameter is only supported for Oracle DB instances.

Mounted DB replicas are included in Oracle Enterprise Edition. The main use case for mounted replicas is cross-Region disaster recovery. The primary database doesn't use Active Data Guard to transmit information to the mounted replica. Because it doesn't accept user connections, a mounted replica can't serve a read-only workload. For more information, see Working with Oracle Read Replicas for Amazon RDS in the Amazon RDS User Guide.

pub fn set_replica_mode(self, input: Option<ReplicaMode>) -> Self

pub fn enable_customer_owned_ip(self, input: bool) -> Self

A value that indicates whether to enable a customer-owned IP address (CoIP) for an RDS on Outposts DB instance.

A CoIP provides local or external connectivity to resources in your Outpost subnets through your on-premises network. For some use cases, a CoIP can provide lower latency for connections to the DB instance from outside of its virtual private cloud (VPC) on your local network.

For more information about RDS on Outposts, see Working with Amazon RDS on Amazon Web Services Outposts in the Amazon RDS User Guide.

For more information about CoIPs, see Customer-owned IP addresses in the Amazon Web Services Outposts User Guide.

pub fn set_enable_customer_owned_ip(self, input: Option<bool>) -> Self

pub fn aws_backup_recovery_point_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the recovery point in Amazon Web Services Backup.

pub fn set_aws_backup_recovery_point_arn(self, input: Option<String>) -> Self