Struct aws_sdk_rds::client::fluent_builders::RestoreDBClusterFromS3

pub struct RestoreDBClusterFromS3<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> RestoreDBClusterFromS3<C>

pub async fn send(
self
) -> Result<RestoreDbClusterFromS3Output, SdkError<RestoreDBClusterFromS3Error>> where
C: SmithyConnector,

pub fn availability_zones(self, inp: impl Into<String>) -> Self

A list of Availability Zones (AZs) where instances in the restored DB cluster can be created.

pub fn set_availability_zones(self, input: Option<Vec<String>>) -> Self

pub fn backup_retention_period(self, input: i32) -> Self

The number of days for which automated backups of the restored DB cluster are retained. You must specify a minimum value of 1.

Default: 1

Constraints:

Must be a value from 1 to 35

pub fn set_backup_retention_period(self, input: Option<i32>) -> Self

pub fn character_set_name(self, input: impl Into<String>) -> Self

A value that indicates that the restored DB cluster should be associated with the specified CharacterSet.

pub fn set_character_set_name(self, input: Option<String>) -> Self

pub fn database_name(self, input: impl Into<String>) -> Self

The database name for the restored DB cluster.

pub fn set_database_name(self, input: Option<String>) -> Self

pub fn db_cluster_identifier(self, input: impl Into<String>) -> Self

The name of the DB cluster to create from the source data in the Amazon S3 bucket. This parameter isn't case-sensitive.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

Example: my-cluster1

pub fn set_db_cluster_identifier(self, input: Option<String>) -> Self

pub fn db_cluster_parameter_group_name(self, input: impl Into<String>) -> Self

The name of the DB cluster parameter group to associate with the restored DB cluster. If this argument is omitted, default.aurora5.6 is used.

Constraints:

If supplied, must match the name of an existing DBClusterParameterGroup.

pub fn set_db_cluster_parameter_group_name(self, input: Option<String>) -> Self

pub fn vpc_security_group_ids(self, inp: impl Into<String>) -> Self

A list of EC2 VPC security groups to associate with the restored DB cluster.

pub fn set_vpc_security_group_ids(self, input: Option<Vec<String>>) -> Self

pub fn db_subnet_group_name(self, input: impl Into<String>) -> Self

A DB subnet group to associate with the restored DB cluster.

Constraints: If supplied, must match the name of an existing DBSubnetGroup.

Example: mySubnetgroup

pub fn set_db_subnet_group_name(self, input: Option<String>) -> Self

pub fn engine(self, input: impl Into<String>) -> Self

The name of the database engine to be used for this DB cluster.

Valid Values: aurora (for MySQL 5.6-compatible Aurora), aurora-mysql (for MySQL 5.7-compatible Aurora), and aurora-postgresql

pub fn set_engine(self, input: Option<String>) -> Self

pub fn engine_version(self, input: impl Into<String>) -> Self

The version number of the database engine to use.

To list all of the available engine versions for aurora (for MySQL 5.6-compatible Aurora), use the following command:

aws rds describe-db-engine-versions --engine aurora --query "DBEngineVersions[].EngineVersion"

To list all of the available engine versions for aurora-mysql (for MySQL 5.7-compatible Aurora), use the following command:

aws rds describe-db-engine-versions --engine aurora-mysql --query "DBEngineVersions[].EngineVersion"

To list all of the available engine versions for aurora-postgresql, use the following command:

aws rds describe-db-engine-versions --engine aurora-postgresql --query "DBEngineVersions[].EngineVersion"

Aurora MySQL

Example: 5.6.10a, 5.6.mysql_aurora.1.19.2, 5.7.12, 5.7.mysql_aurora.2.04.5

Aurora PostgreSQL

Example: 9.6.3, 10.7

pub fn set_engine_version(self, input: Option<String>) -> Self

pub fn port(self, input: i32) -> Self

The port number on which the instances in the restored DB cluster accept connections.

Default: 3306

pub fn set_port(self, input: Option<i32>) -> Self

pub fn master_username(self, input: impl Into<String>) -> Self

The name of the master user for the restored DB cluster.

Constraints:

Must be 1 to 16 letters or numbers.

First character must be a letter.

Can't be a reserved word for the chosen database engine.

pub fn set_master_username(self, input: Option<String>) -> Self

pub fn master_user_password(self, input: impl Into<String>) -> Self

The password for the master database user. This password can contain any printable ASCII character except "/", """, or "@".

Constraints: Must contain from 8 to 41 characters.

pub fn set_master_user_password(self, input: Option<String>) -> Self

pub fn option_group_name(self, input: impl Into<String>) -> Self

A value that indicates that the restored DB cluster should be associated with the specified option group.

Permanent options can't be removed from an option group. An option group can't be removed from a DB cluster once it is associated with a DB cluster.

pub fn set_option_group_name(self, input: Option<String>) -> Self

pub fn preferred_backup_window(self, input: impl Into<String>) -> Self

The daily time range during which automated backups are created if automated backups are enabled using the BackupRetentionPeriod parameter.

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Web Services Region. To view the time blocks available, see Backup window in the Amazon Aurora User Guide.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

pub fn set_preferred_backup_window(self, input: Option<String>) -> Self

pub fn preferred_maintenance_window(self, input: impl Into<String>) -> Self

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Web Services Region, occurring on a random day of the week. To see the time blocks available, see Adjusting the Preferred Maintenance Window in the Amazon Aurora User Guide.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Constraints: Minimum 30-minute window.

pub fn set_preferred_maintenance_window(self, input: Option<String>) -> Self

pub fn tags(self, inp: impl Into<Tag>) -> Self

A list of tags. For more information, see Tagging Amazon RDS Resources in the Amazon RDS User Guide.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn storage_encrypted(self, input: bool) -> Self

A value that indicates whether the restored DB cluster is encrypted.

pub fn set_storage_encrypted(self, input: Option<bool>) -> Self

pub fn kms_key_id(self, input: impl Into<String>) -> Self

The Amazon Web Services KMS key identifier for an encrypted DB cluster.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK). To use a CMK in a different Amazon Web Services account, specify the key ARN or alias ARN.

If the StorageEncrypted parameter is enabled, and you do not specify a value for the KmsKeyId parameter, then Amazon RDS will use your default CMK. There is a default CMK for your Amazon Web Services account. Your Amazon Web Services account has a different default CMK for each Amazon Web Services Region.

pub fn set_kms_key_id(self, input: Option<String>) -> Self

pub fn enable_iam_database_authentication(self, input: bool) -> Self

A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

For more information, see IAM Database Authentication in the Amazon Aurora User Guide.

pub fn set_enable_iam_database_authentication(self, input: Option<bool>) -> Self

pub fn source_engine(self, input: impl Into<String>) -> Self

The identifier for the database engine that was backed up to create the files stored in the Amazon S3 bucket.

Valid values: mysql

pub fn set_source_engine(self, input: Option<String>) -> Self

pub fn source_engine_version(self, input: impl Into<String>) -> Self

The version of the database that the backup files were created from.

MySQL versions 5.5, 5.6, and 5.7 are supported.

Example: 5.6.40, 5.7.28

pub fn set_source_engine_version(self, input: Option<String>) -> Self

pub fn s3_bucket_name(self, input: impl Into<String>) -> Self

The name of the Amazon S3 bucket that contains the data used to create the Amazon Aurora DB cluster.

pub fn set_s3_bucket_name(self, input: Option<String>) -> Self

pub fn s3_prefix(self, input: impl Into<String>) -> Self

The prefix for all of the file names that contain the data used to create the Amazon Aurora DB cluster. If you do not specify a SourceS3Prefix value, then the Amazon Aurora DB cluster is created by using all of the files in the Amazon S3 bucket.

pub fn set_s3_prefix(self, input: Option<String>) -> Self

pub fn s3_ingestion_role_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the Amazon Web Services Identity and Access Management (IAM) role that authorizes Amazon RDS to access the Amazon S3 bucket on your behalf.

pub fn set_s3_ingestion_role_arn(self, input: Option<String>) -> Self

pub fn backtrack_window(self, input: i64) -> Self

The target backtrack window, in seconds. To disable backtracking, set this value to 0.

Currently, Backtrack is only supported for Aurora MySQL DB clusters.

Default: 0

Constraints:

If specified, this value must be set to a number from 0 to 259,200 (72 hours).

pub fn set_backtrack_window(self, input: Option<i64>) -> Self

pub fn enable_cloudwatch_logs_exports(self, inp: impl Into<String>) -> Self

The list of logs that the restored DB cluster is to export to CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see Publishing Database Logs to Amazon CloudWatch Logs in the Amazon Aurora User Guide.

pub fn set_enable_cloudwatch_logs_exports(
self,
input: Option<Vec<String>>
) -> Self

pub fn deletion_protection(self, input: bool) -> Self

A value that indicates whether the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled.

pub fn set_deletion_protection(self, input: Option<bool>) -> Self

pub fn copy_tags_to_snapshot(self, input: bool) -> Self

A value that indicates whether to copy all tags from the restored DB cluster to snapshots of the restored DB cluster. The default is not to copy them.

pub fn set_copy_tags_to_snapshot(self, input: Option<bool>) -> Self

pub fn domain(self, input: impl Into<String>) -> Self

Specify the Active Directory directory ID to restore the DB cluster in. The domain must be created prior to this operation.

For Amazon Aurora DB clusters, Amazon RDS can use Kerberos Authentication to authenticate users that connect to the DB cluster. For more information, see Kerberos Authentication in the Amazon Aurora User Guide.

pub fn set_domain(self, input: Option<String>) -> Self

pub fn domain_iam_role_name(self, input: impl Into<String>) -> Self

Specify the name of the IAM role to be used when making API calls to the Directory Service.

pub fn set_domain_iam_role_name(self, input: Option<String>) -> Self