Struct aws_sdk_rds::client::fluent_builders::RestoreDBClusterFromSnapshot

pub struct RestoreDBClusterFromSnapshot<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> RestoreDBClusterFromSnapshot<C>

pub async fn send(
self
) -> Result<RestoreDbClusterFromSnapshotOutput, SdkError<RestoreDBClusterFromSnapshotError>> where
C: SmithyConnector,

pub fn availability_zones(self, inp: impl Into<String>) -> Self

Provides the list of Availability Zones (AZs) where instances in the restored DB cluster can be created.

pub fn set_availability_zones(self, input: Option<Vec<String>>) -> Self

pub fn db_cluster_identifier(self, input: impl Into<String>) -> Self

The name of the DB cluster to create from the DB snapshot or DB cluster snapshot. This parameter isn't case-sensitive.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

Example: my-snapshot-id

pub fn set_db_cluster_identifier(self, input: Option<String>) -> Self

pub fn snapshot_identifier(self, input: impl Into<String>) -> Self

The identifier for the DB snapshot or DB cluster snapshot to restore from.

You can use either the name or the Amazon Resource Name (ARN) to specify a DB cluster snapshot. However, you can use only the ARN to specify a DB snapshot.

Constraints:

Must match the identifier of an existing Snapshot.

pub fn set_snapshot_identifier(self, input: Option<String>) -> Self

pub fn engine(self, input: impl Into<String>) -> Self

The database engine to use for the new DB cluster.

Default: The same as source

Constraint: Must be compatible with the engine of the source

pub fn set_engine(self, input: Option<String>) -> Self

pub fn engine_version(self, input: impl Into<String>) -> Self

The version of the database engine to use for the new DB cluster.

To list all of the available engine versions for aurora (for MySQL 5.6-compatible Aurora), use the following command:

aws rds describe-db-engine-versions --engine aurora --query "DBEngineVersions[].EngineVersion"

To list all of the available engine versions for aurora-mysql (for MySQL 5.7-compatible Aurora), use the following command:

aws rds describe-db-engine-versions --engine aurora-mysql --query "DBEngineVersions[].EngineVersion"

To list all of the available engine versions for aurora-postgresql, use the following command:

aws rds describe-db-engine-versions --engine aurora-postgresql --query "DBEngineVersions[].EngineVersion"

If you aren't using the default engine version, then you must specify the engine version.

Aurora MySQL

Example: 5.6.10a, 5.6.mysql_aurora.1.19.2, 5.7.12, 5.7.mysql_aurora.2.04.5

Aurora PostgreSQL

Example: 9.6.3, 10.7

pub fn set_engine_version(self, input: Option<String>) -> Self

pub fn port(self, input: i32) -> Self

The port number on which the new DB cluster accepts connections.

Constraints: This value must be 1150-65535

Default: The same port as the original DB cluster.

pub fn set_port(self, input: Option<i32>) -> Self

pub fn db_subnet_group_name(self, input: impl Into<String>) -> Self

The name of the DB subnet group to use for the new DB cluster.

Constraints: If supplied, must match the name of an existing DB subnet group.

Example: mySubnetgroup

pub fn set_db_subnet_group_name(self, input: Option<String>) -> Self

pub fn database_name(self, input: impl Into<String>) -> Self

The database name for the restored DB cluster.

pub fn set_database_name(self, input: Option<String>) -> Self

pub fn option_group_name(self, input: impl Into<String>) -> Self

The name of the option group to use for the restored DB cluster.

pub fn set_option_group_name(self, input: Option<String>) -> Self

pub fn vpc_security_group_ids(self, inp: impl Into<String>) -> Self

A list of VPC security groups that the new DB cluster will belong to.

pub fn set_vpc_security_group_ids(self, input: Option<Vec<String>>) -> Self

pub fn tags(self, inp: impl Into<Tag>) -> Self

The tags to be assigned to the restored DB cluster.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn kms_key_id(self, input: impl Into<String>) -> Self

The Amazon Web Services KMS key identifier to use when restoring an encrypted DB cluster from a DB snapshot or DB cluster snapshot.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK). To use a CMK in a different Amazon Web Services account, specify the key ARN or alias ARN.

When you don't specify a value for the KmsKeyId parameter, then the following occurs:

If the DB snapshot or DB cluster snapshot in SnapshotIdentifier is encrypted, then the restored DB cluster is encrypted using the Amazon Web Services KMS CMK that was used to encrypt the DB snapshot or DB cluster snapshot.

If the DB snapshot or DB cluster snapshot in SnapshotIdentifier isn't encrypted, then the restored DB cluster isn't encrypted.

pub fn set_kms_key_id(self, input: Option<String>) -> Self

pub fn enable_iam_database_authentication(self, input: bool) -> Self

A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

For more information, see IAM Database Authentication in the Amazon Aurora User Guide.

pub fn set_enable_iam_database_authentication(self, input: Option<bool>) -> Self

pub fn backtrack_window(self, input: i64) -> Self

The target backtrack window, in seconds. To disable backtracking, set this value to 0.

Currently, Backtrack is only supported for Aurora MySQL DB clusters.

Default: 0

Constraints:

If specified, this value must be set to a number from 0 to 259,200 (72 hours).

pub fn set_backtrack_window(self, input: Option<i64>) -> Self

pub fn enable_cloudwatch_logs_exports(self, inp: impl Into<String>) -> Self

The list of logs that the restored DB cluster is to export to Amazon CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see Publishing Database Logs to Amazon CloudWatch Logs in the Amazon Aurora User Guide.

pub fn set_enable_cloudwatch_logs_exports(
self,
input: Option<Vec<String>>
) -> Self

pub fn engine_mode(self, input: impl Into<String>) -> Self

The DB engine mode of the DB cluster, either provisioned, serverless, parallelquery, global, or multimaster.

For more information, see CreateDBCluster.

pub fn set_engine_mode(self, input: Option<String>) -> Self

pub fn scaling_configuration(self, input: ScalingConfiguration) -> Self

For DB clusters in serverless DB engine mode, the scaling properties of the DB cluster.

pub fn set_scaling_configuration(
self,
input: Option<ScalingConfiguration>
) -> Self

pub fn db_cluster_parameter_group_name(self, input: impl Into<String>) -> Self

The name of the DB cluster parameter group to associate with this DB cluster. If this argument is omitted, the default DB cluster parameter group for the specified engine is used.

Constraints:

If supplied, must match the name of an existing default DB cluster parameter group.

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

pub fn set_db_cluster_parameter_group_name(self, input: Option<String>) -> Self

pub fn deletion_protection(self, input: bool) -> Self

A value that indicates whether the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled.

pub fn set_deletion_protection(self, input: Option<bool>) -> Self

pub fn copy_tags_to_snapshot(self, input: bool) -> Self

A value that indicates whether to copy all tags from the restored DB cluster to snapshots of the restored DB cluster. The default is not to copy them.

pub fn set_copy_tags_to_snapshot(self, input: Option<bool>) -> Self

pub fn domain(self, input: impl Into<String>) -> Self

Specify the Active Directory directory ID to restore the DB cluster in. The domain must be created prior to this operation. Currently, only MySQL, Microsoft SQL Server, Oracle, and PostgreSQL DB instances can be created in an Active Directory Domain.

For more information, see Kerberos Authentication in the Amazon RDS User Guide.

pub fn set_domain(self, input: Option<String>) -> Self

pub fn domain_iam_role_name(self, input: impl Into<String>) -> Self

Specify the name of the IAM role to be used when making API calls to the Directory Service.

pub fn set_domain_iam_role_name(self, input: Option<String>) -> Self