Struct aws_sdk_rds::client::fluent_builders::RestoreDBInstanceFromDBSnapshot

pub struct RestoreDBInstanceFromDBSnapshot<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> RestoreDBInstanceFromDBSnapshot<C>

pub async fn send(
self
) -> Result<RestoreDbInstanceFromDbSnapshotOutput, SdkError<RestoreDBInstanceFromDBSnapshotError>> where
C: SmithyConnector,

pub fn db_instance_identifier(self, input: impl Into<String>) -> Self

Name of the DB instance to create from the DB snapshot. This parameter isn't case-sensitive.

Constraints:

Must contain from 1 to 63 numbers, letters, or hyphens

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

Example: my-snapshot-id

pub fn set_db_instance_identifier(self, input: Option<String>) -> Self

pub fn db_snapshot_identifier(self, input: impl Into<String>) -> Self

The identifier for the DB snapshot to restore from.

Constraints:

Must match the identifier of an existing DBSnapshot.

If you are restoring from a shared manual DB snapshot, the DBSnapshotIdentifier must be the ARN of the shared DB snapshot.

pub fn set_db_snapshot_identifier(self, input: Option<String>) -> Self

pub fn db_instance_class(self, input: impl Into<String>) -> Self

The compute and memory capacity of the Amazon RDS DB instance, for example, db.m4.large. Not all DB instance classes are available in all Amazon Web Services Regions, or for all database engines. For the full list of DB instance classes, and availability for your engine, see DB Instance Class in the Amazon RDS User Guide.

Default: The same DBInstanceClass as the original DB instance.

pub fn set_db_instance_class(self, input: Option<String>) -> Self

pub fn port(self, input: i32) -> Self

The port number on which the database accepts connections.

Default: The same port as the original DB instance

Constraints: Value must be 1150-65535

pub fn set_port(self, input: Option<i32>) -> Self

pub fn availability_zone(self, input: impl Into<String>) -> Self

The Availability Zone (AZ) where the DB instance will be created.

Default: A random, system-chosen Availability Zone.

Constraint: You can't specify the AvailabilityZone parameter if the DB instance is a Multi-AZ deployment.

Example: us-east-1a

pub fn set_availability_zone(self, input: Option<String>) -> Self

pub fn db_subnet_group_name(self, input: impl Into<String>) -> Self

The DB subnet group name to use for the new instance.

Constraints: If supplied, must match the name of an existing DBSubnetGroup.

Example: mySubnetgroup

pub fn set_db_subnet_group_name(self, input: Option<String>) -> Self

pub fn multi_az(self, input: bool) -> Self

A value that indicates whether the DB instance is a Multi-AZ deployment.

Constraint: You can't specify the AvailabilityZone parameter if the DB instance is a Multi-AZ deployment.

pub fn set_multi_az(self, input: Option<bool>) -> Self

pub fn publicly_accessible(self, input: bool) -> Self

A value that indicates whether the DB instance is publicly accessible.

When the DB instance is publicly accessible, its DNS endpoint resolves to the private IP address from within the DB instance's VPC, and to the public IP address from outside of the DB instance's VPC. Access to the DB instance is ultimately controlled by the security group it uses, and that public access is not permitted if the security group assigned to the DB instance doesn't permit it.

When the DB instance isn't publicly accessible, it is an internal DB instance with a DNS name that resolves to a private IP address.

For more information, see CreateDBInstance.

pub fn set_publicly_accessible(self, input: Option<bool>) -> Self

pub fn auto_minor_version_upgrade(self, input: bool) -> Self

A value that indicates whether minor version upgrades are applied automatically to the DB instance during the maintenance window.

pub fn set_auto_minor_version_upgrade(self, input: Option<bool>) -> Self

pub fn license_model(self, input: impl Into<String>) -> Self

License model information for the restored DB instance.

Default: Same as source.

Valid values: license-included | bring-your-own-license | general-public-license

pub fn set_license_model(self, input: Option<String>) -> Self

pub fn db_name(self, input: impl Into<String>) -> Self

The database name for the restored DB instance.

This parameter doesn't apply to the MySQL, PostgreSQL, or MariaDB engines.

pub fn set_db_name(self, input: Option<String>) -> Self

pub fn engine(self, input: impl Into<String>) -> Self

The database engine to use for the new instance.

Default: The same as source

Constraint: Must be compatible with the engine of the source. For example, you can restore a MariaDB 10.1 DB instance from a MySQL 5.6 snapshot.

Valid Values:

mariadb

mysql

oracle-ee

oracle-ee-cdb

oracle-se2

oracle-se2-cdb

postgres

sqlserver-ee

sqlserver-se

sqlserver-ex

sqlserver-web

pub fn set_engine(self, input: Option<String>) -> Self

pub fn iops(self, input: i32) -> Self

Specifies the amount of provisioned IOPS for the DB instance, expressed in I/O operations per second. If this parameter isn't specified, the IOPS value is taken from the backup. If this parameter is set to 0, the new instance is converted to a non-PIOPS instance. The conversion takes additional time, though your DB instance is available for connections before the conversion starts.

The provisioned IOPS value must follow the requirements for your database engine. For more information, see Amazon RDS Provisioned IOPS Storage to Improve Performance in the Amazon RDS User Guide.

Constraints: Must be an integer greater than 1000.

pub fn set_iops(self, input: Option<i32>) -> Self

pub fn option_group_name(self, input: impl Into<String>) -> Self

The name of the option group to be used for the restored DB instance.

Permanent options, such as the TDE option for Oracle Advanced Security TDE, can't be removed from an option group, and that option group can't be removed from a DB instance once it is associated with a DB instance

pub fn set_option_group_name(self, input: Option<String>) -> Self

pub fn tags(self, inp: impl Into<Tag>) -> Self

A list of tags. For more information, see Tagging Amazon RDS Resources in the Amazon RDS User Guide.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn storage_type(self, input: impl Into<String>) -> Self

Specifies the storage type to be associated with the DB instance.

Valid values: standard | gp2 | io1

If you specify io1, you must also include a value for the Iops parameter.

Default: io1 if the Iops parameter is specified, otherwise gp2

pub fn set_storage_type(self, input: Option<String>) -> Self

pub fn tde_credential_arn(self, input: impl Into<String>) -> Self

The ARN from the key store with which to associate the instance for TDE encryption.

pub fn set_tde_credential_arn(self, input: Option<String>) -> Self

pub fn tde_credential_password(self, input: impl Into<String>) -> Self

The password for the given ARN from the key store in order to access the device.

pub fn set_tde_credential_password(self, input: Option<String>) -> Self

pub fn vpc_security_group_ids(self, inp: impl Into<String>) -> Self

A list of EC2 VPC security groups to associate with this DB instance.

Default: The default EC2 VPC security group for the DB subnet group's VPC.

pub fn set_vpc_security_group_ids(self, input: Option<Vec<String>>) -> Self

pub fn domain(self, input: impl Into<String>) -> Self

Specify the Active Directory directory ID to restore the DB instance in. The domain must be created prior to this operation. Currently, only MySQL, Microsoft SQL Server, Oracle, and PostgreSQL DB instances can be created in an Active Directory Domain.

For more information, see Kerberos Authentication in the Amazon RDS User Guide.

pub fn set_domain(self, input: Option<String>) -> Self

pub fn copy_tags_to_snapshot(self, input: bool) -> Self

A value that indicates whether to copy all tags from the restored DB instance to snapshots of the DB instance. By default, tags are not copied.

pub fn set_copy_tags_to_snapshot(self, input: Option<bool>) -> Self

pub fn domain_iam_role_name(self, input: impl Into<String>) -> Self

Specify the name of the IAM role to be used when making API calls to the Directory Service.

pub fn set_domain_iam_role_name(self, input: Option<String>) -> Self

pub fn enable_iam_database_authentication(self, input: bool) -> Self

A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

For more information about IAM database authentication, see IAM Database Authentication for MySQL and PostgreSQL in the Amazon RDS User Guide.

pub fn set_enable_iam_database_authentication(self, input: Option<bool>) -> Self

pub fn enable_cloudwatch_logs_exports(self, inp: impl Into<String>) -> Self

The list of logs that the restored DB instance is to export to CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see Publishing Database Logs to Amazon CloudWatch Logs in the Amazon RDS User Guide.

pub fn set_enable_cloudwatch_logs_exports(
self,
input: Option<Vec<String>>
) -> Self

pub fn processor_features(self, inp: impl Into<ProcessorFeature>) -> Self

The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

pub fn set_processor_features(
self,
input: Option<Vec<ProcessorFeature>>
) -> Self

pub fn use_default_processor_features(self, input: bool) -> Self

A value that indicates whether the DB instance class of the DB instance uses its default processor features.

pub fn set_use_default_processor_features(self, input: Option<bool>) -> Self

pub fn db_parameter_group_name(self, input: impl Into<String>) -> Self

The name of the DB parameter group to associate with this DB instance.

If you do not specify a value for DBParameterGroupName, then the default DBParameterGroup for the specified DB engine is used.

Constraints:

If supplied, must match the name of an existing DBParameterGroup.

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

pub fn set_db_parameter_group_name(self, input: Option<String>) -> Self

pub fn deletion_protection(self, input: bool) -> Self

A value that indicates whether the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled. For more information, see Deleting a DB Instance.

pub fn set_deletion_protection(self, input: Option<bool>) -> Self

pub fn enable_customer_owned_ip(self, input: bool) -> Self

A value that indicates whether to enable a customer-owned IP address (CoIP) for an RDS on Outposts DB instance.

A CoIP provides local or external connectivity to resources in your Outpost subnets through your on-premises network. For some use cases, a CoIP can provide lower latency for connections to the DB instance from outside of its virtual private cloud (VPC) on your local network.

For more information about RDS on Outposts, see Working with Amazon RDS on Amazon Web Services Outposts in the Amazon RDS User Guide.

For more information about CoIPs, see Customer-owned IP addresses in the Amazon Web Services Outposts User Guide.

pub fn set_enable_customer_owned_ip(self, input: Option<bool>) -> Self