Struct aws_sdk_rds::client::fluent_builders::RestoreDBInstanceFromS3

pub struct RestoreDBInstanceFromS3<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> RestoreDBInstanceFromS3<C>

pub async fn send(
self
) -> Result<RestoreDbInstanceFromS3Output, SdkError<RestoreDBInstanceFromS3Error>> where
C: SmithyConnector,

pub fn db_name(self, input: impl Into<String>) -> Self

The name of the database to create when the DB instance is created. Follow the naming rules specified in CreateDBInstance.

pub fn set_db_name(self, input: Option<String>) -> Self

pub fn db_instance_identifier(self, input: impl Into<String>) -> Self

The DB instance identifier. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

Example: mydbinstance

pub fn set_db_instance_identifier(self, input: Option<String>) -> Self

pub fn allocated_storage(self, input: i32) -> Self

The amount of storage (in gigabytes) to allocate initially for the DB instance. Follow the allocation rules specified in CreateDBInstance.

Be sure to allocate enough memory for your new DB instance so that the restore operation can succeed. You can also allocate additional memory for future growth.

pub fn set_allocated_storage(self, input: Option<i32>) -> Self

pub fn db_instance_class(self, input: impl Into<String>) -> Self

The compute and memory capacity of the DB instance, for example, db.m4.large. Not all DB instance classes are available in all Amazon Web Services Regions, or for all database engines. For the full list of DB instance classes, and availability for your engine, see DB Instance Class in the Amazon RDS User Guide.

Importing from Amazon S3 isn't supported on the db.t2.micro DB instance class.

pub fn set_db_instance_class(self, input: Option<String>) -> Self

pub fn engine(self, input: impl Into<String>) -> Self

The name of the database engine to be used for this instance.

Valid Values: mysql

pub fn set_engine(self, input: Option<String>) -> Self

pub fn master_username(self, input: impl Into<String>) -> Self

The name for the master user.

Constraints:

Must be 1 to 16 letters or numbers.

First character must be a letter.

Can't be a reserved word for the chosen database engine.

pub fn set_master_username(self, input: Option<String>) -> Self

pub fn master_user_password(self, input: impl Into<String>) -> Self

The password for the master user. The password can include any printable ASCII character except "/", """, or "@".

Constraints: Must contain from 8 to 41 characters.

pub fn set_master_user_password(self, input: Option<String>) -> Self

pub fn db_security_groups(self, inp: impl Into<String>) -> Self

A list of DB security groups to associate with this DB instance.

Default: The default DB security group for the database engine.

pub fn set_db_security_groups(self, input: Option<Vec<String>>) -> Self

pub fn vpc_security_group_ids(self, inp: impl Into<String>) -> Self

A list of VPC security groups to associate with this DB instance.

pub fn set_vpc_security_group_ids(self, input: Option<Vec<String>>) -> Self

pub fn availability_zone(self, input: impl Into<String>) -> Self

The Availability Zone that the DB instance is created in. For information about Amazon Web Services Regions and Availability Zones, see Regions and Availability Zones in the Amazon RDS User Guide.

Default: A random, system-chosen Availability Zone in the endpoint's Amazon Web Services Region.

Example: us-east-1d

Constraint: The AvailabilityZone parameter can't be specified if the DB instance is a Multi-AZ deployment. The specified Availability Zone must be in the same Amazon Web Services Region as the current endpoint.

pub fn set_availability_zone(self, input: Option<String>) -> Self

pub fn db_subnet_group_name(self, input: impl Into<String>) -> Self

A DB subnet group to associate with this DB instance.

pub fn set_db_subnet_group_name(self, input: Option<String>) -> Self

pub fn preferred_maintenance_window(self, input: impl Into<String>) -> Self

The time range each week during which system maintenance can occur, in Universal Coordinated Time (UTC). For more information, see Amazon RDS Maintenance Window in the Amazon RDS User Guide.

Constraints:

Must be in the format ddd:hh24:mi-ddd:hh24:mi.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred backup window.

Must be at least 30 minutes.

pub fn set_preferred_maintenance_window(self, input: Option<String>) -> Self

pub fn db_parameter_group_name(self, input: impl Into<String>) -> Self

The name of the DB parameter group to associate with this DB instance.

If you do not specify a value for DBParameterGroupName, then the default DBParameterGroup for the specified DB engine is used.

pub fn set_db_parameter_group_name(self, input: Option<String>) -> Self

pub fn backup_retention_period(self, input: i32) -> Self

The number of days for which automated backups are retained. Setting this parameter to a positive number enables backups. For more information, see CreateDBInstance.

pub fn set_backup_retention_period(self, input: Option<i32>) -> Self

pub fn preferred_backup_window(self, input: impl Into<String>) -> Self

The time range each day during which automated backups are created if automated backups are enabled. For more information, see Backup window in the Amazon RDS User Guide.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

pub fn set_preferred_backup_window(self, input: Option<String>) -> Self

pub fn port(self, input: i32) -> Self

The port number on which the database accepts connections.

Type: Integer

Valid Values: 1150-65535

Default: 3306

pub fn set_port(self, input: Option<i32>) -> Self

pub fn multi_az(self, input: bool) -> Self

A value that indicates whether the DB instance is a Multi-AZ deployment. If the DB instance is a Multi-AZ deployment, you can't set the AvailabilityZone parameter.

pub fn set_multi_az(self, input: Option<bool>) -> Self

pub fn engine_version(self, input: impl Into<String>) -> Self

The version number of the database engine to use. Choose the latest minor version of your database engine. For information about engine versions, see CreateDBInstance, or call DescribeDBEngineVersions.

pub fn set_engine_version(self, input: Option<String>) -> Self

pub fn auto_minor_version_upgrade(self, input: bool) -> Self

A value that indicates whether minor engine upgrades are applied automatically to the DB instance during the maintenance window. By default, minor engine upgrades are not applied automatically.

pub fn set_auto_minor_version_upgrade(self, input: Option<bool>) -> Self

pub fn license_model(self, input: impl Into<String>) -> Self

The license model for this DB instance. Use general-public-license.

pub fn set_license_model(self, input: Option<String>) -> Self

pub fn iops(self, input: i32) -> Self

The amount of Provisioned IOPS (input/output operations per second) to allocate initially for the DB instance. For information about valid Iops values, see Amazon RDS Provisioned IOPS Storage to Improve Performance in the Amazon RDS User Guide.

pub fn set_iops(self, input: Option<i32>) -> Self

pub fn option_group_name(self, input: impl Into<String>) -> Self

The name of the option group to associate with this DB instance. If this argument is omitted, the default option group for the specified engine is used.

pub fn set_option_group_name(self, input: Option<String>) -> Self

pub fn publicly_accessible(self, input: bool) -> Self

A value that indicates whether the DB instance is publicly accessible.

When the DB instance is publicly accessible, its DNS endpoint resolves to the private IP address from within the DB instance's VPC, and to the public IP address from outside of the DB instance's VPC. Access to the DB instance is ultimately controlled by the security group it uses, and that public access is not permitted if the security group assigned to the DB instance doesn't permit it.

When the DB instance isn't publicly accessible, it is an internal DB instance with a DNS name that resolves to a private IP address.

For more information, see CreateDBInstance.

pub fn set_publicly_accessible(self, input: Option<bool>) -> Self

pub fn tags(self, inp: impl Into<Tag>) -> Self

A list of tags to associate with this DB instance. For more information, see Tagging Amazon RDS Resources in the Amazon RDS User Guide.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn storage_type(self, input: impl Into<String>) -> Self

Specifies the storage type to be associated with the DB instance.

Valid values: standard | gp2 | io1

If you specify io1, you must also include a value for the Iops parameter.

Default: io1 if the Iops parameter is specified; otherwise gp2

pub fn set_storage_type(self, input: Option<String>) -> Self

pub fn storage_encrypted(self, input: bool) -> Self

A value that indicates whether the new DB instance is encrypted or not.

pub fn set_storage_encrypted(self, input: Option<bool>) -> Self

pub fn kms_key_id(self, input: impl Into<String>) -> Self

The Amazon Web Services KMS key identifier for an encrypted DB instance.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK). To use a CMK in a different Amazon Web Services account, specify the key ARN or alias ARN.

If the StorageEncrypted parameter is enabled, and you do not specify a value for the KmsKeyId parameter, then Amazon RDS will use your default CMK. There is a default CMK for your Amazon Web Services account. Your Amazon Web Services account has a different default CMK for each Amazon Web Services Region.

pub fn set_kms_key_id(self, input: Option<String>) -> Self

pub fn copy_tags_to_snapshot(self, input: bool) -> Self

A value that indicates whether to copy all tags from the DB instance to snapshots of the DB instance. By default, tags are not copied.

pub fn set_copy_tags_to_snapshot(self, input: Option<bool>) -> Self

pub fn monitoring_interval(self, input: i32) -> Self

The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the DB instance. To disable collecting Enhanced Monitoring metrics, specify 0.

If MonitoringRoleArn is specified, then you must also set MonitoringInterval to a value other than 0.

Valid Values: 0, 1, 5, 10, 15, 30, 60

Default: 0

pub fn set_monitoring_interval(self, input: Option<i32>) -> Self

pub fn monitoring_role_arn(self, input: impl Into<String>) -> Self

The ARN for the IAM role that permits RDS to send enhanced monitoring metrics to Amazon CloudWatch Logs. For example, arn:aws:iam:123456789012:role/emaccess. For information on creating a monitoring role, see Setting Up and Enabling Enhanced Monitoring in the Amazon RDS User Guide.

If MonitoringInterval is set to a value other than 0, then you must supply a MonitoringRoleArn value.

pub fn set_monitoring_role_arn(self, input: Option<String>) -> Self

pub fn enable_iam_database_authentication(self, input: bool) -> Self

A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

For more information about IAM database authentication, see IAM Database Authentication for MySQL and PostgreSQL in the Amazon RDS User Guide.

pub fn set_enable_iam_database_authentication(self, input: Option<bool>) -> Self

pub fn source_engine(self, input: impl Into<String>) -> Self

The name of the engine of your source database.

Valid Values: mysql

pub fn set_source_engine(self, input: Option<String>) -> Self

pub fn source_engine_version(self, input: impl Into<String>) -> Self

The version of the database that the backup files were created from.

MySQL versions 5.6 and 5.7 are supported.

Example: 5.6.40

pub fn set_source_engine_version(self, input: Option<String>) -> Self

pub fn s3_bucket_name(self, input: impl Into<String>) -> Self

The name of your Amazon S3 bucket that contains your database backup file.

pub fn set_s3_bucket_name(self, input: Option<String>) -> Self

pub fn s3_prefix(self, input: impl Into<String>) -> Self

The prefix of your Amazon S3 bucket.

pub fn set_s3_prefix(self, input: Option<String>) -> Self

pub fn s3_ingestion_role_arn(self, input: impl Into<String>) -> Self

An Amazon Web Services Identity and Access Management (IAM) role to allow Amazon RDS to access your Amazon S3 bucket.

pub fn set_s3_ingestion_role_arn(self, input: Option<String>) -> Self

pub fn enable_performance_insights(self, input: bool) -> Self

A value that indicates whether to enable Performance Insights for the DB instance.

For more information, see Using Amazon Performance Insights in the Amazon Relational Database Service User Guide.

pub fn set_enable_performance_insights(self, input: Option<bool>) -> Self

pub fn performance_insights_kms_key_id(self, input: impl Into<String>) -> Self

The Amazon Web Services KMS key identifier for encryption of Performance Insights data.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK).

If you do not specify a value for PerformanceInsightsKMSKeyId, then Amazon RDS uses your default CMK. There is a default CMK for your Amazon Web Services account. Your Amazon Web Services account has a different default CMK for each Amazon Web Services Region.

pub fn set_performance_insights_kms_key_id(self, input: Option<String>) -> Self

pub fn performance_insights_retention_period(self, input: i32) -> Self

The amount of time, in days, to retain Performance Insights data. Valid values are 7 or 731 (2 years).

pub fn set_performance_insights_retention_period(
self,
input: Option<i32>
) -> Self

pub fn enable_cloudwatch_logs_exports(self, inp: impl Into<String>) -> Self

The list of logs that the restored DB instance is to export to CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see Publishing Database Logs to Amazon CloudWatch Logs in the Amazon RDS User Guide.

pub fn set_enable_cloudwatch_logs_exports(
self,
input: Option<Vec<String>>
) -> Self

pub fn processor_features(self, inp: impl Into<ProcessorFeature>) -> Self

The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

pub fn set_processor_features(
self,
input: Option<Vec<ProcessorFeature>>
) -> Self

pub fn use_default_processor_features(self, input: bool) -> Self

A value that indicates whether the DB instance class of the DB instance uses its default processor features.

pub fn set_use_default_processor_features(self, input: Option<bool>) -> Self

pub fn deletion_protection(self, input: bool) -> Self

A value that indicates whether the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled. For more information, see Deleting a DB Instance.

pub fn set_deletion_protection(self, input: Option<bool>) -> Self

pub fn max_allocated_storage(self, input: i32) -> Self

The upper limit to which Amazon RDS can automatically scale the storage of the DB instance.

For more information about this setting, including limitations that apply to it, see Managing capacity automatically with Amazon RDS storage autoscaling in the Amazon RDS User Guide.

pub fn set_max_allocated_storage(self, input: Option<i32>) -> Self