Struct aws_sdk_rds::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for AmazonRDSv19.

This client allows ergonomic access to a AmazonRDSv19-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn add_role_to_db_cluster(&self) -> AddRoleToDBCluster<C, M, R>

pub fn add_role_to_db_instance(&self) -> AddRoleToDBInstance<C, M, R>

pub fn add_source_identifier_to_subscription(
&self
) -> AddSourceIdentifierToSubscription<C, M, R>

pub fn add_tags_to_resource(&self) -> AddTagsToResource<C, M, R>

pub fn apply_pending_maintenance_action(
&self
) -> ApplyPendingMaintenanceAction<C, M, R>

pub fn authorize_db_security_group_ingress(
&self
) -> AuthorizeDBSecurityGroupIngress<C, M, R>

pub fn backtrack_db_cluster(&self) -> BacktrackDBCluster<C, M, R>

pub fn cancel_export_task(&self) -> CancelExportTask<C, M, R>

pub fn copy_db_cluster_parameter_group(
&self
) -> CopyDBClusterParameterGroup<C, M, R>

pub fn copy_db_cluster_snapshot(&self) -> CopyDBClusterSnapshot<C, M, R>

pub fn copy_db_parameter_group(&self) -> CopyDBParameterGroup<C, M, R>

pub fn copy_db_snapshot(&self) -> CopyDBSnapshot<C, M, R>

pub fn copy_option_group(&self) -> CopyOptionGroup<C, M, R>

pub fn create_custom_availability_zone(
&self
) -> CreateCustomAvailabilityZone<C, M, R>

pub fn create_db_cluster(&self) -> CreateDBCluster<C, M, R>

pub fn create_db_cluster_endpoint(&self) -> CreateDBClusterEndpoint<C, M, R>

pub fn create_db_cluster_parameter_group(
&self
) -> CreateDBClusterParameterGroup<C, M, R>

pub fn create_db_cluster_snapshot(&self) -> CreateDBClusterSnapshot<C, M, R>

pub fn create_db_instance(&self) -> CreateDBInstance<C, M, R>

pub fn create_db_instance_read_replica(
&self
) -> CreateDBInstanceReadReplica<C, M, R>

pub fn create_db_parameter_group(&self) -> CreateDBParameterGroup<C, M, R>

pub fn create_db_proxy(&self) -> CreateDBProxy<C, M, R>

pub fn create_db_proxy_endpoint(&self) -> CreateDBProxyEndpoint<C, M, R>

pub fn create_db_security_group(&self) -> CreateDBSecurityGroup<C, M, R>

pub fn create_db_snapshot(&self) -> CreateDBSnapshot<C, M, R>

pub fn create_db_subnet_group(&self) -> CreateDBSubnetGroup<C, M, R>

pub fn create_event_subscription(&self) -> CreateEventSubscription<C, M, R>

pub fn create_global_cluster(&self) -> CreateGlobalCluster<C, M, R>

pub fn create_option_group(&self) -> CreateOptionGroup<C, M, R>

pub fn delete_custom_availability_zone(
&self
) -> DeleteCustomAvailabilityZone<C, M, R>

pub fn delete_db_cluster(&self) -> DeleteDBCluster<C, M, R>

pub fn delete_db_cluster_endpoint(&self) -> DeleteDBClusterEndpoint<C, M, R>

pub fn delete_db_cluster_parameter_group(
&self
) -> DeleteDBClusterParameterGroup<C, M, R>

pub fn delete_db_cluster_snapshot(&self) -> DeleteDBClusterSnapshot<C, M, R>

pub fn delete_db_instance(&self) -> DeleteDBInstance<C, M, R>

pub fn delete_db_instance_automated_backup(
&self
) -> DeleteDBInstanceAutomatedBackup<C, M, R>

pub fn delete_db_parameter_group(&self) -> DeleteDBParameterGroup<C, M, R>

pub fn delete_db_proxy(&self) -> DeleteDBProxy<C, M, R>

pub fn delete_db_proxy_endpoint(&self) -> DeleteDBProxyEndpoint<C, M, R>

pub fn delete_db_security_group(&self) -> DeleteDBSecurityGroup<C, M, R>

pub fn delete_db_snapshot(&self) -> DeleteDBSnapshot<C, M, R>

pub fn delete_db_subnet_group(&self) -> DeleteDBSubnetGroup<C, M, R>

pub fn delete_event_subscription(&self) -> DeleteEventSubscription<C, M, R>

pub fn delete_global_cluster(&self) -> DeleteGlobalCluster<C, M, R>

pub fn delete_installation_media(&self) -> DeleteInstallationMedia<C, M, R>

pub fn delete_option_group(&self) -> DeleteOptionGroup<C, M, R>

pub fn deregister_db_proxy_targets(&self) -> DeregisterDBProxyTargets<C, M, R>

pub fn describe_account_attributes(&self) -> DescribeAccountAttributes<C, M, R>

pub fn describe_certificates(&self) -> DescribeCertificates<C, M, R>

pub fn describe_custom_availability_zones(
&self
) -> DescribeCustomAvailabilityZones<C, M, R>

pub fn describe_db_cluster_backtracks(
&self
) -> DescribeDBClusterBacktracks<C, M, R>

pub fn describe_db_cluster_endpoints(
&self
) -> DescribeDBClusterEndpoints<C, M, R>

pub fn describe_db_cluster_parameter_groups(
&self
) -> DescribeDBClusterParameterGroups<C, M, R>

pub fn describe_db_cluster_parameters(
&self
) -> DescribeDBClusterParameters<C, M, R>

pub fn describe_db_clusters(&self) -> DescribeDBClusters<C, M, R>

pub fn describe_db_cluster_snapshot_attributes(
&self
) -> DescribeDBClusterSnapshotAttributes<C, M, R>

pub fn describe_db_cluster_snapshots(
&self
) -> DescribeDBClusterSnapshots<C, M, R>

pub fn describe_db_engine_versions(&self) -> DescribeDBEngineVersions<C, M, R>

pub fn describe_db_instance_automated_backups(
&self
) -> DescribeDBInstanceAutomatedBackups<C, M, R>

pub fn describe_db_instances(&self) -> DescribeDBInstances<C, M, R>

pub fn describe_db_log_files(&self) -> DescribeDBLogFiles<C, M, R>

pub fn describe_db_parameter_groups(&self) -> DescribeDBParameterGroups<C, M, R>

pub fn describe_db_parameters(&self) -> DescribeDBParameters<C, M, R>

pub fn describe_db_proxies(&self) -> DescribeDBProxies<C, M, R>

pub fn describe_db_proxy_endpoints(&self) -> DescribeDBProxyEndpoints<C, M, R>

pub fn describe_db_proxy_target_groups(
&self
) -> DescribeDBProxyTargetGroups<C, M, R>

pub fn describe_db_proxy_targets(&self) -> DescribeDBProxyTargets<C, M, R>

pub fn describe_db_security_groups(&self) -> DescribeDBSecurityGroups<C, M, R>

pub fn describe_db_snapshot_attributes(
&self
) -> DescribeDBSnapshotAttributes<C, M, R>

pub fn describe_db_snapshots(&self) -> DescribeDBSnapshots<C, M, R>

pub fn describe_db_subnet_groups(&self) -> DescribeDBSubnetGroups<C, M, R>

pub fn describe_engine_default_cluster_parameters(
&self
) -> DescribeEngineDefaultClusterParameters<C, M, R>

pub fn describe_engine_default_parameters(
&self
) -> DescribeEngineDefaultParameters<C, M, R>

pub fn describe_event_categories(&self) -> DescribeEventCategories<C, M, R>

pub fn describe_events(&self) -> DescribeEvents<C, M, R>

pub fn describe_event_subscriptions(
&self
) -> DescribeEventSubscriptions<C, M, R>

pub fn describe_export_tasks(&self) -> DescribeExportTasks<C, M, R>

pub fn describe_global_clusters(&self) -> DescribeGlobalClusters<C, M, R>

pub fn describe_installation_media(&self) -> DescribeInstallationMedia<C, M, R>

pub fn describe_option_group_options(
&self
) -> DescribeOptionGroupOptions<C, M, R>

pub fn describe_option_groups(&self) -> DescribeOptionGroups<C, M, R>

pub fn describe_orderable_db_instance_options(
&self
) -> DescribeOrderableDBInstanceOptions<C, M, R>

pub fn describe_pending_maintenance_actions(
&self
) -> DescribePendingMaintenanceActions<C, M, R>

pub fn describe_reserved_db_instances(
&self
) -> DescribeReservedDBInstances<C, M, R>

pub fn describe_reserved_db_instances_offerings(
&self
) -> DescribeReservedDBInstancesOfferings<C, M, R>

pub fn describe_source_regions(&self) -> DescribeSourceRegions<C, M, R>

pub fn describe_valid_db_instance_modifications(
&self
) -> DescribeValidDBInstanceModifications<C, M, R>

pub fn download_db_log_file_portion(&self) -> DownloadDBLogFilePortion<C, M, R>

pub fn failover_db_cluster(&self) -> FailoverDBCluster<C, M, R>

pub fn failover_global_cluster(&self) -> FailoverGlobalCluster<C, M, R>

pub fn import_installation_media(&self) -> ImportInstallationMedia<C, M, R>

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

pub fn modify_certificates(&self) -> ModifyCertificates<C, M, R>

pub fn modify_current_db_cluster_capacity(
&self
) -> ModifyCurrentDBClusterCapacity<C, M, R>

pub fn modify_db_cluster(&self) -> ModifyDBCluster<C, M, R>

pub fn modify_db_cluster_endpoint(&self) -> ModifyDBClusterEndpoint<C, M, R>

pub fn modify_db_cluster_parameter_group(
&self
) -> ModifyDBClusterParameterGroup<C, M, R>

pub fn modify_db_cluster_snapshot_attribute(
&self
) -> ModifyDBClusterSnapshotAttribute<C, M, R>

pub fn modify_db_instance(&self) -> ModifyDBInstance<C, M, R>

pub fn modify_db_parameter_group(&self) -> ModifyDBParameterGroup<C, M, R>

pub fn modify_db_proxy(&self) -> ModifyDBProxy<C, M, R>

pub fn modify_db_proxy_endpoint(&self) -> ModifyDBProxyEndpoint<C, M, R>

pub fn modify_db_proxy_target_group(&self) -> ModifyDBProxyTargetGroup<C, M, R>

pub fn modify_db_snapshot(&self) -> ModifyDBSnapshot<C, M, R>

pub fn modify_db_snapshot_attribute(&self) -> ModifyDBSnapshotAttribute<C, M, R>

pub fn modify_db_subnet_group(&self) -> ModifyDBSubnetGroup<C, M, R>

pub fn modify_event_subscription(&self) -> ModifyEventSubscription<C, M, R>

pub fn modify_global_cluster(&self) -> ModifyGlobalCluster<C, M, R>

pub fn modify_option_group(&self) -> ModifyOptionGroup<C, M, R>

pub fn promote_read_replica(&self) -> PromoteReadReplica<C, M, R>

pub fn promote_read_replica_db_cluster(
&self
) -> PromoteReadReplicaDBCluster<C, M, R>

pub fn purchase_reserved_db_instances_offering(
&self
) -> PurchaseReservedDBInstancesOffering<C, M, R>

pub fn reboot_db_instance(&self) -> RebootDBInstance<C, M, R>

pub fn register_db_proxy_targets(&self) -> RegisterDBProxyTargets<C, M, R>

pub fn remove_from_global_cluster(&self) -> RemoveFromGlobalCluster<C, M, R>

pub fn remove_role_from_db_cluster(&self) -> RemoveRoleFromDBCluster<C, M, R>

pub fn remove_role_from_db_instance(&self) -> RemoveRoleFromDBInstance<C, M, R>

pub fn remove_source_identifier_from_subscription(
&self
) -> RemoveSourceIdentifierFromSubscription<C, M, R>

pub fn remove_tags_from_resource(&self) -> RemoveTagsFromResource<C, M, R>

pub fn reset_db_cluster_parameter_group(
&self
) -> ResetDBClusterParameterGroup<C, M, R>

pub fn reset_db_parameter_group(&self) -> ResetDBParameterGroup<C, M, R>

pub fn restore_db_cluster_from_s3(&self) -> RestoreDBClusterFromS3<C, M, R>

pub fn restore_db_cluster_from_snapshot(
&self
) -> RestoreDBClusterFromSnapshot<C, M, R>

pub fn restore_db_cluster_to_point_in_time(
&self
) -> RestoreDBClusterToPointInTime<C, M, R>

pub fn restore_db_instance_from_db_snapshot(
&self
) -> RestoreDBInstanceFromDBSnapshot<C, M, R>

pub fn restore_db_instance_from_s3(&self) -> RestoreDBInstanceFromS3<C, M, R>

pub fn restore_db_instance_to_point_in_time(
&self
) -> RestoreDBInstanceToPointInTime<C, M, R>

pub fn revoke_db_security_group_ingress(
&self
) -> RevokeDBSecurityGroupIngress<C, M, R>

pub fn start_activity_stream(&self) -> StartActivityStream<C, M, R>

pub fn start_db_cluster(&self) -> StartDBCluster<C, M, R>

pub fn start_db_instance(&self) -> StartDBInstance<C, M, R>

pub fn start_db_instance_automated_backups_replication(
&self
) -> StartDBInstanceAutomatedBackupsReplication<C, M, R>

pub fn start_export_task(&self) -> StartExportTask<C, M, R>

pub fn stop_activity_stream(&self) -> StopActivityStream<C, M, R>

pub fn stop_db_cluster(&self) -> StopDBCluster<C, M, R>

pub fn stop_db_instance(&self) -> StopDBInstance<C, M, R>

pub fn stop_db_instance_automated_backups_replication(
&self
) -> StopDBInstanceAutomatedBackupsReplication<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self