Module aws_sdk_rds::error

Modules

authorization_already_exists_fault

See AuthorizationAlreadyExistsFault

authorization_not_found_fault

See AuthorizationNotFoundFault

authorization_quota_exceeded_fault

See AuthorizationQuotaExceededFault

backup_policy_not_found_fault

See BackupPolicyNotFoundFault

certificate_not_found_fault

See CertificateNotFoundFault

custom_availability_zone_already_exists_fault

See CustomAvailabilityZoneAlreadyExistsFault

custom_availability_zone_not_found_fault

See CustomAvailabilityZoneNotFoundFault

custom_availability_zone_quota_exceeded_fault

See CustomAvailabilityZoneQuotaExceededFault

db_cluster_already_exists_fault

See DbClusterAlreadyExistsFault

db_cluster_backtrack_not_found_fault

See DbClusterBacktrackNotFoundFault

db_cluster_endpoint_already_exists_fault

See DbClusterEndpointAlreadyExistsFault

db_cluster_endpoint_not_found_fault

See DbClusterEndpointNotFoundFault

db_cluster_endpoint_quota_exceeded_fault

See DbClusterEndpointQuotaExceededFault

db_cluster_not_found_fault

See DbClusterNotFoundFault

db_cluster_parameter_group_not_found_fault

See DbClusterParameterGroupNotFoundFault

db_cluster_quota_exceeded_fault

See DbClusterQuotaExceededFault

db_cluster_role_already_exists_fault

See DbClusterRoleAlreadyExistsFault

db_cluster_role_not_found_fault

See DbClusterRoleNotFoundFault

db_cluster_role_quota_exceeded_fault

See DbClusterRoleQuotaExceededFault

db_cluster_snapshot_already_exists_fault

See DbClusterSnapshotAlreadyExistsFault

db_cluster_snapshot_not_found_fault

See DbClusterSnapshotNotFoundFault

db_instance_already_exists_fault

See DbInstanceAlreadyExistsFault

db_instance_automated_backup_not_found_fault

See DbInstanceAutomatedBackupNotFoundFault

db_instance_automated_backup_quota_exceeded_fault

See DbInstanceAutomatedBackupQuotaExceededFault

db_instance_not_found_fault

See DbInstanceNotFoundFault

db_instance_role_already_exists_fault

See DbInstanceRoleAlreadyExistsFault

db_instance_role_not_found_fault

See DbInstanceRoleNotFoundFault

db_instance_role_quota_exceeded_fault

See DbInstanceRoleQuotaExceededFault

db_log_file_not_found_fault

See DbLogFileNotFoundFault

db_parameter_group_already_exists_fault

See DbParameterGroupAlreadyExistsFault

db_parameter_group_not_found_fault

See DbParameterGroupNotFoundFault

db_parameter_group_quota_exceeded_fault

See DbParameterGroupQuotaExceededFault

db_proxy_already_exists_fault

See DbProxyAlreadyExistsFault

db_proxy_endpoint_already_exists_fault

See DbProxyEndpointAlreadyExistsFault

db_proxy_endpoint_not_found_fault

See DbProxyEndpointNotFoundFault

db_proxy_endpoint_quota_exceeded_fault

See DbProxyEndpointQuotaExceededFault

db_proxy_not_found_fault

See DbProxyNotFoundFault

db_proxy_quota_exceeded_fault

See DbProxyQuotaExceededFault

db_proxy_target_already_registered_fault

See DbProxyTargetAlreadyRegisteredFault

db_proxy_target_group_not_found_fault

See DbProxyTargetGroupNotFoundFault

db_proxy_target_not_found_fault

See DbProxyTargetNotFoundFault

db_security_group_already_exists_fault

See DbSecurityGroupAlreadyExistsFault

db_security_group_not_found_fault

See DbSecurityGroupNotFoundFault

db_security_group_not_supported_fault

See DbSecurityGroupNotSupportedFault

db_security_group_quota_exceeded_fault

See DbSecurityGroupQuotaExceededFault

db_snapshot_already_exists_fault

See DbSnapshotAlreadyExistsFault

db_snapshot_not_found_fault

See DbSnapshotNotFoundFault

db_subnet_group_already_exists_fault

See DbSubnetGroupAlreadyExistsFault

db_subnet_group_does_not_cover_enough_a_zs

See DbSubnetGroupDoesNotCoverEnoughAZs

db_subnet_group_not_allowed_fault

See DbSubnetGroupNotAllowedFault

db_subnet_group_not_found_fault

See DbSubnetGroupNotFoundFault

db_subnet_group_quota_exceeded_fault

See DbSubnetGroupQuotaExceededFault

db_subnet_quota_exceeded_fault

See DbSubnetQuotaExceededFault

db_upgrade_dependency_failure_fault

See DbUpgradeDependencyFailureFault

domain_not_found_fault

See DomainNotFoundFault

event_subscription_quota_exceeded_fault

See EventSubscriptionQuotaExceededFault

export_task_already_exists_fault

See ExportTaskAlreadyExistsFault

export_task_not_found_fault

See ExportTaskNotFoundFault

global_cluster_already_exists_fault

See GlobalClusterAlreadyExistsFault

global_cluster_not_found_fault

See GlobalClusterNotFoundFault

global_cluster_quota_exceeded_fault

See GlobalClusterQuotaExceededFault

iam_role_missing_permissions_fault

See IamRoleMissingPermissionsFault

iam_role_not_found_fault

See IamRoleNotFoundFault

installation_media_already_exists_fault

See InstallationMediaAlreadyExistsFault

installation_media_not_found_fault

See InstallationMediaNotFoundFault

instance_quota_exceeded_fault

See InstanceQuotaExceededFault

insufficient_available_i_ps_in_subnet_fault

See InsufficientAvailableIPsInSubnetFault

insufficient_db_cluster_capacity_fault

See InsufficientDbClusterCapacityFault

insufficient_db_instance_capacity_fault

See InsufficientDbInstanceCapacityFault

insufficient_storage_cluster_capacity_fault

See InsufficientStorageClusterCapacityFault

invalid_db_cluster_capacity_fault

See InvalidDbClusterCapacityFault

invalid_db_cluster_endpoint_state_fault

See InvalidDbClusterEndpointStateFault

invalid_db_cluster_snapshot_state_fault

See InvalidDbClusterSnapshotStateFault

invalid_db_cluster_state_fault

See InvalidDbClusterStateFault

invalid_db_instance_automated_backup_state_fault

See InvalidDbInstanceAutomatedBackupStateFault

invalid_db_instance_state_fault

See InvalidDbInstanceStateFault

invalid_db_parameter_group_state_fault

See InvalidDbParameterGroupStateFault

invalid_db_proxy_endpoint_state_fault

See InvalidDbProxyEndpointStateFault

invalid_db_proxy_state_fault

See InvalidDbProxyStateFault

invalid_db_security_group_state_fault

See InvalidDbSecurityGroupStateFault

invalid_db_snapshot_state_fault

See InvalidDbSnapshotStateFault

invalid_db_subnet_group_fault

See InvalidDbSubnetGroupFault

invalid_db_subnet_group_state_fault

See InvalidDbSubnetGroupStateFault

invalid_db_subnet_state_fault

See InvalidDbSubnetStateFault

invalid_event_subscription_state_fault

See InvalidEventSubscriptionStateFault

invalid_export_only_fault

See InvalidExportOnlyFault

invalid_export_source_state_fault

See InvalidExportSourceStateFault

invalid_export_task_state_fault

See InvalidExportTaskStateFault

invalid_global_cluster_state_fault

See InvalidGlobalClusterStateFault

invalid_option_group_state_fault

See InvalidOptionGroupStateFault

invalid_restore_fault

See InvalidRestoreFault

invalid_s3_bucket_fault

See InvalidS3BucketFault

invalid_subnet

See InvalidSubnet

invalid_vpc_network_state_fault

See InvalidVpcNetworkStateFault

kms_key_not_accessible_fault

See KmsKeyNotAccessibleFault

option_group_already_exists_fault

See OptionGroupAlreadyExistsFault

option_group_not_found_fault

See OptionGroupNotFoundFault

option_group_quota_exceeded_fault

See OptionGroupQuotaExceededFault

point_in_time_restore_not_enabled_fault

See PointInTimeRestoreNotEnabledFault

provisioned_iops_not_available_in_az_fault

See ProvisionedIopsNotAvailableInAzFault

reserved_db_instance_already_exists_fault

See ReservedDbInstanceAlreadyExistsFault

reserved_db_instance_not_found_fault

See ReservedDbInstanceNotFoundFault

reserved_db_instance_quota_exceeded_fault

See ReservedDbInstanceQuotaExceededFault

reserved_db_instances_offering_not_found_fault

See ReservedDbInstancesOfferingNotFoundFault

resource_not_found_fault

See ResourceNotFoundFault

shared_snapshot_quota_exceeded_fault

See SharedSnapshotQuotaExceededFault

snapshot_quota_exceeded_fault

See SnapshotQuotaExceededFault

sns_invalid_topic_fault

See SnsInvalidTopicFault

sns_no_authorization_fault

See SnsNoAuthorizationFault

sns_topic_arn_not_found_fault

See SnsTopicArnNotFoundFault

source_not_found_fault

See SourceNotFoundFault

storage_quota_exceeded_fault

See StorageQuotaExceededFault

storage_type_not_supported_fault

See StorageTypeNotSupportedFault

subnet_already_in_use

See SubnetAlreadyInUse

subscription_already_exist_fault

See SubscriptionAlreadyExistFault

subscription_category_not_found_fault

See SubscriptionCategoryNotFoundFault

subscription_not_found_fault

See SubscriptionNotFoundFault

Structs

AddRoleToDBClusterError

AddRoleToDBInstanceError

AddSourceIdentifierToSubscriptionError

AddTagsToResourceError

ApplyPendingMaintenanceActionError

AuthorizationAlreadyExistsFault

The specified CIDR IP range or Amazon EC2 security group is already authorized for

AuthorizationNotFoundFault

The specified CIDR IP range or Amazon EC2 security group might not be authorized

AuthorizationQuotaExceededFault

The DB security group authorization quota has been reached.

AuthorizeDBSecurityGroupIngressError

BacktrackDBClusterError

BackupPolicyNotFoundFault

CancelExportTaskError

CertificateNotFoundFault

CopyDBClusterParameterGroupError

CopyDBClusterSnapshotError

CopyDBParameterGroupError

CopyDBSnapshotError

CopyOptionGroupError

CreateCustomAvailabilityZoneError

CreateDBClusterEndpointError

CreateDBClusterError

CreateDBClusterParameterGroupError

CreateDBClusterSnapshotError

CreateDBInstanceError

CreateDBInstanceReadReplicaError

CreateDBParameterGroupError

CreateDBProxyEndpointError

CreateDBProxyError

CreateDBSecurityGroupError

CreateDBSnapshotError

CreateDBSubnetGroupError

CreateEventSubscriptionError

CreateGlobalClusterError

CreateOptionGroupError

CustomAvailabilityZoneAlreadyExistsFault

CustomAvailabilityZoneNotFoundFault

CustomAvailabilityZoneQuotaExceededFault

You have exceeded the maximum number of custom Availability Zones.

DbClusterAlreadyExistsFault

The user already has a DB cluster with the given identifier.

DbClusterBacktrackNotFoundFault

DbClusterEndpointAlreadyExistsFault

The specified custom endpoint can't be created because it already exists.

DbClusterEndpointNotFoundFault

The specified custom endpoint doesn't exist.

DbClusterEndpointQuotaExceededFault

The cluster already has the maximum number of custom endpoints.

DbClusterNotFoundFault

DbClusterParameterGroupNotFoundFault

DbClusterQuotaExceededFault

The user attempted to create a new DB cluster and the user has already reached the

DbClusterRoleAlreadyExistsFault

The specified IAM role Amazon Resource Name (ARN) is already associated with the specified DB cluster.

DbClusterRoleNotFoundFault

The specified IAM role Amazon Resource Name (ARN) isn't associated with the specified DB cluster.

DbClusterRoleQuotaExceededFault

You have exceeded the maximum number of IAM roles that can be associated with the specified DB cluster.

DbClusterSnapshotAlreadyExistsFault

The user already has a DB cluster snapshot with the given identifier.

DbClusterSnapshotNotFoundFault

DbInstanceAlreadyExistsFault

The user already has a DB instance with the given identifier.

DbInstanceAutomatedBackupNotFoundFault

No automated backup for this DB instance was found.

DbInstanceAutomatedBackupQuotaExceededFault

The quota for retained automated backups was exceeded. This prevents you

DbInstanceNotFoundFault

DbInstanceRoleAlreadyExistsFault

The specified RoleArn or FeatureName value is already associated with the DB instance.

DbInstanceRoleNotFoundFault

The specified RoleArn value doesn't match the specified feature for

DbInstanceRoleQuotaExceededFault

You can't associate any more Amazon Web Services Identity and Access Management (IAM) roles with the DB instance because the quota has been reached.

DbLogFileNotFoundFault

DbParameterGroupAlreadyExistsFault

A DB parameter group with the same name exists.

DbParameterGroupNotFoundFault

DbParameterGroupQuotaExceededFault

The request would result in the user exceeding the allowed number of DB parameter

DbProxyAlreadyExistsFault

The specified proxy name must be unique for all proxies owned by your Amazon Web Services account in the specified Amazon Web Services Region.

DbProxyEndpointAlreadyExistsFault

The specified DB proxy endpoint name must be unique for all DB proxy endpoints owned by your Amazon Web Services account in the specified Amazon Web Services Region.

DbProxyEndpointNotFoundFault

The DB proxy endpoint doesn't exist.

DbProxyEndpointQuotaExceededFault

The DB proxy already has the maximum number of endpoints.

DbProxyNotFoundFault

The specified proxy name doesn't correspond to a proxy owned by your Amazon Web Services account in the specified Amazon Web Services Region.

DbProxyQuotaExceededFault

Your Amazon Web Services account already has the maximum number of proxies in the specified Amazon Web Services Region.

DbProxyTargetAlreadyRegisteredFault

The proxy is already associated with the specified RDS DB instance or Aurora DB cluster.

DbProxyTargetGroupNotFoundFault

The specified target group isn't available for a proxy owned by your Amazon Web Services account in the specified Amazon Web Services Region.

DbProxyTargetNotFoundFault

The specified RDS DB instance or Aurora DB cluster isn't available for a proxy owned by your Amazon Web Services account in the specified Amazon Web Services Region.

DbSecurityGroupAlreadyExistsFault

DbSecurityGroupNotFoundFault

DbSecurityGroupNotSupportedFault

A DB security group isn't allowed for this action.

DbSecurityGroupQuotaExceededFault

The request would result in the user exceeding the allowed number of DB security

DbSnapshotAlreadyExistsFault

DbSnapshotNotFoundFault

DbSubnetGroupAlreadyExistsFault

DbSubnetGroupDoesNotCoverEnoughAZs

Subnets in the DB subnet group should cover at least two Availability Zones unless there is only one Availability Zone.

DbSubnetGroupNotAllowedFault

The DBSubnetGroup shouldn't be specified while creating read replicas that lie

DbSubnetGroupNotFoundFault

DbSubnetGroupQuotaExceededFault

The request would result in the user exceeding the allowed number of DB subnet

DbSubnetQuotaExceededFault

The request would result in the user exceeding the allowed number of subnets in a

DbUpgradeDependencyFailureFault

The DB upgrade failed because a resource the DB depends on can't be

DeleteCustomAvailabilityZoneError

DeleteDBClusterEndpointError

DeleteDBClusterError

DeleteDBClusterParameterGroupError

DeleteDBClusterSnapshotError

DeleteDBInstanceAutomatedBackupError

DeleteDBInstanceError

DeleteDBParameterGroupError

DeleteDBProxyEndpointError

DeleteDBProxyError

DeleteDBSecurityGroupError

DeleteDBSnapshotError

DeleteDBSubnetGroupError

DeleteEventSubscriptionError

DeleteGlobalClusterError

DeleteInstallationMediaError

DeleteOptionGroupError

DeregisterDBProxyTargetsError

DescribeAccountAttributesError

DescribeCertificatesError

DescribeCustomAvailabilityZonesError

DescribeDBClusterBacktracksError

DescribeDBClusterEndpointsError

DescribeDBClusterParameterGroupsError

DescribeDBClusterParametersError

DescribeDBClusterSnapshotAttributesError

DescribeDBClusterSnapshotsError

DescribeDBClustersError

DescribeDBEngineVersionsError

DescribeDBInstanceAutomatedBackupsError

DescribeDBInstancesError

DescribeDBLogFilesError

DescribeDBParameterGroupsError

DescribeDBParametersError

DescribeDBProxiesError

DescribeDBProxyEndpointsError

DescribeDBProxyTargetGroupsError

DescribeDBProxyTargetsError

DescribeDBSecurityGroupsError

DescribeDBSnapshotAttributesError

DescribeDBSnapshotsError

DescribeDBSubnetGroupsError

DescribeEngineDefaultClusterParametersError

DescribeEngineDefaultParametersError

DescribeEventCategoriesError

DescribeEventSubscriptionsError

DescribeEventsError

DescribeExportTasksError

DescribeGlobalClustersError

DescribeInstallationMediaError

DescribeOptionGroupOptionsError

DescribeOptionGroupsError

DescribeOrderableDBInstanceOptionsError

DescribePendingMaintenanceActionsError

DescribeReservedDBInstancesError

DescribeReservedDBInstancesOfferingsError

DescribeSourceRegionsError

DescribeValidDBInstanceModificationsError

DomainNotFoundFault

DownloadDBLogFilePortionError

EventSubscriptionQuotaExceededFault

You have reached the maximum number of event subscriptions.

ExportTaskAlreadyExistsFault

You can't start an export task that's already running.

ExportTaskNotFoundFault

The export task doesn't exist.

FailoverDBClusterError

FailoverGlobalClusterError

GlobalClusterAlreadyExistsFault

The GlobalClusterIdentifier already exists. Choose a new global database identifier (unique name) to create a new global database cluster.

GlobalClusterNotFoundFault

The GlobalClusterIdentifier doesn't refer to an existing global database cluster.

GlobalClusterQuotaExceededFault

The number of global database clusters for this account is already at the maximum allowed.

IamRoleMissingPermissionsFault

The IAM role requires additional permissions to export to an Amazon S3 bucket.

IamRoleNotFoundFault

The IAM role is missing for exporting to an Amazon S3 bucket.

ImportInstallationMediaError

InstallationMediaAlreadyExistsFault

The specified installation medium has already been imported.

InstallationMediaNotFoundFault

InstanceQuotaExceededFault

The request would result in the user exceeding the allowed number of DB

InsufficientAvailableIPsInSubnetFault

The requested operation can't be performed because there aren't enough available IP addresses

InsufficientDbClusterCapacityFault

The DB cluster doesn't have enough capacity for the current operation.

InsufficientDbInstanceCapacityFault

The specified DB instance class isn't available in the specified Availability

InsufficientStorageClusterCapacityFault

There is insufficient storage available for the current action. You might be able to

InvalidDbClusterCapacityFault

InvalidDbClusterEndpointStateFault

The requested operation can't be performed on the endpoint while the endpoint is in this state.

InvalidDbClusterSnapshotStateFault

The supplied value isn't a valid DB cluster snapshot state.

InvalidDbClusterStateFault

The requested operation can't be performed while the cluster is in this state.

InvalidDbInstanceAutomatedBackupStateFault

The automated backup is in an invalid state.

InvalidDbInstanceStateFault

The DB instance isn't in a valid state.

InvalidDbParameterGroupStateFault

The DB parameter group is in use or is in an invalid state. If you are attempting

InvalidDbProxyEndpointStateFault

You can't perform this operation while the DB proxy endpoint is in a particular state.

InvalidDbProxyStateFault

The requested operation can't be performed while the proxy is in this state.

InvalidDbSecurityGroupStateFault

The state of the DB security group doesn't allow deletion.

InvalidDbSnapshotStateFault

The state of the DB snapshot doesn't allow deletion.

InvalidDbSubnetGroupFault

The DBSubnetGroup doesn't belong to the same VPC as that of an existing

InvalidDbSubnetGroupStateFault

The DB subnet group cannot be deleted because it's in use.

InvalidDbSubnetStateFault

InvalidEventSubscriptionStateFault

This error can occur if someone else is modifying a subscription. You should retry the action.

InvalidExportOnlyFault

The export is invalid for exporting to an Amazon S3 bucket.

InvalidExportSourceStateFault

The state of the export snapshot is invalid for exporting to an Amazon S3 bucket.

InvalidExportTaskStateFault

You can't cancel an export task that has completed.

InvalidGlobalClusterStateFault

The global cluster is in an invalid state and can't perform the requested operation.

InvalidOptionGroupStateFault

InvalidRestoreFault

Cannot restore from VPC backup to non-VPC DB instance.

InvalidS3BucketFault

The specified Amazon S3 bucket name can't be found or Amazon RDS isn't

InvalidSubnet

The requested subnet is invalid, or multiple subnets were requested that are not all in a common VPC.

InvalidVpcNetworkStateFault

The DB subnet group doesn't cover all Availability Zones after it's

KmsKeyNotAccessibleFault

An error occurred accessing an Amazon Web Services KMS key.

ListTagsForResourceError

ModifyCertificatesError

ModifyCurrentDBClusterCapacityError

ModifyDBClusterEndpointError

ModifyDBClusterError

ModifyDBClusterParameterGroupError

ModifyDBClusterSnapshotAttributeError

ModifyDBInstanceError

ModifyDBParameterGroupError

ModifyDBProxyEndpointError

ModifyDBProxyError

ModifyDBProxyTargetGroupError

ModifyDBSnapshotAttributeError

ModifyDBSnapshotError

ModifyDBSubnetGroupError

ModifyEventSubscriptionError

ModifyGlobalClusterError

ModifyOptionGroupError

OptionGroupAlreadyExistsFault

The option group you are trying to create already exists.

OptionGroupNotFoundFault

The specified option group could not be found.

OptionGroupQuotaExceededFault

The quota of 20 option groups was exceeded for this Amazon Web Services account.

PointInTimeRestoreNotEnabledFault

PromoteReadReplicaDBClusterError

PromoteReadReplicaError

ProvisionedIopsNotAvailableInAzFault

Provisioned IOPS not available in the specified Availability Zone.

PurchaseReservedDBInstancesOfferingError

RebootDBInstanceError

RegisterDBProxyTargetsError

RemoveFromGlobalClusterError

RemoveRoleFromDBClusterError

RemoveRoleFromDBInstanceError

RemoveSourceIdentifierFromSubscriptionError

RemoveTagsFromResourceError

ReservedDbInstanceAlreadyExistsFault

User already has a reservation with the given identifier.

ReservedDbInstanceNotFoundFault

The specified reserved DB Instance not found.

ReservedDbInstanceQuotaExceededFault

Request would exceed the user's DB Instance quota.

ReservedDbInstancesOfferingNotFoundFault

Specified offering does not exist.

ResetDBClusterParameterGroupError

ResetDBParameterGroupError

ResourceNotFoundFault

The specified resource ID was not found.

RestoreDBClusterFromS3Error

RestoreDBClusterFromSnapshotError

RestoreDBClusterToPointInTimeError

RestoreDBInstanceFromDBSnapshotError

RestoreDBInstanceFromS3Error

RestoreDBInstanceToPointInTimeError

RevokeDBSecurityGroupIngressError

SharedSnapshotQuotaExceededFault

You have exceeded the maximum number of accounts that you can share a manual DB snapshot with.

SnapshotQuotaExceededFault

The request would result in the user exceeding the allowed number of DB

SnsInvalidTopicFault

SNS has responded that there is a problem with the SND topic specified.

SnsNoAuthorizationFault

You do not have permission to publish to the SNS topic ARN.

SnsTopicArnNotFoundFault

The SNS topic ARN does not exist.

SourceNotFoundFault

The requested source could not be found.

StartActivityStreamError

StartDBClusterError

StartDBInstanceAutomatedBackupsReplicationError

StartDBInstanceError

StartExportTaskError

StopActivityStreamError

StopDBClusterError

StopDBInstanceAutomatedBackupsReplicationError

StopDBInstanceError

StorageQuotaExceededFault

The request would result in the user exceeding the allowed amount of storage

StorageTypeNotSupportedFault

Storage of the StorageType specified can't be associated

SubnetAlreadyInUse

The DB subnet is already in use in the Availability Zone.

SubscriptionAlreadyExistFault

The supplied subscription name already exists.

SubscriptionCategoryNotFoundFault

The supplied category does not exist.

SubscriptionNotFoundFault

The subscription name does not exist.

Enums

AddRoleToDBClusterErrorKind

AddRoleToDBInstanceErrorKind

AddSourceIdentifierToSubscriptionErrorKind

AddTagsToResourceErrorKind

ApplyPendingMaintenanceActionErrorKind

AuthorizeDBSecurityGroupIngressErrorKind

BacktrackDBClusterErrorKind

CancelExportTaskErrorKind

CopyDBClusterParameterGroupErrorKind

CopyDBClusterSnapshotErrorKind

CopyDBParameterGroupErrorKind

CopyDBSnapshotErrorKind

CopyOptionGroupErrorKind

CreateCustomAvailabilityZoneErrorKind

CreateDBClusterEndpointErrorKind

CreateDBClusterErrorKind

CreateDBClusterParameterGroupErrorKind

CreateDBClusterSnapshotErrorKind

CreateDBInstanceErrorKind

CreateDBInstanceReadReplicaErrorKind

CreateDBParameterGroupErrorKind

CreateDBProxyEndpointErrorKind

CreateDBProxyErrorKind

CreateDBSecurityGroupErrorKind

CreateDBSnapshotErrorKind

CreateDBSubnetGroupErrorKind

CreateEventSubscriptionErrorKind

CreateGlobalClusterErrorKind

CreateOptionGroupErrorKind

DeleteCustomAvailabilityZoneErrorKind

DeleteDBClusterEndpointErrorKind

DeleteDBClusterErrorKind

DeleteDBClusterParameterGroupErrorKind

DeleteDBClusterSnapshotErrorKind

DeleteDBInstanceAutomatedBackupErrorKind

DeleteDBInstanceErrorKind

DeleteDBParameterGroupErrorKind

DeleteDBProxyEndpointErrorKind

DeleteDBProxyErrorKind

DeleteDBSecurityGroupErrorKind

DeleteDBSnapshotErrorKind

DeleteDBSubnetGroupErrorKind

DeleteEventSubscriptionErrorKind

DeleteGlobalClusterErrorKind

DeleteInstallationMediaErrorKind

DeleteOptionGroupErrorKind

DeregisterDBProxyTargetsErrorKind

DescribeAccountAttributesErrorKind

DescribeCertificatesErrorKind

DescribeCustomAvailabilityZonesErrorKind

DescribeDBClusterBacktracksErrorKind

DescribeDBClusterEndpointsErrorKind

DescribeDBClusterParameterGroupsErrorKind

DescribeDBClusterParametersErrorKind

DescribeDBClusterSnapshotAttributesErrorKind

DescribeDBClusterSnapshotsErrorKind

DescribeDBClustersErrorKind

DescribeDBEngineVersionsErrorKind

DescribeDBInstanceAutomatedBackupsErrorKind

DescribeDBInstancesErrorKind

DescribeDBLogFilesErrorKind

DescribeDBParameterGroupsErrorKind

DescribeDBParametersErrorKind

DescribeDBProxiesErrorKind

DescribeDBProxyEndpointsErrorKind

DescribeDBProxyTargetGroupsErrorKind

DescribeDBProxyTargetsErrorKind

DescribeDBSecurityGroupsErrorKind

DescribeDBSnapshotAttributesErrorKind

DescribeDBSnapshotsErrorKind

DescribeDBSubnetGroupsErrorKind

DescribeEngineDefaultClusterParametersErrorKind

DescribeEngineDefaultParametersErrorKind

DescribeEventCategoriesErrorKind

DescribeEventSubscriptionsErrorKind

DescribeEventsErrorKind

DescribeExportTasksErrorKind

DescribeGlobalClustersErrorKind

DescribeInstallationMediaErrorKind

DescribeOptionGroupOptionsErrorKind

DescribeOptionGroupsErrorKind

DescribeOrderableDBInstanceOptionsErrorKind

DescribePendingMaintenanceActionsErrorKind

DescribeReservedDBInstancesErrorKind

DescribeReservedDBInstancesOfferingsErrorKind

DescribeSourceRegionsErrorKind

DescribeValidDBInstanceModificationsErrorKind

DownloadDBLogFilePortionErrorKind

FailoverDBClusterErrorKind

FailoverGlobalClusterErrorKind

ImportInstallationMediaErrorKind

ListTagsForResourceErrorKind

ModifyCertificatesErrorKind

ModifyCurrentDBClusterCapacityErrorKind

ModifyDBClusterEndpointErrorKind

ModifyDBClusterErrorKind

ModifyDBClusterParameterGroupErrorKind

ModifyDBClusterSnapshotAttributeErrorKind

ModifyDBInstanceErrorKind

ModifyDBParameterGroupErrorKind

ModifyDBProxyEndpointErrorKind

ModifyDBProxyErrorKind

ModifyDBProxyTargetGroupErrorKind

ModifyDBSnapshotAttributeErrorKind

ModifyDBSnapshotErrorKind

ModifyDBSubnetGroupErrorKind

ModifyEventSubscriptionErrorKind

ModifyGlobalClusterErrorKind

ModifyOptionGroupErrorKind

PromoteReadReplicaDBClusterErrorKind

PromoteReadReplicaErrorKind

PurchaseReservedDBInstancesOfferingErrorKind

RebootDBInstanceErrorKind

RegisterDBProxyTargetsErrorKind

RemoveFromGlobalClusterErrorKind

RemoveRoleFromDBClusterErrorKind

RemoveRoleFromDBInstanceErrorKind

RemoveSourceIdentifierFromSubscriptionErrorKind

RemoveTagsFromResourceErrorKind

ResetDBClusterParameterGroupErrorKind

ResetDBParameterGroupErrorKind

RestoreDBClusterFromS3ErrorKind

RestoreDBClusterFromSnapshotErrorKind

RestoreDBClusterToPointInTimeErrorKind

RestoreDBInstanceFromDBSnapshotErrorKind

RestoreDBInstanceFromS3ErrorKind

RestoreDBInstanceToPointInTimeErrorKind

RevokeDBSecurityGroupIngressErrorKind

StartActivityStreamErrorKind

StartDBClusterErrorKind

StartDBInstanceAutomatedBackupsReplicationErrorKind

StartDBInstanceErrorKind

StartExportTaskErrorKind

StopActivityStreamErrorKind

StopDBClusterErrorKind

StopDBInstanceAutomatedBackupsReplicationErrorKind

StopDBInstanceErrorKind