Struct aws_sdk_rds::input::create_db_instance_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for CreateDbInstanceInput

Implementations

impl Builder

pub fn db_name(self, input: impl Into<String>) -> Self

The meaning of this parameter differs according to the database engine you use.

MySQL

The name of the database to create when the DB instance is created. If this parameter isn't specified, no database is created in the DB instance.

Constraints:

Must contain 1 to 64 letters or numbers.

Must begin with a letter. Subsequent characters can be letters, underscores, or digits (0-9).

Can't be a word reserved by the specified database engine

MariaDB

The name of the database to create when the DB instance is created. If this parameter isn't specified, no database is created in the DB instance.

Constraints:

Must contain 1 to 64 letters or numbers.

Must begin with a letter. Subsequent characters can be letters, underscores, or digits (0-9).

Can't be a word reserved by the specified database engine

PostgreSQL

The name of the database to create when the DB instance is created. If this parameter isn't specified, a database named postgres is created in the DB instance.

Constraints:

Must contain 1 to 63 letters, numbers, or underscores.

Must begin with a letter. Subsequent characters can be letters, underscores, or digits (0-9).

Can't be a word reserved by the specified database engine

Oracle

The Oracle System ID (SID) of the created DB instance. If you specify null, the default value ORCL is used. You can't specify the string NULL, or any other reserved word, for DBName.

Default: ORCL

Constraints:

Can't be longer than 8 characters

SQL Server

Not applicable. Must be null.

Amazon Aurora MySQL

The name of the database to create when the primary DB instance of the Aurora MySQL DB cluster is created. If this parameter isn't specified for an Aurora MySQL DB cluster, no database is created in the DB cluster.

Constraints:

It must contain 1 to 64 alphanumeric characters.

It can't be a word reserved by the database engine.

Amazon Aurora PostgreSQL

The name of the database to create when the primary DB instance of the Aurora PostgreSQL DB cluster is created. If this parameter isn't specified for an Aurora PostgreSQL DB cluster, a database named postgres is created in the DB cluster.

Constraints:

It must contain 1 to 63 alphanumeric characters.

It must begin with a letter or an underscore. Subsequent characters can be letters, underscores, or digits (0 to 9).

It can't be a word reserved by the database engine.

pub fn set_db_name(self, input: Option<String>) -> Self

pub fn db_instance_identifier(self, input: impl Into<String>) -> Self

The DB instance identifier. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

Example: mydbinstance

pub fn set_db_instance_identifier(self, input: Option<String>) -> Self

pub fn allocated_storage(self, input: i32) -> Self

The amount of storage (in gibibytes) to allocate for the DB instance.

Type: Integer

Amazon Aurora

Not applicable. Aurora cluster volumes automatically grow as the amount of data in your database increases, though you are only charged for the space that you use in an Aurora cluster volume.

MySQL

Constraints to the amount of storage for each storage type are the following:

General Purpose (SSD) storage (gp2): Must be an integer from 20 to 65536.

Provisioned IOPS storage (io1): Must be an integer from 100 to 65536.

Magnetic storage (standard): Must be an integer from 5 to 3072.

MariaDB

Constraints to the amount of storage for each storage type are the following:

General Purpose (SSD) storage (gp2): Must be an integer from 20 to 65536.

Provisioned IOPS storage (io1): Must be an integer from 100 to 65536.

Magnetic storage (standard): Must be an integer from 5 to 3072.

PostgreSQL

Constraints to the amount of storage for each storage type are the following:

General Purpose (SSD) storage (gp2): Must be an integer from 20 to 65536.

Provisioned IOPS storage (io1): Must be an integer from 100 to 65536.

Magnetic storage (standard): Must be an integer from 5 to 3072.

Oracle

Constraints to the amount of storage for each storage type are the following:

General Purpose (SSD) storage (gp2): Must be an integer from 20 to 65536.

Provisioned IOPS storage (io1): Must be an integer from 100 to 65536.

Magnetic storage (standard): Must be an integer from 10 to 3072.

SQL Server

Constraints to the amount of storage for each storage type are the following:

General Purpose (SSD) storage (gp2):

Enterprise and Standard editions: Must be an integer from 200 to 16384.

Web and Express editions: Must be an integer from 20 to 16384.

Provisioned IOPS storage (io1):

Enterprise and Standard editions: Must be an integer from 200 to 16384.

Web and Express editions: Must be an integer from 100 to 16384.

Magnetic storage (standard):

Enterprise and Standard editions: Must be an integer from 200 to 1024.

Web and Express editions: Must be an integer from 20 to 1024.

pub fn set_allocated_storage(self, input: Option<i32>) -> Self

pub fn db_instance_class(self, input: impl Into<String>) -> Self

The compute and memory capacity of the DB instance, for example, db.m4.large. Not all DB instance classes are available in all Amazon Web Services Regions, or for all database engines. For the full list of DB instance classes, and availability for your engine, see DB Instance Class in the Amazon RDS User Guide.

pub fn set_db_instance_class(self, input: Option<String>) -> Self

pub fn engine(self, input: impl Into<String>) -> Self

The name of the database engine to be used for this instance.

Not every database engine is available for every Amazon Web Services Region.

Valid Values:

aurora (for MySQL 5.6-compatible Aurora)

aurora-mysql (for MySQL 5.7-compatible Aurora)

aurora-postgresql

mariadb

mysql

oracle-ee

oracle-ee-cdb

oracle-se2

oracle-se2-cdb

postgres

sqlserver-ee

sqlserver-se

sqlserver-ex

sqlserver-web

pub fn set_engine(self, input: Option<String>) -> Self

pub fn master_username(self, input: impl Into<String>) -> Self

The name for the master user.

Amazon Aurora

Not applicable. The name for the master user is managed by the DB cluster.

MariaDB

Constraints:

Required for MariaDB.

Must be 1 to 16 letters or numbers.

Can't be a reserved word for the chosen database engine.

Microsoft SQL Server

Constraints:

Required for SQL Server.

Must be 1 to 128 letters or numbers.

The first character must be a letter.

Can't be a reserved word for the chosen database engine.

MySQL

Constraints:

Required for MySQL.

Must be 1 to 16 letters or numbers.

First character must be a letter.

Can't be a reserved word for the chosen database engine.

Oracle

Constraints:

Required for Oracle.

Must be 1 to 30 letters or numbers.

First character must be a letter.

Can't be a reserved word for the chosen database engine.

PostgreSQL

Constraints:

Required for PostgreSQL.

Must be 1 to 63 letters or numbers.

First character must be a letter.

Can't be a reserved word for the chosen database engine.

pub fn set_master_username(self, input: Option<String>) -> Self

pub fn master_user_password(self, input: impl Into<String>) -> Self

The password for the master user. The password can include any printable ASCII character except "/", """, or "@".

Amazon Aurora

Not applicable. The password for the master user is managed by the DB cluster.

MariaDB

Constraints: Must contain from 8 to 41 characters.

Microsoft SQL Server

Constraints: Must contain from 8 to 128 characters.

MySQL

Constraints: Must contain from 8 to 41 characters.

Oracle

Constraints: Must contain from 8 to 30 characters.

PostgreSQL

Constraints: Must contain from 8 to 128 characters.

pub fn set_master_user_password(self, input: Option<String>) -> Self

pub fn db_security_groups(self, input: impl Into<String>) -> Self

pub fn set_db_security_groups(self, input: Option<Vec<String>>) -> Self

pub fn vpc_security_group_ids(self, input: impl Into<String>) -> Self

pub fn set_vpc_security_group_ids(self, input: Option<Vec<String>>) -> Self

pub fn availability_zone(self, input: impl Into<String>) -> Self

The Availability Zone (AZ) where the database will be created. For information on Amazon Web Services Regions and Availability Zones, see Regions and Availability Zones.

Default: A random, system-chosen Availability Zone in the endpoint's Amazon Web Services Region.

Example: us-east-1d

Constraint: The AvailabilityZone parameter can't be specified if the DB instance is a Multi-AZ deployment. The specified Availability Zone must be in the same Amazon Web Services Region as the current endpoint.

If you're creating a DB instance in an RDS on VMware environment, specify the identifier of the custom Availability Zone to create the DB instance in.

For more information about RDS on VMware, see the RDS on VMware User Guide.

pub fn set_availability_zone(self, input: Option<String>) -> Self

pub fn db_subnet_group_name(self, input: impl Into<String>) -> Self

A DB subnet group to associate with this DB instance.

If there is no DB subnet group, then it is a non-VPC DB instance.

pub fn set_db_subnet_group_name(self, input: Option<String>) -> Self

pub fn preferred_maintenance_window(self, input: impl Into<String>) -> Self

The time range each week during which system maintenance can occur, in Universal Coordinated Time (UTC). For more information, see Amazon RDS Maintenance Window.

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Web Services Region, occurring on a random day of the week.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Constraints: Minimum 30-minute window.

pub fn set_preferred_maintenance_window(self, input: Option<String>) -> Self

pub fn db_parameter_group_name(self, input: impl Into<String>) -> Self

The name of the DB parameter group to associate with this DB instance. If you do not specify a value, then the default DB parameter group for the specified DB engine and version is used.

Constraints:

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

pub fn set_db_parameter_group_name(self, input: Option<String>) -> Self

pub fn backup_retention_period(self, input: i32) -> Self

The number of days for which automated backups are retained. Setting this parameter to a positive number enables backups. Setting this parameter to 0 disables automated backups.

Amazon Aurora

Not applicable. The retention period for automated backups is managed by the DB cluster.

Default: 1

Constraints:

Must be a value from 0 to 35

Can't be set to 0 if the DB instance is a source to read replicas

pub fn set_backup_retention_period(self, input: Option<i32>) -> Self

pub fn preferred_backup_window(self, input: impl Into<String>) -> Self

The daily time range during which automated backups are created if automated backups are enabled, using the BackupRetentionPeriod parameter. The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Web Services Region. For more information, see Backup window in the Amazon RDS User Guide.

Amazon Aurora

Not applicable. The daily time range for creating automated backups is managed by the DB cluster.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

pub fn set_preferred_backup_window(self, input: Option<String>) -> Self

pub fn port(self, input: i32) -> Self

The port number on which the database accepts connections.

MySQL

Default: 3306

Valid values: 1150-65535

Type: Integer

MariaDB

Default: 3306

Valid values: 1150-65535

Type: Integer

PostgreSQL

Default: 5432

Valid values: 1150-65535

Type: Integer

Oracle

Default: 1521

Valid values: 1150-65535

SQL Server

Default: 1433

Valid values: 1150-65535 except 1234, 1434, 3260, 3343, 3389, 47001, and 49152-49156.

Amazon Aurora

Default: 3306

Valid values: 1150-65535

Type: Integer

pub fn set_port(self, input: Option<i32>) -> Self

pub fn multi_az(self, input: bool) -> Self

A value that indicates whether the DB instance is a Multi-AZ deployment. You can't set the AvailabilityZone parameter if the DB instance is a Multi-AZ deployment.

pub fn set_multi_az(self, input: Option<bool>) -> Self

pub fn engine_version(self, input: impl Into<String>) -> Self

The version number of the database engine to use.

For a list of valid engine versions, use the DescribeDBEngineVersions action.

The following are the database engines and links to information about the major and minor versions that are available with Amazon RDS. Not every database engine is available for every Amazon Web Services Region.

Amazon Aurora

Not applicable. The version number of the database engine to be used by the DB instance is managed by the DB cluster.

MariaDB

See MariaDB on Amazon RDS Versions in the Amazon RDS User Guide.

Microsoft SQL Server

See Microsoft SQL Server Versions on Amazon RDS in the Amazon RDS User Guide.

MySQL

See MySQL on Amazon RDS Versions in the Amazon RDS User Guide.

Oracle

See Oracle Database Engine Release Notes in the Amazon RDS User Guide.

PostgreSQL

See Amazon RDS for PostgreSQL versions and extensions in the Amazon RDS User Guide.

pub fn set_engine_version(self, input: Option<String>) -> Self

pub fn auto_minor_version_upgrade(self, input: bool) -> Self

A value that indicates whether minor engine upgrades are applied automatically to the DB instance during the maintenance window. By default, minor engine upgrades are applied automatically.

pub fn set_auto_minor_version_upgrade(self, input: Option<bool>) -> Self

pub fn license_model(self, input: impl Into<String>) -> Self

License model information for this DB instance.

Valid values: license-included | bring-your-own-license | general-public-license

pub fn set_license_model(self, input: Option<String>) -> Self

pub fn iops(self, input: i32) -> Self

The amount of Provisioned IOPS (input/output operations per second) to be initially allocated for the DB instance. For information about valid Iops values, see Amazon RDS Provisioned IOPS Storage to Improve Performance in the Amazon RDS User Guide.

Constraints: For MariaDB, MySQL, Oracle, and PostgreSQL DB instances, must be a multiple between .5 and 50 of the storage amount for the DB instance. For SQL Server DB instances, must be a multiple between 1 and 50 of the storage amount for the DB instance.

pub fn set_iops(self, input: Option<i32>) -> Self

pub fn option_group_name(self, input: impl Into<String>) -> Self

A value that indicates that the DB instance should be associated with the specified option group.

Permanent options, such as the TDE option for Oracle Advanced Security TDE, can't be removed from an option group. Also, that option group can't be removed from a DB instance once it is associated with a DB instance

pub fn set_option_group_name(self, input: Option<String>) -> Self

pub fn character_set_name(self, input: impl Into<String>) -> Self

For supported engines, indicates that the DB instance should be associated with the specified CharacterSet.

Amazon Aurora

Not applicable. The character set is managed by the DB cluster. For more information, see CreateDBCluster.

pub fn set_character_set_name(self, input: Option<String>) -> Self

pub fn nchar_character_set_name(self, input: impl Into<String>) -> Self

The name of the NCHAR character set for the Oracle DB instance.

pub fn set_nchar_character_set_name(self, input: Option<String>) -> Self

pub fn publicly_accessible(self, input: bool) -> Self

A value that indicates whether the DB instance is publicly accessible.

When the DB instance is publicly accessible, its DNS endpoint resolves to the private IP address from within the DB instance's VPC, and to the public IP address from outside of the DB instance's VPC. Access to the DB instance is ultimately controlled by the security group it uses, and that public access is not permitted if the security group assigned to the DB instance doesn't permit it.

When the DB instance isn't publicly accessible, it is an internal DB instance with a DNS name that resolves to a private IP address.

Default: The default behavior varies depending on whether DBSubnetGroupName is specified.

If DBSubnetGroupName isn't specified, and PubliclyAccessible isn't specified, the following applies:

If the default VPC in the target region doesn’t have an Internet gateway attached to it, the DB instance is private.

If the default VPC in the target region has an Internet gateway attached to it, the DB instance is public.

If DBSubnetGroupName is specified, and PubliclyAccessible isn't specified, the following applies:

If the subnets are part of a VPC that doesn’t have an Internet gateway attached to it, the DB instance is private.

If the subnets are part of a VPC that has an Internet gateway attached to it, the DB instance is public.

pub fn set_publicly_accessible(self, input: Option<bool>) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn db_cluster_identifier(self, input: impl Into<String>) -> Self

The identifier of the DB cluster that the instance will belong to.

pub fn set_db_cluster_identifier(self, input: Option<String>) -> Self

pub fn storage_type(self, input: impl Into<String>) -> Self

Specifies the storage type to be associated with the DB instance.

Valid values: standard | gp2 | io1

If you specify io1, you must also include a value for the Iops parameter.

Default: io1 if the Iops parameter is specified, otherwise gp2

pub fn set_storage_type(self, input: Option<String>) -> Self

pub fn tde_credential_arn(self, input: impl Into<String>) -> Self

The ARN from the key store with which to associate the instance for TDE encryption.

pub fn set_tde_credential_arn(self, input: Option<String>) -> Self

pub fn tde_credential_password(self, input: impl Into<String>) -> Self

The password for the given ARN from the key store in order to access the device.

pub fn set_tde_credential_password(self, input: Option<String>) -> Self

pub fn storage_encrypted(self, input: bool) -> Self

A value that indicates whether the DB instance is encrypted. By default, it isn't encrypted.

Amazon Aurora

Not applicable. The encryption for DB instances is managed by the DB cluster.

pub fn set_storage_encrypted(self, input: Option<bool>) -> Self

pub fn kms_key_id(self, input: impl Into<String>) -> Self

The Amazon Web Services KMS key identifier for an encrypted DB instance.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK). To use a CMK in a different Amazon Web Services account, specify the key ARN or alias ARN.

Amazon Aurora

Not applicable. The Amazon Web Services KMS key identifier is managed by the DB cluster. For more information, see CreateDBCluster.

If StorageEncrypted is enabled, and you do not specify a value for the KmsKeyId parameter, then Amazon RDS uses your default CMK. There is a default CMK for your Amazon Web Services account. Your Amazon Web Services account has a different default CMK for each Amazon Web Services Region.

pub fn set_kms_key_id(self, input: Option<String>) -> Self

pub fn domain(self, input: impl Into<String>) -> Self

The Active Directory directory ID to create the DB instance in. Currently, only MySQL, Microsoft SQL Server, Oracle, and PostgreSQL DB instances can be created in an Active Directory Domain.

For more information, see Kerberos Authentication in the Amazon RDS User Guide.

pub fn set_domain(self, input: Option<String>) -> Self

pub fn copy_tags_to_snapshot(self, input: bool) -> Self

A value that indicates whether to copy tags from the DB instance to snapshots of the DB instance. By default, tags are not copied.

Amazon Aurora

Not applicable. Copying tags to snapshots is managed by the DB cluster. Setting this value for an Aurora DB instance has no effect on the DB cluster setting.

pub fn set_copy_tags_to_snapshot(self, input: Option<bool>) -> Self

pub fn monitoring_interval(self, input: i32) -> Self

The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the DB instance. To disable collecting Enhanced Monitoring metrics, specify 0. The default is 0.

If MonitoringRoleArn is specified, then you must also set MonitoringInterval to a value other than 0.

Valid Values: 0, 1, 5, 10, 15, 30, 60

pub fn set_monitoring_interval(self, input: Option<i32>) -> Self

pub fn monitoring_role_arn(self, input: impl Into<String>) -> Self

The ARN for the IAM role that permits RDS to send enhanced monitoring metrics to Amazon CloudWatch Logs. For example, arn:aws:iam:123456789012:role/emaccess. For information on creating a monitoring role, go to Setting Up and Enabling Enhanced Monitoring in the Amazon RDS User Guide.

If MonitoringInterval is set to a value other than 0, then you must supply a MonitoringRoleArn value.

pub fn set_monitoring_role_arn(self, input: Option<String>) -> Self

pub fn domain_iam_role_name(self, input: impl Into<String>) -> Self

Specify the name of the IAM role to be used when making API calls to the Directory Service.

pub fn set_domain_iam_role_name(self, input: Option<String>) -> Self

pub fn promotion_tier(self, input: i32) -> Self

A value that specifies the order in which an Aurora Replica is promoted to the primary instance after a failure of the existing primary instance. For more information, see Fault Tolerance for an Aurora DB Cluster in the Amazon Aurora User Guide.

Default: 1

Valid Values: 0 - 15

pub fn set_promotion_tier(self, input: Option<i32>) -> Self

pub fn timezone(self, input: impl Into<String>) -> Self

The time zone of the DB instance. The time zone parameter is currently supported only by Microsoft SQL Server.

pub fn set_timezone(self, input: Option<String>) -> Self

pub fn enable_iam_database_authentication(self, input: bool) -> Self

A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

This setting doesn't apply to Amazon Aurora. Mapping Amazon Web Services IAM accounts to database accounts is managed by the DB cluster.

For more information, see IAM Database Authentication for MySQL and PostgreSQL in the Amazon RDS User Guide.

pub fn set_enable_iam_database_authentication(self, input: Option<bool>) -> Self

pub fn enable_performance_insights(self, input: bool) -> Self

A value that indicates whether to enable Performance Insights for the DB instance.

For more information, see Using Amazon Performance Insights in the Amazon Relational Database Service User Guide.

pub fn set_enable_performance_insights(self, input: Option<bool>) -> Self

pub fn performance_insights_kms_key_id(self, input: impl Into<String>) -> Self

The Amazon Web Services KMS key identifier for encryption of Performance Insights data.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK).

If you do not specify a value for PerformanceInsightsKMSKeyId, then Amazon RDS uses your default CMK. There is a default CMK for your Amazon Web Services account. Your Amazon Web Services account has a different default CMK for each Amazon Web Services Region.

pub fn set_performance_insights_kms_key_id(self, input: Option<String>) -> Self

pub fn performance_insights_retention_period(self, input: i32) -> Self

The amount of time, in days, to retain Performance Insights data. Valid values are 7 or 731 (2 years).

pub fn set_performance_insights_retention_period(
self,
input: Option<i32>
) -> Self

pub fn enable_cloudwatch_logs_exports(self, input: impl Into<String>) -> Self

pub fn set_enable_cloudwatch_logs_exports(
self,
input: Option<Vec<String>>
) -> Self

pub fn processor_features(self, input: impl Into<ProcessorFeature>) -> Self

pub fn set_processor_features(
self,
input: Option<Vec<ProcessorFeature>>
) -> Self

pub fn deletion_protection(self, input: bool) -> Self

A value that indicates whether the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled. For more information, see Deleting a DB Instance.

Amazon Aurora

Not applicable. You can enable or disable deletion protection for the DB cluster. For more information, see CreateDBCluster. DB instances in a DB cluster can be deleted even when deletion protection is enabled for the DB cluster.

pub fn set_deletion_protection(self, input: Option<bool>) -> Self

pub fn max_allocated_storage(self, input: i32) -> Self

The upper limit to which Amazon RDS can automatically scale the storage of the DB instance.

For more information about this setting, including limitations that apply to it, see Managing capacity automatically with Amazon RDS storage autoscaling in the Amazon RDS User Guide.

pub fn set_max_allocated_storage(self, input: Option<i32>) -> Self

pub fn enable_customer_owned_ip(self, input: bool) -> Self

A value that indicates whether to enable a customer-owned IP address (CoIP) for an RDS on Outposts DB instance.

A CoIP provides local or external connectivity to resources in your Outpost subnets through your on-premises network. For some use cases, a CoIP can provide lower latency for connections to the DB instance from outside of its virtual private cloud (VPC) on your local network.

For more information about RDS on Outposts, see Working with Amazon RDS on Amazon Web Services Outposts in the Amazon RDS User Guide.

For more information about CoIPs, see Customer-owned IP addresses in the Amazon Web Services Outposts User Guide.

pub fn set_enable_customer_owned_ip(self, input: Option<bool>) -> Self

pub fn build(self) -> Result<CreateDbInstanceInput, BuildError>

Consumes the builder and constructs a CreateDbInstanceInput