Struct aws_sdk_rds::input::create_db_instance_read_replica_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for CreateDbInstanceReadReplicaInput

Implementations

impl Builder

pub fn db_instance_identifier(self, input: impl Into<String>) -> Self

The DB instance identifier of the read replica. This identifier is the unique key that identifies a DB instance. This parameter is stored as a lowercase string.

pub fn set_db_instance_identifier(self, input: Option<String>) -> Self

pub fn source_db_instance_identifier(self, input: impl Into<String>) -> Self

The identifier of the DB instance that will act as the source for the read replica. Each DB instance can have up to five read replicas.

Constraints:

Must be the identifier of an existing MySQL, MariaDB, Oracle, PostgreSQL, or SQL Server DB instance.

Can specify a DB instance that is a MySQL read replica only if the source is running MySQL 5.6 or later.

For the limitations of Oracle read replicas, see Read Replica Limitations with Oracle in the Amazon RDS User Guide.

For the limitations of SQL Server read replicas, see Read Replica Limitations with Microsoft SQL Server in the Amazon RDS User Guide.

Can specify a PostgreSQL DB instance only if the source is running PostgreSQL 9.3.5 or later (9.4.7 and higher for cross-region replication).

The specified DB instance must have automatic backups enabled, that is, its backup retention period must be greater than 0.

If the source DB instance is in the same Amazon Web Services Region as the read replica, specify a valid DB instance identifier.

If the source DB instance is in a different Amazon Web Services Region from the read replica, specify a valid DB instance ARN. For more information, see Constructing an ARN for Amazon RDS in the Amazon RDS User Guide. This doesn't apply to SQL Server, which doesn't support cross-region replicas.

pub fn set_source_db_instance_identifier(self, input: Option<String>) -> Self

pub fn db_instance_class(self, input: impl Into<String>) -> Self

The compute and memory capacity of the read replica, for example, db.m4.large. Not all DB instance classes are available in all Amazon Web Services Regions, or for all database engines. For the full list of DB instance classes, and availability for your engine, see DB Instance Class in the Amazon RDS User Guide.

Default: Inherits from the source DB instance.

pub fn set_db_instance_class(self, input: Option<String>) -> Self

pub fn availability_zone(self, input: impl Into<String>) -> Self

The Availability Zone (AZ) where the read replica will be created.

Default: A random, system-chosen Availability Zone in the endpoint's Amazon Web Services Region.

Example: us-east-1d

pub fn set_availability_zone(self, input: Option<String>) -> Self

pub fn port(self, input: i32) -> Self

The port number that the DB instance uses for connections.

Default: Inherits from the source DB instance

Valid Values: 1150-65535

pub fn set_port(self, input: Option<i32>) -> Self

pub fn multi_az(self, input: bool) -> Self

A value that indicates whether the read replica is in a Multi-AZ deployment.

You can create a read replica as a Multi-AZ DB instance. RDS creates a standby of your replica in another Availability Zone for failover support for the replica. Creating your read replica as a Multi-AZ DB instance is independent of whether the source database is a Multi-AZ DB instance.

pub fn set_multi_az(self, input: Option<bool>) -> Self

pub fn auto_minor_version_upgrade(self, input: bool) -> Self

A value that indicates whether minor engine upgrades are applied automatically to the read replica during the maintenance window.

Default: Inherits from the source DB instance

pub fn set_auto_minor_version_upgrade(self, input: Option<bool>) -> Self

pub fn iops(self, input: i32) -> Self

The amount of Provisioned IOPS (input/output operations per second) to be initially allocated for the DB instance.

pub fn set_iops(self, input: Option<i32>) -> Self

pub fn option_group_name(self, input: impl Into<String>) -> Self

The option group the DB instance is associated with. If omitted, the option group associated with the source instance is used.

For SQL Server, you must use the option group associated with the source instance.

pub fn set_option_group_name(self, input: Option<String>) -> Self

pub fn db_parameter_group_name(self, input: impl Into<String>) -> Self

The name of the DB parameter group to associate with this DB instance.

If you do not specify a value for DBParameterGroupName, then Amazon RDS uses the DBParameterGroup of source DB instance for a same region read replica, or the default DBParameterGroup for the specified DB engine for a cross region read replica.

Currently, specifying a parameter group for this operation is only supported for Oracle DB instances.

Constraints:

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

pub fn set_db_parameter_group_name(self, input: Option<String>) -> Self

pub fn publicly_accessible(self, input: bool) -> Self

A value that indicates whether the DB instance is publicly accessible.

When the DB instance is publicly accessible, its DNS endpoint resolves to the private IP address from within the DB instance's VPC, and to the public IP address from outside of the DB instance's VPC. Access to the DB instance is ultimately controlled by the security group it uses, and that public access is not permitted if the security group assigned to the DB instance doesn't permit it.

When the DB instance isn't publicly accessible, it is an internal DB instance with a DNS name that resolves to a private IP address.

For more information, see CreateDBInstance.

pub fn set_publicly_accessible(self, input: Option<bool>) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn db_subnet_group_name(self, input: impl Into<String>) -> Self

Specifies a DB subnet group for the DB instance. The new DB instance is created in the VPC associated with the DB subnet group. If no DB subnet group is specified, then the new DB instance isn't created in a VPC.

Constraints:

Can only be specified if the source DB instance identifier specifies a DB instance in another Amazon Web Services Region.

If supplied, must match the name of an existing DBSubnetGroup.

The specified DB subnet group must be in the same Amazon Web Services Region in which the operation is running.

All read replicas in one Amazon Web Services Region that are created from the same source DB instance must either:>

Specify DB subnet groups from the same VPC. All these read replicas are created in the same VPC.

Not specify a DB subnet group. All these read replicas are created outside of any VPC.

Example: mySubnetgroup

pub fn set_db_subnet_group_name(self, input: Option<String>) -> Self

pub fn vpc_security_group_ids(self, input: impl Into<String>) -> Self

pub fn set_vpc_security_group_ids(self, input: Option<Vec<String>>) -> Self

pub fn storage_type(self, input: impl Into<String>) -> Self

Specifies the storage type to be associated with the read replica.

Valid values: standard | gp2 | io1

If you specify io1, you must also include a value for the Iops parameter.

Default: io1 if the Iops parameter is specified, otherwise gp2

pub fn set_storage_type(self, input: Option<String>) -> Self

pub fn copy_tags_to_snapshot(self, input: bool) -> Self

A value that indicates whether to copy all tags from the read replica to snapshots of the read replica. By default, tags are not copied.

pub fn set_copy_tags_to_snapshot(self, input: Option<bool>) -> Self

pub fn monitoring_interval(self, input: i32) -> Self

The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the read replica. To disable collecting Enhanced Monitoring metrics, specify 0. The default is 0.

If MonitoringRoleArn is specified, then you must also set MonitoringInterval to a value other than 0.

Valid Values: 0, 1, 5, 10, 15, 30, 60

pub fn set_monitoring_interval(self, input: Option<i32>) -> Self

pub fn monitoring_role_arn(self, input: impl Into<String>) -> Self

The ARN for the IAM role that permits RDS to send enhanced monitoring metrics to Amazon CloudWatch Logs. For example, arn:aws:iam:123456789012:role/emaccess. For information on creating a monitoring role, go to To create an IAM role for Amazon RDS Enhanced Monitoring in the Amazon RDS User Guide.

If MonitoringInterval is set to a value other than 0, then you must supply a MonitoringRoleArn value.

pub fn set_monitoring_role_arn(self, input: Option<String>) -> Self

pub fn kms_key_id(self, input: impl Into<String>) -> Self

The Amazon Web Services KMS key identifier for an encrypted read replica.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS CMK.

If you create an encrypted read replica in the same Amazon Web Services Region as the source DB instance, then do not specify a value for this parameter. A read replica in the same Region is always encrypted with the same Amazon Web Services KMS CMK as the source DB instance.

If you create an encrypted read replica in a different Amazon Web Services Region, then you must specify a Amazon Web Services KMS key identifier for the destination Amazon Web Services Region. Amazon Web Services KMS CMKs are specific to the Amazon Web Services Region that they are created in, and you can't use CMKs from one Amazon Web Services Region in another Amazon Web Services Region.

You can't create an encrypted read replica from an unencrypted DB instance.

pub fn set_kms_key_id(self, input: Option<String>) -> Self

pub fn pre_signed_url(self, input: impl Into<String>) -> Self

The URL that contains a Signature Version 4 signed request for the CreateDBInstanceReadReplica API action in the source Amazon Web Services Region that contains the source DB instance.

You must specify this parameter when you create an encrypted read replica from another Amazon Web Services Region by using the Amazon RDS API. Don't specify PreSignedUrl when you are creating an encrypted read replica in the same Amazon Web Services Region.

The presigned URL must be a valid request for the CreateDBInstanceReadReplica API action that can be executed in the source Amazon Web Services Region that contains the encrypted source DB instance. The presigned URL request must contain the following parameter values:

DestinationRegion - The Amazon Web Services Region that the encrypted read replica is created in. This Amazon Web Services Region is the same one where the CreateDBInstanceReadReplica action is called that contains this presigned URL.

For example, if you create an encrypted DB instance in the us-west-1 Amazon Web Services Region, from a source DB instance in the us-east-2 Amazon Web Services Region, then you call the CreateDBInstanceReadReplica action in the us-east-1 Amazon Web Services Region and provide a presigned URL that contains a call to the CreateDBInstanceReadReplica action in the us-west-2 Amazon Web Services Region. For this example, the DestinationRegion in the presigned URL must be set to the us-east-1 Amazon Web Services Region.

KmsKeyId - The Amazon Web Services KMS key identifier for the key to use to encrypt the read replica in the destination Amazon Web Services Region. This is the same identifier for both the CreateDBInstanceReadReplica action that is called in the destination Amazon Web Services Region, and the action contained in the presigned URL.

SourceDBInstanceIdentifier - The DB instance identifier for the encrypted DB instance to be replicated. This identifier must be in the Amazon Resource Name (ARN) format for the source Amazon Web Services Region. For example, if you are creating an encrypted read replica from a DB instance in the us-west-2 Amazon Web Services Region, then your SourceDBInstanceIdentifier looks like the following example: arn:aws:rds:us-west-2:123456789012:instance:mysql-instance1-20161115.

To learn how to generate a Signature Version 4 signed request, see Authenticating Requests: Using Query Parameters (Amazon Web Services Signature Version 4) and Signature Version 4 Signing Process.

If you are using an Amazon Web Services SDK tool or the CLI, you can specify SourceRegion (or --source-region for the CLI) instead of specifying PreSignedUrl manually. Specifying SourceRegion autogenerates a presigned URL that is a valid request for the operation that can be executed in the source Amazon Web Services Region.

SourceRegion isn't supported for SQL Server, because SQL Server on Amazon RDS doesn't support cross-region read replicas.

pub fn set_pre_signed_url(self, input: Option<String>) -> Self

pub fn enable_iam_database_authentication(self, input: bool) -> Self

A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

For more information about IAM database authentication, see IAM Database Authentication for MySQL and PostgreSQL in the Amazon RDS User Guide.

pub fn set_enable_iam_database_authentication(self, input: Option<bool>) -> Self

pub fn enable_performance_insights(self, input: bool) -> Self

A value that indicates whether to enable Performance Insights for the read replica.

For more information, see Using Amazon Performance Insights in the Amazon RDS User Guide.

pub fn set_enable_performance_insights(self, input: Option<bool>) -> Self

pub fn performance_insights_kms_key_id(self, input: impl Into<String>) -> Self

The Amazon Web Services KMS key identifier for encryption of Performance Insights data.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK).

If you do not specify a value for PerformanceInsightsKMSKeyId, then Amazon RDS uses your default CMK. There is a default CMK for your Amazon Web Services account. Your Amazon Web Services account has a different default CMK for each Amazon Web Services Region.

pub fn set_performance_insights_kms_key_id(self, input: Option<String>) -> Self

pub fn performance_insights_retention_period(self, input: i32) -> Self

The amount of time, in days, to retain Performance Insights data. Valid values are 7 or 731 (2 years).

pub fn set_performance_insights_retention_period(
self,
input: Option<i32>
) -> Self

pub fn enable_cloudwatch_logs_exports(self, input: impl Into<String>) -> Self

pub fn set_enable_cloudwatch_logs_exports(
self,
input: Option<Vec<String>>
) -> Self

pub fn processor_features(self, input: impl Into<ProcessorFeature>) -> Self

pub fn set_processor_features(
self,
input: Option<Vec<ProcessorFeature>>
) -> Self

pub fn use_default_processor_features(self, input: bool) -> Self

A value that indicates whether the DB instance class of the DB instance uses its default processor features.

pub fn set_use_default_processor_features(self, input: Option<bool>) -> Self

pub fn deletion_protection(self, input: bool) -> Self

A value that indicates whether the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled. For more information, see Deleting a DB Instance.

pub fn set_deletion_protection(self, input: Option<bool>) -> Self

pub fn domain(self, input: impl Into<String>) -> Self

The Active Directory directory ID to create the DB instance in. Currently, only MySQL, Microsoft SQL Server, Oracle, and PostgreSQL DB instances can be created in an Active Directory Domain.

For more information, see Kerberos Authentication in the Amazon RDS User Guide.

pub fn set_domain(self, input: Option<String>) -> Self

pub fn domain_iam_role_name(self, input: impl Into<String>) -> Self

Specify the name of the IAM role to be used when making API calls to the Directory Service.

pub fn set_domain_iam_role_name(self, input: Option<String>) -> Self

pub fn replica_mode(self, input: ReplicaMode) -> Self

The open mode of the replica database: mounted or read-only.

This parameter is only supported for Oracle DB instances.

Mounted DB replicas are included in Oracle Enterprise Edition. The main use case for mounted replicas is cross-Region disaster recovery. The primary database doesn't use Active Data Guard to transmit information to the mounted replica. Because it doesn't accept user connections, a mounted replica can't serve a read-only workload.

You can create a combination of mounted and read-only DB replicas for the same primary DB instance. For more information, see Working with Oracle Read Replicas for Amazon RDS in the Amazon RDS User Guide.

pub fn set_replica_mode(self, input: Option<ReplicaMode>) -> Self

pub fn max_allocated_storage(self, input: i32) -> Self

The upper limit to which Amazon RDS can automatically scale the storage of the DB instance.

For more information about this setting, including limitations that apply to it, see Managing capacity automatically with Amazon RDS storage autoscaling in the Amazon RDS User Guide.

pub fn set_max_allocated_storage(self, input: Option<i32>) -> Self

pub fn build(self) -> Result<CreateDbInstanceReadReplicaInput, BuildError>

Consumes the builder and constructs a CreateDbInstanceReadReplicaInput