Struct aws_sdk_rds::input::modify_db_cluster_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for ModifyDbClusterInput

Implementations

impl Builder

pub fn db_cluster_identifier(self, input: impl Into<String>) -> Self

The DB cluster identifier for the cluster being modified. This parameter isn't case-sensitive.

Constraints: This identifier must match the identifier of an existing DB cluster.

pub fn set_db_cluster_identifier(self, input: Option<String>) -> Self

pub fn new_db_cluster_identifier(self, input: impl Into<String>) -> Self

The new DB cluster identifier for the DB cluster when renaming a DB cluster. This value is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens

The first character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

Example: my-cluster2

pub fn set_new_db_cluster_identifier(self, input: Option<String>) -> Self

pub fn apply_immediately(self, input: bool) -> Self

A value that indicates whether the modifications in this request and any pending modifications are asynchronously applied as soon as possible, regardless of the PreferredMaintenanceWindow setting for the DB cluster. If this parameter is disabled, changes to the DB cluster are applied during the next maintenance window.

The ApplyImmediately parameter only affects the EnableIAMDatabaseAuthentication, MasterUserPassword, and NewDBClusterIdentifier values. If the ApplyImmediately parameter is disabled, then changes to the EnableIAMDatabaseAuthentication, MasterUserPassword, and NewDBClusterIdentifier values are applied during the next maintenance window. All other changes are applied immediately, regardless of the value of the ApplyImmediately parameter.

By default, this parameter is disabled.

pub fn set_apply_immediately(self, input: Option<bool>) -> Self

pub fn backup_retention_period(self, input: i32) -> Self

The number of days for which automated backups are retained. You must specify a minimum value of 1.

Default: 1

Constraints:

Must be a value from 1 to 35

pub fn set_backup_retention_period(self, input: Option<i32>) -> Self

pub fn db_cluster_parameter_group_name(self, input: impl Into<String>) -> Self

The name of the DB cluster parameter group to use for the DB cluster.

pub fn set_db_cluster_parameter_group_name(self, input: Option<String>) -> Self

pub fn vpc_security_group_ids(self, input: impl Into<String>) -> Self

pub fn set_vpc_security_group_ids(self, input: Option<Vec<String>>) -> Self

pub fn port(self, input: i32) -> Self

The port number on which the DB cluster accepts connections.

Constraints: Value must be 1150-65535

Default: The same port as the original DB cluster.

pub fn set_port(self, input: Option<i32>) -> Self

pub fn master_user_password(self, input: impl Into<String>) -> Self

The new password for the master database user. This password can contain any printable ASCII character except "/", """, or "@".

Constraints: Must contain from 8 to 41 characters.

pub fn set_master_user_password(self, input: Option<String>) -> Self

pub fn option_group_name(self, input: impl Into<String>) -> Self

A value that indicates that the DB cluster should be associated with the specified option group. Changing this parameter doesn't result in an outage except in the following case, and the change is applied during the next maintenance window unless the ApplyImmediately is enabled for this request. If the parameter change results in an option group that enables OEM, this change can cause a brief (sub-second) period during which new connections are rejected but existing connections are not interrupted.

Permanent options can't be removed from an option group. The option group can't be removed from a DB cluster once it is associated with a DB cluster.

pub fn set_option_group_name(self, input: Option<String>) -> Self

pub fn preferred_backup_window(self, input: impl Into<String>) -> Self

The daily time range during which automated backups are created if automated backups are enabled, using the BackupRetentionPeriod parameter.

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Web Services Region. To view the time blocks available, see Backup window in the Amazon Aurora User Guide.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

pub fn set_preferred_backup_window(self, input: Option<String>) -> Self

pub fn preferred_maintenance_window(self, input: impl Into<String>) -> Self

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Web Services Region, occurring on a random day of the week. To see the time blocks available, see Adjusting the Preferred DB Cluster Maintenance Window in the Amazon Aurora User Guide.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Constraints: Minimum 30-minute window.

pub fn set_preferred_maintenance_window(self, input: Option<String>) -> Self

pub fn enable_iam_database_authentication(self, input: bool) -> Self

A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

For more information, see IAM Database Authentication in the Amazon Aurora User Guide.

pub fn set_enable_iam_database_authentication(self, input: Option<bool>) -> Self

pub fn backtrack_window(self, input: i64) -> Self

The target backtrack window, in seconds. To disable backtracking, set this value to 0.

Currently, Backtrack is only supported for Aurora MySQL DB clusters.

Default: 0

Constraints:

If specified, this value must be set to a number from 0 to 259,200 (72 hours).

pub fn set_backtrack_window(self, input: Option<i64>) -> Self

pub fn cloudwatch_logs_export_configuration(
self,
input: CloudwatchLogsExportConfiguration
) -> Self

The configuration setting for the log types to be enabled for export to CloudWatch Logs for a specific DB cluster.

pub fn set_cloudwatch_logs_export_configuration(
self,
input: Option<CloudwatchLogsExportConfiguration>
) -> Self

pub fn engine_version(self, input: impl Into<String>) -> Self

The version number of the database engine to which you want to upgrade. Changing this parameter results in an outage. The change is applied during the next maintenance window unless ApplyImmediately is enabled.

To list all of the available engine versions for aurora (for MySQL 5.6-compatible Aurora), use the following command:

aws rds describe-db-engine-versions --engine aurora --query "DBEngineVersions[].EngineVersion"

To list all of the available engine versions for aurora-mysql (for MySQL 5.7-compatible Aurora), use the following command:

aws rds describe-db-engine-versions --engine aurora-mysql --query "DBEngineVersions[].EngineVersion"

To list all of the available engine versions for aurora-postgresql, use the following command:

aws rds describe-db-engine-versions --engine aurora-postgresql --query "DBEngineVersions[].EngineVersion"

pub fn set_engine_version(self, input: Option<String>) -> Self

pub fn allow_major_version_upgrade(self, input: bool) -> Self

A value that indicates whether major version upgrades are allowed.

Constraints: You must allow major version upgrades when specifying a value for the EngineVersion parameter that is a different major version than the DB cluster's current version.

pub fn set_allow_major_version_upgrade(self, input: Option<bool>) -> Self

pub fn db_instance_parameter_group_name(self, input: impl Into<String>) -> Self

The name of the DB parameter group to apply to all instances of the DB cluster.

When you apply a parameter group using the DBInstanceParameterGroupName parameter, the DB cluster isn't rebooted automatically. Also, parameter changes aren't applied during the next maintenance window but instead are applied immediately.

Default: The existing name setting

Constraints:

The DB parameter group must be in the same DB parameter group family as this DB cluster.

The DBInstanceParameterGroupName parameter is only valid in combination with the AllowMajorVersionUpgrade parameter.

pub fn set_db_instance_parameter_group_name(self, input: Option<String>) -> Self

pub fn domain(self, input: impl Into<String>) -> Self

The Active Directory directory ID to move the DB cluster to. Specify none to remove the cluster from its current domain. The domain must be created prior to this operation.

For more information, see Kerberos Authentication in the Amazon Aurora User Guide.

pub fn set_domain(self, input: Option<String>) -> Self

pub fn domain_iam_role_name(self, input: impl Into<String>) -> Self

Specify the name of the IAM role to be used when making API calls to the Directory Service.

pub fn set_domain_iam_role_name(self, input: Option<String>) -> Self

pub fn scaling_configuration(self, input: ScalingConfiguration) -> Self

The scaling properties of the DB cluster. You can only modify scaling properties for DB clusters in serverless DB engine mode.

pub fn set_scaling_configuration(
self,
input: Option<ScalingConfiguration>
) -> Self

pub fn deletion_protection(self, input: bool) -> Self

A value that indicates whether the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled.

pub fn set_deletion_protection(self, input: Option<bool>) -> Self

pub fn enable_http_endpoint(self, input: bool) -> Self

A value that indicates whether to enable the HTTP endpoint for an Aurora Serverless DB cluster. By default, the HTTP endpoint is disabled.

When enabled, the HTTP endpoint provides a connectionless web service API for running SQL queries on the Aurora Serverless DB cluster. You can also query your database from inside the RDS console with the query editor.

For more information, see Using the Data API for Aurora Serverless in the Amazon Aurora User Guide.

pub fn set_enable_http_endpoint(self, input: Option<bool>) -> Self

pub fn copy_tags_to_snapshot(self, input: bool) -> Self

A value that indicates whether to copy all tags from the DB cluster to snapshots of the DB cluster. The default is not to copy them.

pub fn set_copy_tags_to_snapshot(self, input: Option<bool>) -> Self

pub fn enable_global_write_forwarding(self, input: bool) -> Self

A value that indicates whether to enable this DB cluster to forward write operations to the primary cluster of an Aurora global database (GlobalCluster). By default, write operations are not allowed on Aurora DB clusters that are secondary clusters in an Aurora global database.

You can set this value only on Aurora DB clusters that are members of an Aurora global database. With this parameter enabled, a secondary cluster can forward writes to the current primary cluster and the resulting changes are replicated back to this cluster. For the primary DB cluster of an Aurora global database, this value is used immediately if the primary is demoted by the FailoverGlobalCluster API operation, but it does nothing until then.

pub fn set_enable_global_write_forwarding(self, input: Option<bool>) -> Self

pub fn build(self) -> Result<ModifyDbClusterInput, BuildError>

Consumes the builder and constructs a ModifyDbClusterInput