Struct aws_sdk_rds::input::restore_db_instance_to_point_in_time_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for RestoreDbInstanceToPointInTimeInput

Implementations

impl Builder

pub fn source_db_instance_identifier(self, input: impl Into<String>) -> Self

The identifier of the source DB instance from which to restore.

Constraints:

Must match the identifier of an existing DB instance.

pub fn set_source_db_instance_identifier(self, input: Option<String>) -> Self

The identifier of the source DB instance from which to restore.

Constraints:

Must match the identifier of an existing DB instance.

pub fn target_db_instance_identifier(self, input: impl Into<String>) -> Self

The name of the new DB instance to be created.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

pub fn set_target_db_instance_identifier(self, input: Option<String>) -> Self

The name of the new DB instance to be created.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

pub fn restore_time(self, input: Instant) -> Self

The date and time to restore from.

Valid Values: Value must be a time in Universal Coordinated Time (UTC) format

Constraints:

Must be before the latest restorable time for the DB instance

Can't be specified if the UseLatestRestorableTime parameter is enabled

Example: 2009-09-07T23:45:00Z

pub fn set_restore_time(self, input: Option<Instant>) -> Self

The date and time to restore from.

Valid Values: Value must be a time in Universal Coordinated Time (UTC) format

Constraints:

Must be before the latest restorable time for the DB instance

Can't be specified if the UseLatestRestorableTime parameter is enabled

Example: 2009-09-07T23:45:00Z

pub fn use_latest_restorable_time(self, input: bool) -> Self

A value that indicates whether the DB instance is restored from the latest backup time. By default, the DB instance isn't restored from the latest backup time.

Constraints: Can't be specified if the RestoreTime parameter is provided.

pub fn set_use_latest_restorable_time(self, input: Option<bool>) -> Self

A value that indicates whether the DB instance is restored from the latest backup time. By default, the DB instance isn't restored from the latest backup time.

Constraints: Can't be specified if the RestoreTime parameter is provided.

pub fn db_instance_class(self, input: impl Into<String>) -> Self

The compute and memory capacity of the Amazon RDS DB instance, for example, db.m4.large. Not all DB instance classes are available in all Amazon Web Services Regions, or for all database engines. For the full list of DB instance classes, and availability for your engine, see DB Instance Class in the Amazon RDS User Guide.

Default: The same DBInstanceClass as the original DB instance.

pub fn set_db_instance_class(self, input: Option<String>) -> Self

The compute and memory capacity of the Amazon RDS DB instance, for example, db.m4.large. Not all DB instance classes are available in all Amazon Web Services Regions, or for all database engines. For the full list of DB instance classes, and availability for your engine, see DB Instance Class in the Amazon RDS User Guide.

Default: The same DBInstanceClass as the original DB instance.

pub fn port(self, input: i32) -> Self

The port number on which the database accepts connections.

Constraints: Value must be 1150-65535

Default: The same port as the original DB instance.

pub fn set_port(self, input: Option<i32>) -> Self

The port number on which the database accepts connections.

Constraints: Value must be 1150-65535

Default: The same port as the original DB instance.

pub fn availability_zone(self, input: impl Into<String>) -> Self

The Availability Zone (AZ) where the DB instance will be created.

Default: A random, system-chosen Availability Zone.

Constraint: You can't specify the AvailabilityZone parameter if the DB instance is a Multi-AZ deployment.

Example: us-east-1a

pub fn set_availability_zone(self, input: Option<String>) -> Self

The Availability Zone (AZ) where the DB instance will be created.

Default: A random, system-chosen Availability Zone.

Constraint: You can't specify the AvailabilityZone parameter if the DB instance is a Multi-AZ deployment.

Example: us-east-1a

pub fn db_subnet_group_name(self, input: impl Into<String>) -> Self

The DB subnet group name to use for the new instance.

Constraints: If supplied, must match the name of an existing DBSubnetGroup.

Example: mySubnetgroup

pub fn set_db_subnet_group_name(self, input: Option<String>) -> Self

The DB subnet group name to use for the new instance.

Constraints: If supplied, must match the name of an existing DBSubnetGroup.

Example: mySubnetgroup

pub fn multi_az(self, input: bool) -> Self

A value that indicates whether the DB instance is a Multi-AZ deployment.

This setting doesn't apply to RDS Custom.

Constraint: You can't specify the AvailabilityZone parameter if the DB instance is a Multi-AZ deployment.

pub fn set_multi_az(self, input: Option<bool>) -> Self

A value that indicates whether the DB instance is a Multi-AZ deployment.

This setting doesn't apply to RDS Custom.

Constraint: You can't specify the AvailabilityZone parameter if the DB instance is a Multi-AZ deployment.

pub fn publicly_accessible(self, input: bool) -> Self

A value that indicates whether the DB instance is publicly accessible.

When the DB instance is publicly accessible, its DNS endpoint resolves to the private IP address from within the DB instance's VPC, and to the public IP address from outside of the DB instance's VPC. Access to the DB instance is ultimately controlled by the security group it uses, and that public access is not permitted if the security group assigned to the DB instance doesn't permit it.

When the DB instance isn't publicly accessible, it is an internal DB instance with a DNS name that resolves to a private IP address.

For more information, see CreateDBInstance.

pub fn set_publicly_accessible(self, input: Option<bool>) -> Self

A value that indicates whether the DB instance is publicly accessible.

When the DB instance is publicly accessible, its DNS endpoint resolves to the private IP address from within the DB instance's VPC, and to the public IP address from outside of the DB instance's VPC. Access to the DB instance is ultimately controlled by the security group it uses, and that public access is not permitted if the security group assigned to the DB instance doesn't permit it.

When the DB instance isn't publicly accessible, it is an internal DB instance with a DNS name that resolves to a private IP address.

For more information, see CreateDBInstance.

pub fn auto_minor_version_upgrade(self, input: bool) -> Self

A value that indicates whether minor version upgrades are applied automatically to the DB instance during the maintenance window.

This setting doesn't apply to RDS Custom.

pub fn set_auto_minor_version_upgrade(self, input: Option<bool>) -> Self

A value that indicates whether minor version upgrades are applied automatically to the DB instance during the maintenance window.

This setting doesn't apply to RDS Custom.

pub fn license_model(self, input: impl Into<String>) -> Self

License model information for the restored DB instance.

This setting doesn't apply to RDS Custom.

Default: Same as source.

Valid values: license-included | bring-your-own-license | general-public-license

pub fn set_license_model(self, input: Option<String>) -> Self

License model information for the restored DB instance.

This setting doesn't apply to RDS Custom.

Default: Same as source.

Valid values: license-included | bring-your-own-license | general-public-license

pub fn db_name(self, input: impl Into<String>) -> Self

The database name for the restored DB instance.

This parameter isn't supported for the MySQL or MariaDB engines. It also doesn't apply to RDS Custom.

pub fn set_db_name(self, input: Option<String>) -> Self

The database name for the restored DB instance.

This parameter isn't supported for the MySQL or MariaDB engines. It also doesn't apply to RDS Custom.

pub fn engine(self, input: impl Into<String>) -> Self

The database engine to use for the new instance.

This setting doesn't apply to RDS Custom.

Default: The same as source

Constraint: Must be compatible with the engine of the source

Valid Values:

mariadb

mysql

oracle-ee

oracle-ee-cdb

oracle-se2

oracle-se2-cdb

postgres

sqlserver-ee

sqlserver-se

sqlserver-ex

sqlserver-web

pub fn set_engine(self, input: Option<String>) -> Self

The database engine to use for the new instance.

This setting doesn't apply to RDS Custom.

Default: The same as source

Constraint: Must be compatible with the engine of the source

Valid Values:

mariadb

mysql

oracle-ee

oracle-ee-cdb

oracle-se2

oracle-se2-cdb

postgres

sqlserver-ee

sqlserver-se

sqlserver-ex

sqlserver-web

pub fn iops(self, input: i32) -> Self

The amount of Provisioned IOPS (input/output operations per second) to be initially allocated for the DB instance.

Constraints: Must be an integer greater than 1000.

SQL Server

Setting the IOPS value for the SQL Server database engine isn't supported.

pub fn set_iops(self, input: Option<i32>) -> Self

The amount of Provisioned IOPS (input/output operations per second) to be initially allocated for the DB instance.

Constraints: Must be an integer greater than 1000.

SQL Server

Setting the IOPS value for the SQL Server database engine isn't supported.

pub fn option_group_name(self, input: impl Into<String>) -> Self

The name of the option group to be used for the restored DB instance.

Permanent options, such as the TDE option for Oracle Advanced Security TDE, can't be removed from an option group, and that option group can't be removed from a DB instance after it is associated with a DB instance

This setting doesn't apply to RDS Custom.

pub fn set_option_group_name(self, input: Option<String>) -> Self

The name of the option group to be used for the restored DB instance.

Permanent options, such as the TDE option for Oracle Advanced Security TDE, can't be removed from an option group, and that option group can't be removed from a DB instance after it is associated with a DB instance

This setting doesn't apply to RDS Custom.

pub fn copy_tags_to_snapshot(self, input: bool) -> Self

A value that indicates whether to copy all tags from the restored DB instance to snapshots of the DB instance. By default, tags are not copied.

pub fn set_copy_tags_to_snapshot(self, input: Option<bool>) -> Self

A value that indicates whether to copy all tags from the restored DB instance to snapshots of the DB instance. By default, tags are not copied.

pub fn tags(self, input: impl Into<Tag>) -> Self

Appends an item to tags.

To override the contents of this collection use set_tags.

A list of tags. For more information, see Tagging Amazon RDS Resources in the Amazon RDS User Guide.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

A list of tags. For more information, see Tagging Amazon RDS Resources in the Amazon RDS User Guide.

pub fn storage_type(self, input: impl Into<String>) -> Self

Specifies the storage type to be associated with the DB instance.

Valid values: standard | gp2 | io1

If you specify io1, you must also include a value for the Iops parameter.

Default: io1 if the Iops parameter is specified, otherwise gp2

pub fn set_storage_type(self, input: Option<String>) -> Self

Specifies the storage type to be associated with the DB instance.

Valid values: standard | gp2 | io1

If you specify io1, you must also include a value for the Iops parameter.

Default: io1 if the Iops parameter is specified, otherwise gp2

pub fn tde_credential_arn(self, input: impl Into<String>) -> Self

The ARN from the key store with which to associate the instance for TDE encryption.

This setting doesn't apply to RDS Custom.

pub fn set_tde_credential_arn(self, input: Option<String>) -> Self

The ARN from the key store with which to associate the instance for TDE encryption.

This setting doesn't apply to RDS Custom.

pub fn tde_credential_password(self, input: impl Into<String>) -> Self

The password for the given ARN from the key store in order to access the device.

This setting doesn't apply to RDS Custom.

pub fn set_tde_credential_password(self, input: Option<String>) -> Self

The password for the given ARN from the key store in order to access the device.

This setting doesn't apply to RDS Custom.

pub fn vpc_security_group_ids(self, input: impl Into<String>) -> Self

Appends an item to vpc_security_group_ids.

To override the contents of this collection use set_vpc_security_group_ids.

A list of EC2 VPC security groups to associate with this DB instance.

Default: The default EC2 VPC security group for the DB subnet group's VPC.

pub fn set_vpc_security_group_ids(self, input: Option<Vec<String>>) -> Self

A list of EC2 VPC security groups to associate with this DB instance.

Default: The default EC2 VPC security group for the DB subnet group's VPC.

pub fn domain(self, input: impl Into<String>) -> Self

Specify the Active Directory directory ID to restore the DB instance in. Create the domain before running this command. Currently, you can create only the MySQL, Microsoft SQL Server, Oracle, and PostgreSQL DB instances in an Active Directory Domain.

This setting doesn't apply to RDS Custom.

For more information, see Kerberos Authentication in the Amazon RDS User Guide.

pub fn set_domain(self, input: Option<String>) -> Self

Specify the Active Directory directory ID to restore the DB instance in. Create the domain before running this command. Currently, you can create only the MySQL, Microsoft SQL Server, Oracle, and PostgreSQL DB instances in an Active Directory Domain.

This setting doesn't apply to RDS Custom.

For more information, see Kerberos Authentication in the Amazon RDS User Guide.

pub fn domain_iam_role_name(self, input: impl Into<String>) -> Self

Specify the name of the IAM role to be used when making API calls to the Directory Service.

This setting doesn't apply to RDS Custom.

pub fn set_domain_iam_role_name(self, input: Option<String>) -> Self

Specify the name of the IAM role to be used when making API calls to the Directory Service.

This setting doesn't apply to RDS Custom.

pub fn enable_iam_database_authentication(self, input: bool) -> Self

A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

This setting doesn't apply to RDS Custom.

For more information about IAM database authentication, see IAM Database Authentication for MySQL and PostgreSQL in the Amazon RDS User Guide.

pub fn set_enable_iam_database_authentication(self, input: Option<bool>) -> Self

A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

This setting doesn't apply to RDS Custom.

For more information about IAM database authentication, see IAM Database Authentication for MySQL and PostgreSQL in the Amazon RDS User Guide.

pub fn enable_cloudwatch_logs_exports(self, input: impl Into<String>) -> Self

Appends an item to enable_cloudwatch_logs_exports.

To override the contents of this collection use set_enable_cloudwatch_logs_exports.

The list of logs that the restored DB instance is to export to CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see Publishing Database Logs to Amazon CloudWatch Logs in the Amazon RDS User Guide.

This setting doesn't apply to RDS Custom.

pub fn set_enable_cloudwatch_logs_exports(
self,
input: Option<Vec<String>>
) -> Self

The list of logs that the restored DB instance is to export to CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see Publishing Database Logs to Amazon CloudWatch Logs in the Amazon RDS User Guide.

This setting doesn't apply to RDS Custom.

pub fn processor_features(self, input: impl Into<ProcessorFeature>) -> Self

Appends an item to processor_features.

To override the contents of this collection use set_processor_features.

The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

This setting doesn't apply to RDS Custom.

pub fn set_processor_features(
self,
input: Option<Vec<ProcessorFeature>>
) -> Self

The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

This setting doesn't apply to RDS Custom.

pub fn use_default_processor_features(self, input: bool) -> Self

A value that indicates whether the DB instance class of the DB instance uses its default processor features.

This setting doesn't apply to RDS Custom.

pub fn set_use_default_processor_features(self, input: Option<bool>) -> Self

A value that indicates whether the DB instance class of the DB instance uses its default processor features.

This setting doesn't apply to RDS Custom.

pub fn db_parameter_group_name(self, input: impl Into<String>) -> Self

The name of the DB parameter group to associate with this DB instance.

If you do not specify a value for DBParameterGroupName, then the default DBParameterGroup for the specified DB engine is used.

This setting doesn't apply to RDS Custom.

Constraints:

If supplied, must match the name of an existing DBParameterGroup.

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

pub fn set_db_parameter_group_name(self, input: Option<String>) -> Self

The name of the DB parameter group to associate with this DB instance.

If you do not specify a value for DBParameterGroupName, then the default DBParameterGroup for the specified DB engine is used.

This setting doesn't apply to RDS Custom.

Constraints:

If supplied, must match the name of an existing DBParameterGroup.

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

pub fn deletion_protection(self, input: bool) -> Self

A value that indicates whether the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled. For more information, see Deleting a DB Instance.

pub fn set_deletion_protection(self, input: Option<bool>) -> Self

A value that indicates whether the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled. For more information, see Deleting a DB Instance.

pub fn source_dbi_resource_id(self, input: impl Into<String>) -> Self

The resource ID of the source DB instance from which to restore.

pub fn set_source_dbi_resource_id(self, input: Option<String>) -> Self

The resource ID of the source DB instance from which to restore.

pub fn max_allocated_storage(self, input: i32) -> Self

The upper limit in gibibytes (GiB) to which Amazon RDS can automatically scale the storage of the DB instance.

For more information about this setting, including limitations that apply to it, see Managing capacity automatically with Amazon RDS storage autoscaling in the Amazon RDS User Guide.

This setting doesn't apply to RDS Custom.

pub fn set_max_allocated_storage(self, input: Option<i32>) -> Self

The upper limit in gibibytes (GiB) to which Amazon RDS can automatically scale the storage of the DB instance.

For more information about this setting, including limitations that apply to it, see Managing capacity automatically with Amazon RDS storage autoscaling in the Amazon RDS User Guide.

This setting doesn't apply to RDS Custom.

pub fn source_db_instance_automated_backups_arn(
self,
input: impl Into<String>
) -> Self

The Amazon Resource Name (ARN) of the replicated automated backups from which to restore, for example, arn:aws:rds:useast-1:123456789012:auto-backup:ab-L2IJCEXJP7XQ7HOJ4SIEXAMPLE.

This setting doesn't apply to RDS Custom.

pub fn set_source_db_instance_automated_backups_arn(
self,
input: Option<String>
) -> Self

The Amazon Resource Name (ARN) of the replicated automated backups from which to restore, for example, arn:aws:rds:useast-1:123456789012:auto-backup:ab-L2IJCEXJP7XQ7HOJ4SIEXAMPLE.

This setting doesn't apply to RDS Custom.

pub fn enable_customer_owned_ip(self, input: bool) -> Self

A value that indicates whether to enable a customer-owned IP address (CoIP) for an RDS on Outposts DB instance.

A CoIP provides local or external connectivity to resources in your Outpost subnets through your on-premises network. For some use cases, a CoIP can provide lower latency for connections to the DB instance from outside of its virtual private cloud (VPC) on your local network.

This setting doesn't apply to RDS Custom.

For more information about RDS on Outposts, see Working with Amazon RDS on Amazon Web Services Outposts in the Amazon RDS User Guide.

For more information about CoIPs, see Customer-owned IP addresses in the Amazon Web Services Outposts User Guide.

pub fn set_enable_customer_owned_ip(self, input: Option<bool>) -> Self

A value that indicates whether to enable a customer-owned IP address (CoIP) for an RDS on Outposts DB instance.

A CoIP provides local or external connectivity to resources in your Outpost subnets through your on-premises network. For some use cases, a CoIP can provide lower latency for connections to the DB instance from outside of its virtual private cloud (VPC) on your local network.

This setting doesn't apply to RDS Custom.

For more information about RDS on Outposts, see Working with Amazon RDS on Amazon Web Services Outposts in the Amazon RDS User Guide.

For more information about CoIPs, see Customer-owned IP addresses in the Amazon Web Services Outposts User Guide.

pub fn custom_iam_instance_profile(self, input: impl Into<String>) -> Self

The instance profile associated with the underlying Amazon EC2 instance of an RDS Custom DB instance. The instance profile must meet the following requirements:

The profile must exist in your account.

The profile must have an IAM role that Amazon EC2 has permissions to assume.

The instance profile name and the associated IAM role name must start with the prefix AWSRDSCustom.

For the list of permissions required for the IAM role, see Configure IAM and your VPC in the Amazon Relational Database Service User Guide.

This setting is required for RDS Custom.

pub fn set_custom_iam_instance_profile(self, input: Option<String>) -> Self

The instance profile associated with the underlying Amazon EC2 instance of an RDS Custom DB instance. The instance profile must meet the following requirements:

The profile must exist in your account.

The profile must have an IAM role that Amazon EC2 has permissions to assume.

The instance profile name and the associated IAM role name must start with the prefix AWSRDSCustom.

For the list of permissions required for the IAM role, see Configure IAM and your VPC in the Amazon Relational Database Service User Guide.

This setting is required for RDS Custom.

pub fn build(self) -> Result<RestoreDbInstanceToPointInTimeInput, BuildError>

Consumes the builder and constructs a RestoreDbInstanceToPointInTimeInput