Struct aws_sdk_rds::input::RestoreDbInstanceToPointInTimeInput

#[non_exhaustive]
pub struct RestoreDbInstanceToPointInTimeInput {

    pub source_db_instance_identifier: Option<String>,
    pub target_db_instance_identifier: Option<String>,
    pub restore_time: Option<Instant>,
    pub use_latest_restorable_time: bool,
    pub db_instance_class: Option<String>,
    pub port: Option<i32>,
    pub availability_zone: Option<String>,
    pub db_subnet_group_name: Option<String>,
    pub multi_az: Option<bool>,
    pub publicly_accessible: Option<bool>,
    pub auto_minor_version_upgrade: Option<bool>,
    pub license_model: Option<String>,
    pub db_name: Option<String>,
    pub engine: Option<String>,
    pub iops: Option<i32>,
    pub option_group_name: Option<String>,
    pub copy_tags_to_snapshot: Option<bool>,
    pub tags: Option<Vec<Tag>>,
    pub storage_type: Option<String>,
    pub tde_credential_arn: Option<String>,
    pub tde_credential_password: Option<String>,
    pub vpc_security_group_ids: Option<Vec<String>>,
    pub domain: Option<String>,
    pub domain_iam_role_name: Option<String>,
    pub enable_iam_database_authentication: Option<bool>,
    pub enable_cloudwatch_logs_exports: Option<Vec<String>>,
    pub processor_features: Option<Vec<ProcessorFeature>>,
    pub use_default_processor_features: Option<bool>,
    pub db_parameter_group_name: Option<String>,
    pub deletion_protection: Option<bool>,
    pub source_dbi_resource_id: Option<String>,
    pub max_allocated_storage: Option<i32>,
    pub source_db_instance_automated_backups_arn: Option<String>,
    pub enable_customer_owned_ip: Option<bool>,

}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

source_db_instance_identifier: Option<String>

The identifier of the source DB instance from which to restore.

Constraints:

Must match the identifier of an existing DB instance.

target_db_instance_identifier: Option<String>

The name of the new DB instance to be created.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

restore_time: Option<Instant>

The date and time to restore from.

Valid Values: Value must be a time in Universal Coordinated Time (UTC) format

Constraints:

Must be before the latest restorable time for the DB instance

Can't be specified if the UseLatestRestorableTime parameter is enabled

Example: 2009-09-07T23:45:00Z

use_latest_restorable_time: bool

A value that indicates whether the DB instance is restored from the latest backup time. By default, the DB instance isn't restored from the latest backup time.

Constraints: Can't be specified if the RestoreTime parameter is provided.

db_instance_class: Option<String>

The compute and memory capacity of the Amazon RDS DB instance, for example, db.m4.large. Not all DB instance classes are available in all Amazon Web Services Regions, or for all database engines. For the full list of DB instance classes, and availability for your engine, see DB Instance Class in the Amazon RDS User Guide.

Default: The same DBInstanceClass as the original DB instance.

port: Option<i32>

The port number on which the database accepts connections.

Constraints: Value must be 1150-65535

Default: The same port as the original DB instance.

availability_zone: Option<String>

The Availability Zone (AZ) where the DB instance will be created.

Default: A random, system-chosen Availability Zone.

Constraint: You can't specify the AvailabilityZone parameter if the DB instance is a Multi-AZ deployment.

Example: us-east-1a

db_subnet_group_name: Option<String>

The DB subnet group name to use for the new instance.

Constraints: If supplied, must match the name of an existing DBSubnetGroup.

Example: mySubnetgroup

multi_az: Option<bool>

A value that indicates whether the DB instance is a Multi-AZ deployment.

Constraint: You can't specify the AvailabilityZone parameter if the DB instance is a Multi-AZ deployment.

publicly_accessible: Option<bool>

A value that indicates whether the DB instance is publicly accessible.

When the DB instance is publicly accessible, its DNS endpoint resolves to the private IP address from within the DB instance's VPC, and to the public IP address from outside of the DB instance's VPC. Access to the DB instance is ultimately controlled by the security group it uses, and that public access is not permitted if the security group assigned to the DB instance doesn't permit it.

When the DB instance isn't publicly accessible, it is an internal DB instance with a DNS name that resolves to a private IP address.

For more information, see CreateDBInstance.

auto_minor_version_upgrade: Option<bool>

A value that indicates whether minor version upgrades are applied automatically to the DB instance during the maintenance window.

license_model: Option<String>

License model information for the restored DB instance.

Default: Same as source.

Valid values: license-included | bring-your-own-license | general-public-license

db_name: Option<String>

The database name for the restored DB instance.

This parameter isn't used for the MySQL or MariaDB engines.

engine: Option<String>

The database engine to use for the new instance.

Default: The same as source

Constraint: Must be compatible with the engine of the source

Valid Values:

mariadb

mysql

oracle-ee

oracle-ee-cdb

oracle-se2

oracle-se2-cdb

postgres

sqlserver-ee

sqlserver-se

sqlserver-ex

sqlserver-web

iops: Option<i32>

The amount of Provisioned IOPS (input/output operations per second) to be initially allocated for the DB instance.

Constraints: Must be an integer greater than 1000.

SQL Server

Setting the IOPS value for the SQL Server database engine isn't supported.

option_group_name: Option<String>

The name of the option group to be used for the restored DB instance.

Permanent options, such as the TDE option for Oracle Advanced Security TDE, can't be removed from an option group, and that option group can't be removed from a DB instance once it is associated with a DB instance

copy_tags_to_snapshot: Option<bool>

A value that indicates whether to copy all tags from the restored DB instance to snapshots of the DB instance. By default, tags are not copied.

tags: Option<Vec<Tag>>

A list of tags. For more information, see Tagging Amazon RDS Resources in the Amazon RDS User Guide.

storage_type: Option<String>

Specifies the storage type to be associated with the DB instance.

Valid values: standard | gp2 | io1

If you specify io1, you must also include a value for the Iops parameter.

Default: io1 if the Iops parameter is specified, otherwise gp2

tde_credential_arn: Option<String>

The ARN from the key store with which to associate the instance for TDE encryption.

tde_credential_password: Option<String>

The password for the given ARN from the key store in order to access the device.

vpc_security_group_ids: Option<Vec<String>>

A list of EC2 VPC security groups to associate with this DB instance.

Default: The default EC2 VPC security group for the DB subnet group's VPC.

domain: Option<String>

Specify the Active Directory directory ID to restore the DB instance in. The domain must be created prior to this operation. Currently, only MySQL, Microsoft SQL Server, Oracle, and PostgreSQL DB instances can be created in an Active Directory Domain.

For more information, see Kerberos Authentication in the Amazon RDS User Guide.

domain_iam_role_name: Option<String>

Specify the name of the IAM role to be used when making API calls to the Directory Service.

enable_iam_database_authentication: Option<bool>

A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

For more information about IAM database authentication, see IAM Database Authentication for MySQL and PostgreSQL in the Amazon RDS User Guide.

enable_cloudwatch_logs_exports: Option<Vec<String>>

The list of logs that the restored DB instance is to export to CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see Publishing Database Logs to Amazon CloudWatch Logs in the Amazon RDS User Guide.

processor_features: Option<Vec<ProcessorFeature>>

The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

use_default_processor_features: Option<bool>

A value that indicates whether the DB instance class of the DB instance uses its default processor features.

db_parameter_group_name: Option<String>

The name of the DB parameter group to associate with this DB instance.

If you do not specify a value for DBParameterGroupName, then the default DBParameterGroup for the specified DB engine is used.

Constraints:

If supplied, must match the name of an existing DBParameterGroup.

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

deletion_protection: Option<bool>

A value that indicates whether the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled. For more information, see Deleting a DB Instance.

source_dbi_resource_id: Option<String>

The resource ID of the source DB instance from which to restore.

max_allocated_storage: Option<i32>

The upper limit to which Amazon RDS can automatically scale the storage of the DB instance.

For more information about this setting, including limitations that apply to it, see Managing capacity automatically with Amazon RDS storage autoscaling in the Amazon RDS User Guide.

source_db_instance_automated_backups_arn: Option<String>

The Amazon Resource Name (ARN) of the replicated automated backups from which to restore, for example, arn:aws:rds:useast-1:123456789012:auto-backup:ab-L2IJCEXJP7XQ7HOJ4SIEXAMPLE.

enable_customer_owned_ip: Option<bool>

A value that indicates whether to enable a customer-owned IP address (CoIP) for an RDS on Outposts DB instance.

A CoIP provides local or external connectivity to resources in your Outpost subnets through your on-premises network. For some use cases, a CoIP can provide lower latency for connections to the DB instance from outside of its virtual private cloud (VPC) on your local network.

For more information about RDS on Outposts, see Working with Amazon RDS on Amazon Web Services Outposts in the Amazon RDS User Guide.

For more information about CoIPs, see Customer-owned IP addresses in the Amazon Web Services Outposts User Guide.

Implementations

impl RestoreDbInstanceToPointInTimeInput

pub fn make_operation(
&self,
_config: &Config
) -> Result<Operation<RestoreDBInstanceToPointInTime, AwsErrorRetryPolicy>, BuildError>

Consumes the builder and constructs an Operation<RestoreDBInstanceToPointInTime>

pub fn builder() -> Builder

Creates a new builder-style object to manufacture RestoreDbInstanceToPointInTimeInput