Module aws_sdk_rds::model

Modules

account_quota

See AccountQuota

availability_zone

See AvailabilityZone

available_processor_feature

See AvailableProcessorFeature

certificate

See Certificate

character_set

See CharacterSet

cloudwatch_logs_export_configuration

See CloudwatchLogsExportConfiguration

cluster_pending_modified_values

See ClusterPendingModifiedValues

connection_pool_configuration

See ConnectionPoolConfiguration

connection_pool_configuration_info

See ConnectionPoolConfigurationInfo

custom_availability_zone

See CustomAvailabilityZone

db_cluster

See DbCluster

db_cluster_backtrack

See DbClusterBacktrack

db_cluster_endpoint

See DbClusterEndpoint

db_cluster_member

See DbClusterMember

db_cluster_option_group_status

See DbClusterOptionGroupStatus

db_cluster_parameter_group

See DbClusterParameterGroup

db_cluster_role

See DbClusterRole

db_cluster_snapshot

See DbClusterSnapshot

db_cluster_snapshot_attribute

See DbClusterSnapshotAttribute

db_cluster_snapshot_attributes_result

See DbClusterSnapshotAttributesResult

db_engine_version

See DbEngineVersion

db_instance

See DbInstance

db_instance_automated_backup

See DbInstanceAutomatedBackup

db_instance_automated_backups_replication

See DbInstanceAutomatedBackupsReplication

db_instance_role

See DbInstanceRole

db_instance_status_info

See DbInstanceStatusInfo

db_parameter_group

See DbParameterGroup

db_parameter_group_status

See DbParameterGroupStatus

db_proxy

See DbProxy

db_proxy_endpoint

See DbProxyEndpoint

db_proxy_target

See DbProxyTarget

db_proxy_target_group

See DbProxyTargetGroup

db_security_group

See DbSecurityGroup

db_security_group_membership

See DbSecurityGroupMembership

db_snapshot

See DbSnapshot

db_snapshot_attribute

See DbSnapshotAttribute

db_snapshot_attributes_result

See DbSnapshotAttributesResult

db_subnet_group

See DbSubnetGroup

describe_db_log_files_details

See DescribeDbLogFilesDetails

domain_membership

See DomainMembership

double_range

See DoubleRange

ec2_security_group

See Ec2SecurityGroup

endpoint

See Endpoint

engine_defaults

See EngineDefaults

event

See Event

event_categories_map

See EventCategoriesMap

event_subscription

See EventSubscription

export_task

See ExportTask

failover_state

See FailoverState

filter

See Filter

global_cluster

See GlobalCluster

global_cluster_member

See GlobalClusterMember

installation_media

See InstallationMedia

installation_media_failure_cause

See InstallationMediaFailureCause

ip_range

See IpRange

minimum_engine_version_per_allowed_value

See MinimumEngineVersionPerAllowedValue

option

See Option

option_configuration

See OptionConfiguration

option_group

See OptionGroup

option_group_membership

See OptionGroupMembership

option_group_option

See OptionGroupOption

option_group_option_setting

See OptionGroupOptionSetting

option_setting

See OptionSetting

option_version

See OptionVersion

orderable_db_instance_option

See OrderableDbInstanceOption

outpost

See Outpost

parameter

See Parameter

pending_cloudwatch_logs_exports

See PendingCloudwatchLogsExports

pending_maintenance_action

See PendingMaintenanceAction

pending_modified_values

See PendingModifiedValues

processor_feature

See ProcessorFeature

range

See Range

recurring_charge

See RecurringCharge

reserved_db_instance

See ReservedDbInstance

reserved_db_instances_offering

See ReservedDbInstancesOffering

resource_pending_maintenance_actions

See ResourcePendingMaintenanceActions

restore_window

See RestoreWindow

scaling_configuration

See ScalingConfiguration

scaling_configuration_info

See ScalingConfigurationInfo

source_region

See SourceRegion

subnet

See Subnet

tag

See Tag

target_health

See TargetHealth

timezone

See Timezone

upgrade_target

See UpgradeTarget

user_auth_config

See UserAuthConfig

user_auth_config_info

See UserAuthConfigInfo

valid_db_instance_modifications_message

See ValidDbInstanceModificationsMessage

valid_storage_options

See ValidStorageOptions

vpc_security_group_membership

See VpcSecurityGroupMembership

vpn_details

See VpnDetails

Structs

AccountQuota

Describes a quota for an Amazon Web Services account.

AvailabilityZone

Contains Availability Zone information.

AvailableProcessorFeature

Contains the available processor feature information for the DB instance class of a DB instance.

Certificate

A CA certificate for an Amazon Web Services account.

CharacterSet

CloudwatchLogsExportConfiguration

The configuration setting for the log types to be enabled for export to CloudWatch

ClusterPendingModifiedValues

This data type is used as a response element in the ModifyDBCluster operation and

ConnectionPoolConfiguration

Specifies the settings that control the size and behavior of the connection pool associated with a DBProxyTargetGroup.

ConnectionPoolConfigurationInfo

Displays the settings that control the size and behavior of the connection pool associated with a DBProxyTarget.

CustomAvailabilityZone

A custom Availability Zone (AZ) is an on-premises AZ that is integrated with a VMware vSphere cluster.

DbCluster

Contains the details of an Amazon Aurora DB cluster.

DbClusterBacktrack

This data type is used as a response element in the DescribeDBClusterBacktracks action.

DbClusterEndpoint

This data type represents the information you need to connect to an Amazon Aurora DB cluster.

DbClusterMember

Contains information about an instance that is part of a DB cluster.

DbClusterOptionGroupStatus

Contains status information for a DB cluster option group.

DbClusterParameterGroup

Contains the details of an Amazon RDS DB cluster parameter group.

DbClusterRole

Describes an Amazon Web Services Identity and Access Management (IAM) role that is associated with a DB cluster.

DbClusterSnapshot

Contains the details for an Amazon RDS DB cluster snapshot

DbClusterSnapshotAttribute

Contains the name and values of a manual DB cluster snapshot attribute.

DbClusterSnapshotAttributesResult

Contains the results of a successful call to the DescribeDBClusterSnapshotAttributes

DbEngineVersion

DbInstance

Contains the details of an Amazon RDS DB instance.

DbInstanceAutomatedBackup

An automated backup of a DB instance. It consists of system backups, transaction logs, and the database instance properties that

DbInstanceAutomatedBackupsReplication

Automated backups of a DB instance replicated to another Amazon Web Services Region. They consist of system backups, transaction logs, and database instance properties.

DbInstanceRole

Describes an Amazon Web Services Identity and Access Management (IAM) role that is associated with a DB instance.

DbInstanceStatusInfo

Provides a list of status information for a DB instance.

DbParameterGroup

Contains the details of an Amazon RDS DB parameter group.

DbParameterGroupStatus

The status of the DB parameter group.

DbProxy

The data structure representing a proxy managed by the RDS Proxy.

DbProxyEndpoint

The data structure representing an endpoint associated with a DB proxy. RDS automatically creates one

DbProxyTarget

Contains the details for an RDS Proxy target. It represents an RDS DB instance or Aurora DB cluster

DbProxyTargetGroup

Represents a set of RDS DB instances, Aurora DB clusters, or both that a proxy can connect to. Currently, each target group

DbSecurityGroup

Contains the details for an Amazon RDS DB security group.

DbSecurityGroupMembership

This data type is used as a response element in the following actions:

DbSnapshot

Contains the details of an Amazon RDS DB snapshot.

DbSnapshotAttribute

Contains the name and values of a manual DB snapshot attribute

DbSnapshotAttributesResult

Contains the results of a successful call to the DescribeDBSnapshotAttributes

DbSubnetGroup

Contains the details of an Amazon RDS DB subnet group.

DescribeDbLogFilesDetails

This data type is used as a response element to DescribeDBLogFiles.

DomainMembership

An Active Directory Domain membership record associated with the DB instance or cluster.

DoubleRange

A range of double values.

Ec2SecurityGroup

This data type is used as a response element in the following actions:

Endpoint

This data type represents the information you need to connect to an Amazon RDS DB instance.

EngineDefaults

Event

EventCategoriesMap

Contains the results of a successful invocation of the DescribeEventCategories operation.

EventSubscription

Contains the results of a successful invocation of the DescribeEventSubscriptions action.

ExportTask

Contains the details of a snapshot export to Amazon S3.

FailoverState

Contains the state of scheduled or in-process failover operations on an

Filter

A filter name and value pair that is used to return a more specific list of results

GlobalCluster

A data type representing an Aurora global database.

GlobalClusterMember

InstallationMedia

Contains the installation media for a DB engine that requires an on-premises

InstallationMediaFailureCause

Contains the cause of an installation media failure. Installation media is used

IpRange

MinimumEngineVersionPerAllowedValue

The minimum DB engine version required for each corresponding allowed value for an option setting.

Option

Option details.

OptionConfiguration

A list of all available options

OptionGroup

OptionGroupMembership

Provides information on the option groups the DB instance is a member of.

OptionGroupOption

Available option.

OptionGroupOptionSetting

Option group option settings are used to display settings available for each option with their default values and other information. These values are used with the DescribeOptionGroupOptions action.

OptionSetting

Option settings are the actual settings being applied or configured for that option. It is used when you modify an option group or describe option groups. For example, the NATIVE_NETWORK_ENCRYPTION option has a setting called SQLNET.ENCRYPTION_SERVER that can have several different values.

OptionVersion

The version for an option. Option group option versions are returned by

OrderableDbInstanceOption

Contains a list of available options for a DB instance.

Outpost

A data type that represents an Outpost.

Parameter

PendingCloudwatchLogsExports

A list of the log types whose configuration is still pending. In other words, these log types are in the process of being activated or deactivated.

PendingMaintenanceAction

Provides information about a pending maintenance action for a resource.

PendingModifiedValues

ProcessorFeature

Contains the processor features of a DB instance class.

Range

A range of integer values.

RecurringCharge

ReservedDbInstance

ReservedDbInstancesOffering

ResourcePendingMaintenanceActions

Describes the pending maintenance actions for a resource.

RestoreWindow

Earliest and latest time an instance can be restored to:

ScalingConfiguration

Contains the scaling configuration of an Aurora Serverless DB cluster.

ScalingConfigurationInfo

Shows the scaling configuration for an Aurora DB cluster in serverless DB engine mode.

SourceRegion

Contains an Amazon Web Services Region name as the result of a successful call to the DescribeSourceRegions action.

Subnet

Tag

Metadata assigned to an Amazon RDS resource consisting of a key-value pair.

TargetHealth

Information about the connection health of an RDS Proxy target.

Timezone

A time zone associated with a

UpgradeTarget

The version of the database engine that a DB instance can be upgraded to.

UserAuthConfig

Specifies the details of authentication used by a proxy to log in as a specific database user.

UserAuthConfigInfo

Returns the details of authentication used by a proxy to log in as a specific database user.

ValidDbInstanceModificationsMessage

Information about valid modifications that you can make to your DB instance.

ValidStorageOptions

Information about valid modifications that you can make to your DB instance.

VpcSecurityGroupMembership

This data type is used as a response element for queries on VPC security group membership.

VpnDetails

Information about the virtual private network (VPN) between the VMware vSphere cluster and the Amazon Web Services website.

Enums

ActivityStreamMode

ActivityStreamStatus

ApplyMethod

AuthScheme

DbProxyEndpointStatus

DbProxyEndpointTargetRole

DbProxyStatus

EngineFamily

FailoverStatus

IamAuthMode

ReplicaMode

SourceType

TargetHealthReason

TargetRole

NOTE: TargetRole::Unknown has been renamed to ::UnknownValue.

TargetState

TargetType

WriteForwardingStatus

NOTE: WriteForwardingStatus::Unknown has been renamed to ::UnknownValue.