Struct aws_sdk_rds::model::db_snapshot::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for DbSnapshot

Implementations

impl Builder

pub fn db_snapshot_identifier(self, input: impl Into<String>) -> Self

Specifies the identifier for the DB snapshot.

pub fn set_db_snapshot_identifier(self, input: Option<String>) -> Self

pub fn db_instance_identifier(self, input: impl Into<String>) -> Self

Specifies the DB instance identifier of the DB instance this DB snapshot was created from.

pub fn set_db_instance_identifier(self, input: Option<String>) -> Self

pub fn snapshot_create_time(self, input: Instant) -> Self

Specifies when the snapshot was taken in Coordinated Universal Time (UTC).

pub fn set_snapshot_create_time(self, input: Option<Instant>) -> Self

pub fn engine(self, input: impl Into<String>) -> Self

Specifies the name of the database engine.

pub fn set_engine(self, input: Option<String>) -> Self

pub fn allocated_storage(self, input: i32) -> Self

Specifies the allocated storage size in gibibytes (GiB).

pub fn set_allocated_storage(self, input: Option<i32>) -> Self

pub fn status(self, input: impl Into<String>) -> Self

Specifies the status of this DB snapshot.

pub fn set_status(self, input: Option<String>) -> Self

pub fn port(self, input: i32) -> Self

Specifies the port that the database engine was listening on at the time of the snapshot.

pub fn set_port(self, input: Option<i32>) -> Self

pub fn availability_zone(self, input: impl Into<String>) -> Self

Specifies the name of the Availability Zone the DB instance was located in at the time of the DB snapshot.

pub fn set_availability_zone(self, input: Option<String>) -> Self

pub fn vpc_id(self, input: impl Into<String>) -> Self

Provides the VPC ID associated with the DB snapshot.

pub fn set_vpc_id(self, input: Option<String>) -> Self

pub fn instance_create_time(self, input: Instant) -> Self

Specifies the time in Coordinated Universal Time (UTC) when the DB instance, from which the snapshot was taken, was created.

pub fn set_instance_create_time(self, input: Option<Instant>) -> Self

pub fn master_username(self, input: impl Into<String>) -> Self

Provides the master username for the DB snapshot.

pub fn set_master_username(self, input: Option<String>) -> Self

pub fn engine_version(self, input: impl Into<String>) -> Self

Specifies the version of the database engine.

pub fn set_engine_version(self, input: Option<String>) -> Self

pub fn license_model(self, input: impl Into<String>) -> Self

License model information for the restored DB instance.

pub fn set_license_model(self, input: Option<String>) -> Self

pub fn snapshot_type(self, input: impl Into<String>) -> Self

Provides the type of the DB snapshot.

pub fn set_snapshot_type(self, input: Option<String>) -> Self

pub fn iops(self, input: i32) -> Self

Specifies the Provisioned IOPS (I/O operations per second) value of the DB instance at the time of the snapshot.

pub fn set_iops(self, input: Option<i32>) -> Self

pub fn option_group_name(self, input: impl Into<String>) -> Self

Provides the option group name for the DB snapshot.

pub fn set_option_group_name(self, input: Option<String>) -> Self

pub fn percent_progress(self, input: i32) -> Self

The percentage of the estimated data that has been transferred.

pub fn set_percent_progress(self, input: Option<i32>) -> Self

pub fn source_region(self, input: impl Into<String>) -> Self

The Amazon Web Services Region that the DB snapshot was created in or copied from.

pub fn set_source_region(self, input: Option<String>) -> Self

pub fn source_db_snapshot_identifier(self, input: impl Into<String>) -> Self

The DB snapshot Amazon Resource Name (ARN) that the DB snapshot was copied from. It only has value in case of cross-customer or cross-region copy.

pub fn set_source_db_snapshot_identifier(self, input: Option<String>) -> Self

pub fn storage_type(self, input: impl Into<String>) -> Self

Specifies the storage type associated with DB snapshot.

pub fn set_storage_type(self, input: Option<String>) -> Self

pub fn tde_credential_arn(self, input: impl Into<String>) -> Self

The ARN from the key store with which to associate the instance for TDE encryption.

pub fn set_tde_credential_arn(self, input: Option<String>) -> Self

pub fn encrypted(self, input: bool) -> Self

Specifies whether the DB snapshot is encrypted.

pub fn set_encrypted(self, input: Option<bool>) -> Self

pub fn kms_key_id(self, input: impl Into<String>) -> Self

If Encrypted is true, the Amazon Web Services KMS key identifier for the encrypted DB snapshot.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK).

pub fn set_kms_key_id(self, input: Option<String>) -> Self

pub fn db_snapshot_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) for the DB snapshot.

pub fn set_db_snapshot_arn(self, input: Option<String>) -> Self

pub fn timezone(self, input: impl Into<String>) -> Self

The time zone of the DB snapshot. In most cases, the Timezone element is empty. Timezone content appears only for snapshots taken from Microsoft SQL Server DB instances that were created with a time zone specified.

pub fn set_timezone(self, input: Option<String>) -> Self

pub fn iam_database_authentication_enabled(self, input: bool) -> Self

True if mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts is enabled, and otherwise false.

pub fn set_iam_database_authentication_enabled(
self,
input: Option<bool>
) -> Self

pub fn processor_features(self, input: impl Into<ProcessorFeature>) -> Self

pub fn set_processor_features(
self,
input: Option<Vec<ProcessorFeature>>
) -> Self

pub fn dbi_resource_id(self, input: impl Into<String>) -> Self

The identifier for the source DB instance, which can't be changed and which is unique to an Amazon Web Services Region.

pub fn set_dbi_resource_id(self, input: Option<String>) -> Self

pub fn tag_list(self, input: impl Into<Tag>) -> Self

pub fn set_tag_list(self, input: Option<Vec<Tag>>) -> Self

pub fn build(self) -> DbSnapshot

Consumes the builder and constructs a DbSnapshot