Struct aws_sdk_rds::model::orderable_db_instance_option::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for OrderableDbInstanceOption

Implementations

impl Builder

pub fn engine(self, input: impl Into<String>) -> Self

The engine type of a DB instance.

pub fn set_engine(self, input: Option<String>) -> Self

pub fn engine_version(self, input: impl Into<String>) -> Self

The engine version of a DB instance.

pub fn set_engine_version(self, input: Option<String>) -> Self

pub fn db_instance_class(self, input: impl Into<String>) -> Self

The DB instance class for a DB instance.

pub fn set_db_instance_class(self, input: Option<String>) -> Self

pub fn license_model(self, input: impl Into<String>) -> Self

The license model for a DB instance.

pub fn set_license_model(self, input: Option<String>) -> Self

pub fn availability_zone_group(self, input: impl Into<String>) -> Self

The Availability Zone group for a DB instance.

pub fn set_availability_zone_group(self, input: Option<String>) -> Self

pub fn availability_zones(self, input: impl Into<AvailabilityZone>) -> Self

pub fn set_availability_zones(
self,
input: Option<Vec<AvailabilityZone>>
) -> Self

pub fn multi_az_capable(self, input: bool) -> Self

Indicates whether a DB instance is Multi-AZ capable.

pub fn set_multi_az_capable(self, input: Option<bool>) -> Self

pub fn read_replica_capable(self, input: bool) -> Self

Indicates whether a DB instance can have a read replica.

pub fn set_read_replica_capable(self, input: Option<bool>) -> Self

pub fn vpc(self, input: bool) -> Self

Indicates whether a DB instance is in a VPC.

pub fn set_vpc(self, input: Option<bool>) -> Self

pub fn supports_storage_encryption(self, input: bool) -> Self

Indicates whether a DB instance supports encrypted storage.

pub fn set_supports_storage_encryption(self, input: Option<bool>) -> Self

pub fn storage_type(self, input: impl Into<String>) -> Self

Indicates the storage type for a DB instance.

pub fn set_storage_type(self, input: Option<String>) -> Self

pub fn supports_iops(self, input: bool) -> Self

Indicates whether a DB instance supports provisioned IOPS.

pub fn set_supports_iops(self, input: Option<bool>) -> Self

pub fn supports_enhanced_monitoring(self, input: bool) -> Self

Indicates whether a DB instance supports Enhanced Monitoring at intervals from 1 to 60 seconds.

pub fn set_supports_enhanced_monitoring(self, input: Option<bool>) -> Self

pub fn supports_iam_database_authentication(self, input: bool) -> Self

Indicates whether a DB instance supports IAM database authentication.

pub fn set_supports_iam_database_authentication(
self,
input: Option<bool>
) -> Self

pub fn supports_performance_insights(self, input: bool) -> Self

True if a DB instance supports Performance Insights, otherwise false.

pub fn set_supports_performance_insights(self, input: Option<bool>) -> Self

pub fn min_storage_size(self, input: i32) -> Self

Minimum storage size for a DB instance.

pub fn set_min_storage_size(self, input: Option<i32>) -> Self

pub fn max_storage_size(self, input: i32) -> Self

Maximum storage size for a DB instance.

pub fn set_max_storage_size(self, input: Option<i32>) -> Self

pub fn min_iops_per_db_instance(self, input: i32) -> Self

Minimum total provisioned IOPS for a DB instance.

pub fn set_min_iops_per_db_instance(self, input: Option<i32>) -> Self

pub fn max_iops_per_db_instance(self, input: i32) -> Self

Maximum total provisioned IOPS for a DB instance.

pub fn set_max_iops_per_db_instance(self, input: Option<i32>) -> Self

pub fn min_iops_per_gib(self, input: f64) -> Self

Minimum provisioned IOPS per GiB for a DB instance.

pub fn set_min_iops_per_gib(self, input: Option<f64>) -> Self

pub fn max_iops_per_gib(self, input: f64) -> Self

Maximum provisioned IOPS per GiB for a DB instance.

pub fn set_max_iops_per_gib(self, input: Option<f64>) -> Self

pub fn available_processor_features(
self,
input: impl Into<AvailableProcessorFeature>
) -> Self

pub fn set_available_processor_features(
self,
input: Option<Vec<AvailableProcessorFeature>>
) -> Self

pub fn supported_engine_modes(self, input: impl Into<String>) -> Self

pub fn set_supported_engine_modes(self, input: Option<Vec<String>>) -> Self

pub fn supports_storage_autoscaling(self, input: bool) -> Self

Whether Amazon RDS can automatically scale storage for DB instances that use the specified DB instance class.

pub fn set_supports_storage_autoscaling(self, input: Option<bool>) -> Self

pub fn supports_kerberos_authentication(self, input: bool) -> Self

Whether a DB instance supports Kerberos Authentication.

pub fn set_supports_kerberos_authentication(self, input: Option<bool>) -> Self

pub fn outpost_capable(self, input: bool) -> Self

Whether a DB instance supports RDS on Outposts.

For more information about RDS on Outposts, see Amazon RDS on Amazon Web Services Outposts in the Amazon RDS User Guide.

pub fn set_outpost_capable(self, input: Option<bool>) -> Self

pub fn supported_activity_stream_modes(self, input: impl Into<String>) -> Self

pub fn set_supported_activity_stream_modes(
self,
input: Option<Vec<String>>
) -> Self

pub fn supports_global_databases(self, input: bool) -> Self

A value that indicates whether you can use Aurora global databases with a specific combination of other DB engine attributes.

pub fn set_supports_global_databases(self, input: Option<bool>) -> Self

pub fn build(self) -> OrderableDbInstanceOption

Consumes the builder and constructs a OrderableDbInstanceOption