Struct aws_sdk_rds::model::DbInstance

#[non_exhaustive]
pub struct DbInstance {

    pub db_instance_identifier: Option<String>,
    pub db_instance_class: Option<String>,
    pub engine: Option<String>,
    pub db_instance_status: Option<String>,
    pub master_username: Option<String>,
    pub db_name: Option<String>,
    pub endpoint: Option<Endpoint>,
    pub allocated_storage: i32,
    pub instance_create_time: Option<Instant>,
    pub preferred_backup_window: Option<String>,
    pub backup_retention_period: i32,
    pub db_security_groups: Option<Vec<DbSecurityGroupMembership>>,
    pub vpc_security_groups: Option<Vec<VpcSecurityGroupMembership>>,
    pub db_parameter_groups: Option<Vec<DbParameterGroupStatus>>,
    pub availability_zone: Option<String>,
    pub db_subnet_group: Option<DbSubnetGroup>,
    pub preferred_maintenance_window: Option<String>,
    pub pending_modified_values: Option<PendingModifiedValues>,
    pub latest_restorable_time: Option<Instant>,
    pub multi_az: bool,
    pub engine_version: Option<String>,
    pub auto_minor_version_upgrade: bool,
    pub read_replica_source_db_instance_identifier: Option<String>,
    pub read_replica_db_instance_identifiers: Option<Vec<String>>,
    pub read_replica_db_cluster_identifiers: Option<Vec<String>>,
    pub replica_mode: Option<ReplicaMode>,
    pub license_model: Option<String>,
    pub iops: Option<i32>,
    pub option_group_memberships: Option<Vec<OptionGroupMembership>>,
    pub character_set_name: Option<String>,
    pub nchar_character_set_name: Option<String>,
    pub secondary_availability_zone: Option<String>,
    pub publicly_accessible: bool,
    pub status_infos: Option<Vec<DbInstanceStatusInfo>>,
    pub storage_type: Option<String>,
    pub tde_credential_arn: Option<String>,
    pub db_instance_port: i32,
    pub db_cluster_identifier: Option<String>,
    pub storage_encrypted: bool,
    pub kms_key_id: Option<String>,
    pub dbi_resource_id: Option<String>,
    pub ca_certificate_identifier: Option<String>,
    pub domain_memberships: Option<Vec<DomainMembership>>,
    pub copy_tags_to_snapshot: bool,
    pub monitoring_interval: Option<i32>,
    pub enhanced_monitoring_resource_arn: Option<String>,
    pub monitoring_role_arn: Option<String>,
    pub promotion_tier: Option<i32>,
    pub db_instance_arn: Option<String>,
    pub timezone: Option<String>,
    pub iam_database_authentication_enabled: bool,
    pub performance_insights_enabled: Option<bool>,
    pub performance_insights_kms_key_id: Option<String>,
    pub performance_insights_retention_period: Option<i32>,
    pub enabled_cloudwatch_logs_exports: Option<Vec<String>>,
    pub processor_features: Option<Vec<ProcessorFeature>>,
    pub deletion_protection: bool,
    pub associated_roles: Option<Vec<DbInstanceRole>>,
    pub listener_endpoint: Option<Endpoint>,
    pub max_allocated_storage: Option<i32>,
    pub tag_list: Option<Vec<Tag>>,
    pub db_instance_automated_backups_replications: Option<Vec<DbInstanceAutomatedBackupsReplication>>,
    pub customer_owned_ip_enabled: Option<bool>,
    pub aws_backup_recovery_point_arn: Option<String>,
    pub activity_stream_status: Option<ActivityStreamStatus>,
    pub activity_stream_kms_key_id: Option<String>,
    pub activity_stream_kinesis_stream_name: Option<String>,
    pub activity_stream_mode: Option<ActivityStreamMode>,
    pub activity_stream_engine_native_audit_fields_included: Option<bool>,

}

Contains the details of an Amazon RDS DB instance.

This data type is used as a response element in the DescribeDBInstances action.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

db_instance_identifier: Option<String>

Contains a user-supplied database identifier. This identifier is the unique key that identifies a DB instance.

db_instance_class: Option<String>

Contains the name of the compute and memory capacity class of the DB instance.

engine: Option<String>

The name of the database engine to be used for this DB instance.

db_instance_status: Option<String>

Specifies the current state of this database.

For information about DB instance statuses, see Viewing DB instance status in the Amazon RDS User Guide.

master_username: Option<String>

Contains the master username for the DB instance.

db_name: Option<String>

The meaning of this parameter differs according to the database engine you use.

MySQL, MariaDB, SQL Server, PostgreSQL

Contains the name of the initial database of this instance that was provided at create time, if one was specified when the DB instance was created. This same name is returned for the life of the DB instance.

Type: String

Oracle

Contains the Oracle System ID (SID) of the created DB instance. Not shown when the returned parameters do not apply to an Oracle DB instance.

endpoint: Option<Endpoint>

Specifies the connection endpoint.

allocated_storage: i32

Specifies the allocated storage size specified in gibibytes.

instance_create_time: Option<Instant>

Provides the date and time the DB instance was created.

preferred_backup_window: Option<String>

Specifies the daily time range during which automated backups are created if automated backups are enabled, as determined by the BackupRetentionPeriod.

backup_retention_period: i32

Specifies the number of days for which automatic DB snapshots are retained.

db_security_groups: Option<Vec<DbSecurityGroupMembership>>

A list of DB security group elements containing DBSecurityGroup.Name and DBSecurityGroup.Status subelements.

vpc_security_groups: Option<Vec<VpcSecurityGroupMembership>>

Provides a list of VPC security group elements that the DB instance belongs to.

db_parameter_groups: Option<Vec<DbParameterGroupStatus>>

Provides the list of DB parameter groups applied to this DB instance.

availability_zone: Option<String>

Specifies the name of the Availability Zone the DB instance is located in.

db_subnet_group: Option<DbSubnetGroup>

Specifies information on the subnet group associated with the DB instance, including the name, description, and subnets in the subnet group.

preferred_maintenance_window: Option<String>

Specifies the weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

pending_modified_values: Option<PendingModifiedValues>

A value that specifies that changes to the DB instance are pending. This element is only included when changes are pending. Specific changes are identified by subelements.

latest_restorable_time: Option<Instant>

Specifies the latest time to which a database can be restored with point-in-time restore.

multi_az: bool

Specifies if the DB instance is a Multi-AZ deployment.

engine_version: Option<String>

Indicates the database engine version.

auto_minor_version_upgrade: bool

A value that indicates that minor version patches are applied automatically.

read_replica_source_db_instance_identifier: Option<String>

Contains the identifier of the source DB instance if this DB instance is a read replica.

read_replica_db_instance_identifiers: Option<Vec<String>>

Contains one or more identifiers of the read replicas associated with this DB instance.

read_replica_db_cluster_identifiers: Option<Vec<String>>

Contains one or more identifiers of Aurora DB clusters to which the RDS DB instance is replicated as a read replica. For example, when you create an Aurora read replica of an RDS MySQL DB instance, the Aurora MySQL DB cluster for the Aurora read replica is shown. This output does not contain information about cross region Aurora read replicas.

Currently, each RDS DB instance can have only one Aurora read replica.

replica_mode: Option<ReplicaMode>

The open mode of an Oracle read replica. The default is open-read-only. For more information, see Working with Oracle Read Replicas for Amazon RDS in the Amazon RDS User Guide.

This attribute is only supported in RDS for Oracle.

license_model: Option<String>

License model information for this DB instance.

iops: Option<i32>

Specifies the Provisioned IOPS (I/O operations per second) value.

option_group_memberships: Option<Vec<OptionGroupMembership>>

Provides the list of option group memberships for this DB instance.

character_set_name: Option<String>

If present, specifies the name of the character set that this instance is associated with.

nchar_character_set_name: Option<String>

The name of the NCHAR character set for the Oracle DB instance. This character set specifies the Unicode encoding for data stored in table columns of type NCHAR, NCLOB, or NVARCHAR2.

secondary_availability_zone: Option<String>

If present, specifies the name of the secondary Availability Zone for a DB instance with multi-AZ support.

publicly_accessible: bool

Specifies the accessibility options for the DB instance.

When the DB instance is publicly accessible, its DNS endpoint resolves to the private IP address from within the DB instance's VPC, and to the public IP address from outside of the DB instance's VPC. Access to the DB instance is ultimately controlled by the security group it uses, and that public access is not permitted if the security group assigned to the DB instance doesn't permit it.

When the DB instance isn't publicly accessible, it is an internal DB instance with a DNS name that resolves to a private IP address.

For more information, see CreateDBInstance.

status_infos: Option<Vec<DbInstanceStatusInfo>>

The status of a read replica. If the instance isn't a read replica, this is blank.

storage_type: Option<String>

Specifies the storage type associated with DB instance.

tde_credential_arn: Option<String>

The ARN from the key store with which the instance is associated for TDE encryption.

db_instance_port: i32

Specifies the port that the DB instance listens on. If the DB instance is part of a DB cluster, this can be a different port than the DB cluster port.

db_cluster_identifier: Option<String>

If the DB instance is a member of a DB cluster, contains the name of the DB cluster that the DB instance is a member of.

storage_encrypted: bool

Specifies whether the DB instance is encrypted.

kms_key_id: Option<String>

If StorageEncrypted is true, the Amazon Web Services KMS key identifier for the encrypted DB instance.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK).

dbi_resource_id: Option<String>

The Amazon Web Services Region-unique, immutable identifier for the DB instance. This identifier is found in Amazon Web Services CloudTrail log entries whenever the Amazon Web Services KMS customer master key (CMK) for the DB instance is accessed.

ca_certificate_identifier: Option<String>

The identifier of the CA certificate for this DB instance.

domain_memberships: Option<Vec<DomainMembership>>

The Active Directory Domain membership records associated with the DB instance.

copy_tags_to_snapshot: bool

Specifies whether tags are copied from the DB instance to snapshots of the DB instance.

Amazon Aurora

Not applicable. Copying tags to snapshots is managed by the DB cluster. Setting this value for an Aurora DB instance has no effect on the DB cluster setting. For more information, see DBCluster.

monitoring_interval: Option<i32>

The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the DB instance.

enhanced_monitoring_resource_arn: Option<String>

The Amazon Resource Name (ARN) of the Amazon CloudWatch Logs log stream that receives the Enhanced Monitoring metrics data for the DB instance.

monitoring_role_arn: Option<String>

The ARN for the IAM role that permits RDS to send Enhanced Monitoring metrics to Amazon CloudWatch Logs.

promotion_tier: Option<i32>

A value that specifies the order in which an Aurora Replica is promoted to the primary instance after a failure of the existing primary instance. For more information, see Fault Tolerance for an Aurora DB Cluster in the Amazon Aurora User Guide.

db_instance_arn: Option<String>

The Amazon Resource Name (ARN) for the DB instance.

timezone: Option<String>

The time zone of the DB instance. In most cases, the Timezone element is empty. Timezone content appears only for Microsoft SQL Server DB instances that were created with a time zone specified.

iam_database_authentication_enabled: bool

True if mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts is enabled, and otherwise false.

IAM database authentication can be enabled for the following database engines

For MySQL 5.6, minor version 5.6.34 or higher

For MySQL 5.7, minor version 5.7.16 or higher

Aurora 5.6 or higher. To enable IAM database authentication for Aurora, see DBCluster Type.

performance_insights_enabled: Option<bool>

True if Performance Insights is enabled for the DB instance, and otherwise false.

performance_insights_kms_key_id: Option<String>

The Amazon Web Services KMS key identifier for encryption of Performance Insights data.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK).

performance_insights_retention_period: Option<i32>

The amount of time, in days, to retain Performance Insights data. Valid values are 7 or 731 (2 years).

enabled_cloudwatch_logs_exports: Option<Vec<String>>

A list of log types that this DB instance is configured to export to CloudWatch Logs.

Log types vary by DB engine. For information about the log types for each DB engine, see Amazon RDS Database Log Files in the Amazon RDS User Guide.

processor_features: Option<Vec<ProcessorFeature>>

The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

deletion_protection: bool

Indicates if the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. For more information, see Deleting a DB Instance.

associated_roles: Option<Vec<DbInstanceRole>>

The Amazon Web Services Identity and Access Management (IAM) roles associated with the DB instance.

listener_endpoint: Option<Endpoint>

Specifies the listener connection endpoint for SQL Server Always On.

max_allocated_storage: Option<i32>

The upper limit to which Amazon RDS can automatically scale the storage of the DB instance.

tag_list: Option<Vec<Tag>>

A list of tags. For more information, see Tagging Amazon RDS Resources in the Amazon RDS User Guide.

db_instance_automated_backups_replications: Option<Vec<DbInstanceAutomatedBackupsReplication>>

The list of replicated automated backups associated with the DB instance.

customer_owned_ip_enabled: Option<bool>

Specifies whether a customer-owned IP address (CoIP) is enabled for an RDS on Outposts DB instance.

A CoIP provides local or external connectivity to resources in your Outpost subnets through your on-premises network. For some use cases, a CoIP can provide lower latency for connections to the DB instance from outside of its virtual private cloud (VPC) on your local network.

For more information about RDS on Outposts, see Working with Amazon RDS on Amazon Web Services Outposts in the Amazon RDS User Guide.

For more information about CoIPs, see Customer-owned IP addresses in the Amazon Web Services Outposts User Guide.

aws_backup_recovery_point_arn: Option<String>

The Amazon Resource Name (ARN) of the recovery point in Amazon Web Services Backup.

activity_stream_status: Option<ActivityStreamStatus>

The status of the database activity stream.

activity_stream_kms_key_id: Option<String>

The Amazon Web Services KMS key identifier used for encrypting messages in the database activity stream. The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK).

activity_stream_kinesis_stream_name: Option<String>

The name of the Amazon Kinesis data stream used for the database activity stream.

activity_stream_mode: Option<ActivityStreamMode>

The mode of the database activity stream. Database events such as a change or access generate an activity stream event. RDS for Oracle always handles these events asynchronously.

activity_stream_engine_native_audit_fields_included: Option<bool>

Indicates whether engine-native audit fields are included in the database activity stream.

Implementations

impl DbInstance

pub fn builder() -> Builder

Creates a new builder-style object to manufacture DbInstance