Struct aws_sdk_rds::model::OptionGroupOption

#[non_exhaustive]
pub struct OptionGroupOption {

    pub name: Option<String>,
    pub description: Option<String>,
    pub engine_name: Option<String>,
    pub major_engine_version: Option<String>,
    pub minimum_required_minor_engine_version: Option<String>,
    pub port_required: bool,
    pub default_port: Option<i32>,
    pub options_depended_on: Option<Vec<String>>,
    pub options_conflicts_with: Option<Vec<String>>,
    pub persistent: bool,
    pub permanent: bool,
    pub requires_auto_minor_engine_version_upgrade: bool,
    pub vpc_only: bool,
    pub supports_option_version_downgrade: Option<bool>,
    pub option_group_option_settings: Option<Vec<OptionGroupOptionSetting>>,
    pub option_group_option_versions: Option<Vec<OptionVersion>>,

}

Available option.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

name: Option<String>

The name of the option.

description: Option<String>

The description of the option.

engine_name: Option<String>

The name of the engine that this option can be applied to.

major_engine_version: Option<String>

Indicates the major engine version that the option is available for.

minimum_required_minor_engine_version: Option<String>

The minimum required engine version for the option to be applied.

port_required: bool

Specifies whether the option requires a port.

default_port: Option<i32>

If the option requires a port, specifies the default port for the option.

options_depended_on: Option<Vec<String>>

The options that are prerequisites for this option.

options_conflicts_with: Option<Vec<String>>

The options that conflict with this option.

persistent: bool

Persistent options can't be removed from an option group while DB instances are associated with the option group. If you disassociate all DB instances from the option group, your can remove the persistent option from the option group.

permanent: bool

Permanent options can never be removed from an option group. An option group containing a permanent option can't be removed from a DB instance.

requires_auto_minor_engine_version_upgrade: bool

If true, you must enable the Auto Minor Version Upgrade setting for your DB instance before you can use this option. You can enable Auto Minor Version Upgrade when you first create your DB instance, or by modifying your DB instance later.

vpc_only: bool

If true, you can only use this option with a DB instance that is in a VPC.

supports_option_version_downgrade: Option<bool>

If true, you can change the option to an earlier version of the option. This only applies to options that have different versions available.

option_group_option_settings: Option<Vec<OptionGroupOptionSetting>>

The option settings that are available (and the default value) for each option in an option group.

option_group_option_versions: Option<Vec<OptionVersion>>

The versions that are available for the option.

Implementations

impl OptionGroupOption

pub fn builder() -> Builder

Creates a new builder-style object to manufacture OptionGroupOption