Module aws_sdk_rds::operation

Structs

AddRoleToDBCluster

Associates an Identity and Access Management (IAM) role from an Amazon Aurora DB cluster.

AddRoleToDBInstance

Associates an Amazon Web Services Identity and Access Management (IAM) role with a DB instance.

AddSourceIdentifierToSubscription

Adds a source identifier to an existing RDS event notification subscription.

AddTagsToResource

Adds metadata tags to an Amazon RDS resource. These tags can also be used with cost allocation reporting to track cost associated with Amazon RDS resources, or used in a Condition statement in an IAM policy for Amazon RDS.

ApplyPendingMaintenanceAction

Applies a pending maintenance action to a resource (for example, to a DB instance).

AuthorizeDBSecurityGroupIngress

Enables ingress to a DBSecurityGroup using one of two forms of authorization. First, EC2 or VPC security

BacktrackDBCluster

Backtracks a DB cluster to a specific time, without creating a new DB cluster.

CancelExportTask

Cancels an export task in progress that is exporting a snapshot to Amazon S3.

CopyDBClusterParameterGroup

Copies the specified DB cluster parameter group.

CopyDBClusterSnapshot

Copies a snapshot of a DB cluster.

CopyDBParameterGroup

Copies the specified DB parameter group.

CopyDBSnapshot

Copies the specified DB snapshot. The source DB snapshot must be in the available state.

CopyOptionGroup

Copies the specified option group.

CreateCustomAvailabilityZone

Creates a custom Availability Zone (AZ).

CreateDBCluster

Creates a new Amazon Aurora DB cluster.

CreateDBClusterEndpoint

Creates a new custom endpoint and associates it with an Amazon Aurora DB cluster.

CreateDBClusterParameterGroup

Creates a new DB cluster parameter group.

CreateDBClusterSnapshot

Creates a snapshot of a DB cluster. For more information on Amazon Aurora, see

CreateDBInstance

Creates a new DB instance.

CreateDBInstanceReadReplica

Creates a new DB instance that acts as a read replica for an existing source DB

CreateDBParameterGroup

Creates a new DB parameter group.

CreateDBProxy

Creates a new DB proxy.

CreateDBProxyEndpoint

CreateDBSecurityGroup

Creates a new DB security group. DB security groups control access to a DB instance.

CreateDBSnapshot

Creates a snapshot of a DB instance. The source DB instance must be in the available or

CreateDBSubnetGroup

Creates a new DB subnet group. DB subnet groups must contain at least one subnet in at least two AZs in the Amazon Web Services Region.

CreateEventSubscription

Creates an RDS event notification subscription. This action requires a topic Amazon

CreateGlobalCluster

CreateOptionGroup

Creates a new option group. You can create up to 20 option groups.

DeleteCustomAvailabilityZone

Deletes a custom Availability Zone (AZ).

DeleteDBCluster

The DeleteDBCluster action deletes a previously provisioned DB cluster.

DeleteDBClusterEndpoint

Deletes a custom endpoint and removes it from an Amazon Aurora DB cluster.

DeleteDBClusterParameterGroup

Deletes a specified DB cluster parameter group. The DB cluster parameter group to be deleted can't be associated with any DB clusters.

DeleteDBClusterSnapshot

Deletes a DB cluster snapshot. If the snapshot is being copied, the copy operation is terminated.

DeleteDBInstance

The DeleteDBInstance action deletes a previously provisioned DB instance.

DeleteDBInstanceAutomatedBackup

Deletes automated backups using the DbiResourceId value of the source DB instance or the Amazon Resource Name (ARN) of the automated backups.

DeleteDBParameterGroup

Deletes a specified DB parameter group. The DB parameter group to be deleted can't be associated with any DB instances.

DeleteDBProxy

Deletes an existing DB proxy.

DeleteDBProxyEndpoint

Deletes a DBProxyEndpoint. Doing so removes the ability to access the DB proxy using the

DeleteDBSecurityGroup

Deletes a DB security group.

DeleteDBSnapshot

Deletes a DB snapshot. If the snapshot is being copied, the copy operation is

DeleteDBSubnetGroup

Deletes a DB subnet group.

DeleteEventSubscription

Deletes an RDS event notification subscription.

DeleteGlobalCluster

DeleteInstallationMedia

Deletes the installation medium for a DB engine that requires an on-premises customer provided license,

DeleteOptionGroup

Deletes an existing option group.

DeregisterDBProxyTargets

Remove the association between one or more DBProxyTarget data structures and a DBProxyTargetGroup.

DescribeAccountAttributes

Lists all of the attributes for a customer account. The attributes include Amazon RDS quotas for the account, such as the number of DB instances allowed. The description for a quota includes the quota name, current usage toward that quota, and the quota's maximum value.

DescribeCertificates

Lists the set of CA certificates provided by Amazon RDS for this Amazon Web Services account.

DescribeCustomAvailabilityZones

Returns information about custom Availability Zones (AZs).

DescribeDBClusterBacktracks

Returns information about backtracks for a DB cluster.

DescribeDBClusterEndpoints

Returns information about endpoints for an Amazon Aurora DB cluster.

DescribeDBClusterParameterGroups

DescribeDBClusterParameters

Returns the detailed parameter list for a particular DB cluster parameter group.

DescribeDBClusterSnapshotAttributes

Returns a list of DB cluster snapshot attribute names and values for a manual DB cluster snapshot.

DescribeDBClusterSnapshots

Returns information about DB cluster snapshots. This API action supports pagination.

DescribeDBClusters

Returns information about provisioned Aurora DB clusters. This API supports pagination.

DescribeDBEngineVersions

Returns a list of the available DB engines.

DescribeDBInstanceAutomatedBackups

Displays backups for both current and deleted

DescribeDBInstances

Returns information about provisioned RDS instances. This API supports pagination.

DescribeDBLogFiles

Returns a list of DB log files for the DB instance.

DescribeDBParameterGroups

DescribeDBParameters

Returns the detailed parameter list for a particular DB parameter group.

DescribeDBProxies

Returns information about DB proxies.

DescribeDBProxyEndpoints

Returns information about DB proxy endpoints.

DescribeDBProxyTargetGroups

Returns information about DB proxy target groups, represented by DBProxyTargetGroup data structures.

DescribeDBProxyTargets

Returns information about DBProxyTarget objects. This API supports pagination.

DescribeDBSecurityGroups

DescribeDBSnapshotAttributes

Returns a list of DB snapshot attribute names and values for a manual DB snapshot.

DescribeDBSnapshots

Returns information about DB snapshots. This API action supports pagination.

DescribeDBSubnetGroups

Returns a list of DBSubnetGroup descriptions. If a DBSubnetGroupName is specified, the list will contain only the descriptions of the specified DBSubnetGroup.

DescribeEngineDefaultClusterParameters

Returns the default engine and system parameter information for the cluster database engine.

DescribeEngineDefaultParameters

Returns the default engine and system parameter information for the specified database engine.

DescribeEventCategories

Displays a list of categories for all event source types, or, if specified, for a specified source type.

DescribeEventSubscriptions

Lists all the subscription descriptions for a customer account. The description for a subscription includes

DescribeEvents

Returns events related to DB instances, DB clusters, DB parameter groups, DB security groups, DB snapshots, and DB cluster snapshots for the past 14 days.

DescribeExportTasks

Returns information about a snapshot export to Amazon S3. This API operation supports

DescribeGlobalClusters

DescribeInstallationMedia

Describes the available installation media for a DB engine that requires an

DescribeOptionGroupOptions

Describes all available options.

DescribeOptionGroups

Describes the available option groups.

DescribeOrderableDBInstanceOptions

Returns a list of orderable DB instance options for the specified engine.

DescribePendingMaintenanceActions

Returns a list of resources (for example, DB instances) that have at least one pending maintenance action.

DescribeReservedDBInstances

Returns information about reserved DB instances for this account, or about a specified reserved DB instance.

DescribeReservedDBInstancesOfferings

Lists available reserved DB instance offerings.

DescribeSourceRegions

Returns a list of the source Amazon Web Services Regions where the current Amazon Web Services Region can create a read replica,

DescribeValidDBInstanceModifications

You can call DescribeValidDBInstanceModifications

DownloadDBLogFilePortion

Downloads all or a portion of the specified log file, up to 1 MB in size.

FailoverDBCluster

Forces a failover for a DB cluster.

FailoverGlobalCluster

Initiates the failover process for an Aurora global database (GlobalCluster).

ImportInstallationMedia

Imports the installation media for a DB engine that requires an on-premises

ListTagsForResource

Lists all tags on an Amazon RDS resource.

ModifyCertificates

Override the system-default Secure Sockets Layer/Transport Layer Security (SSL/TLS)

ModifyCurrentDBClusterCapacity

Set the capacity of an Aurora Serverless DB cluster to a specific value.

ModifyDBCluster

Modify a setting for an Amazon Aurora DB cluster.

ModifyDBClusterEndpoint

Modifies the properties of an endpoint in an Amazon Aurora DB cluster.

ModifyDBClusterParameterGroup

ModifyDBClusterSnapshotAttribute

Adds an attribute and values to, or removes an attribute and values from, a manual DB cluster snapshot.

ModifyDBInstance

Modifies settings for a DB instance.

ModifyDBParameterGroup

ModifyDBProxy

Changes the settings for an existing DB proxy.

ModifyDBProxyEndpoint

Changes the settings for an existing DB proxy endpoint.

ModifyDBProxyTargetGroup

Modifies the properties of a DBProxyTargetGroup.

ModifyDBSnapshot

Updates a manual DB snapshot with a new engine version. The snapshot can be encrypted

ModifyDBSnapshotAttribute

Adds an attribute and values to, or removes an attribute and values from, a manual DB snapshot.

ModifyDBSubnetGroup

Modifies an existing DB subnet group. DB subnet groups must contain at least one subnet in at least two AZs in the Amazon Web Services Region.

ModifyEventSubscription

Modifies an existing RDS event notification subscription. You can't modify the source identifiers using this call. To change

ModifyGlobalCluster

ModifyOptionGroup

Modifies an existing option group.

PromoteReadReplica

Promotes a read replica DB instance to a standalone DB instance.

PromoteReadReplicaDBCluster

Promotes a read replica DB cluster to a standalone DB cluster.

PurchaseReservedDBInstancesOffering

Purchases a reserved DB instance offering.

RebootDBInstance

You might need to reboot your DB instance, usually for maintenance reasons.

RegisterDBProxyTargets

Associate one or more DBProxyTarget data structures with a DBProxyTargetGroup.

RemoveFromGlobalCluster

RemoveRoleFromDBCluster

Disassociates an Amazon Web Services Identity and Access Management (IAM) role from an Amazon Aurora DB cluster.

RemoveRoleFromDBInstance

Disassociates an Amazon Web Services Identity and Access Management (IAM) role from a DB instance.

RemoveSourceIdentifierFromSubscription

Removes a source identifier from an existing RDS event notification subscription.

RemoveTagsFromResource

Removes metadata tags from an Amazon RDS resource.

ResetDBClusterParameterGroup

ResetDBParameterGroup

Modifies the parameters of a DB parameter group to the engine/system default value.

RestoreDBClusterFromS3

Creates an Amazon Aurora DB cluster from MySQL data stored in an Amazon S3 bucket.

RestoreDBClusterFromSnapshot

Creates a new DB cluster from a DB snapshot or DB cluster snapshot. This action

RestoreDBClusterToPointInTime

Restores a DB cluster to an arbitrary point in time. Users can restore to any point

RestoreDBInstanceFromDBSnapshot

Creates a new DB instance from a DB snapshot. The target database is created from the source database restore point with most

RestoreDBInstanceFromS3

Amazon Relational Database Service (Amazon RDS)

RestoreDBInstanceToPointInTime

Restores a DB instance to an arbitrary point in time. You can restore to any point in time before the time identified by the LatestRestorableTime property. You can restore to a point up to the number of days specified by the BackupRetentionPeriod property.

RevokeDBSecurityGroupIngress

Revokes ingress from a DBSecurityGroup for previously authorized IP ranges or EC2 or VPC security groups. Required parameters for this API are one of CIDRIP, EC2SecurityGroupId for VPC, or (EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId).

StartActivityStream

Starts a database activity stream to monitor activity on the database.

StartDBCluster

Starts an Amazon Aurora DB cluster that was stopped using the Amazon Web Services console, the stop-db-cluster

StartDBInstance

StartDBInstanceAutomatedBackupsReplication

Enables replication of automated backups to a different Amazon Web Services Region.

StartExportTask

Starts an export of a snapshot to Amazon S3.

StopActivityStream

Stops a database activity stream that was started using the Amazon Web Services console,

StopDBCluster

StopDBInstance

StopDBInstanceAutomatedBackupsReplication

Stops automated backup replication for a DB instance.