Struct aws_sdk_rds::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for AmazonRDSv19.

This client allows ergonomic access to a AmazonRDSv19-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

Creates a client with the given service configuration.

pub fn conf(&self) -> &Config

Returns the client’s configuration.

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn add_role_to_db_cluster(&self) -> AddRoleToDBCluster<C, M, R>

Constructs a fluent builder for the AddRoleToDBCluster operation.

See AddRoleToDBCluster for more information about the operation and its arguments.

pub fn add_role_to_db_instance(&self) -> AddRoleToDBInstance<C, M, R>

Constructs a fluent builder for the AddRoleToDBInstance operation.

See AddRoleToDBInstance for more information about the operation and its arguments.

pub fn add_source_identifier_to_subscription(
&self
) -> AddSourceIdentifierToSubscription<C, M, R>

Constructs a fluent builder for the AddSourceIdentifierToSubscription operation.

See AddSourceIdentifierToSubscription for more information about the operation and its arguments.

pub fn add_tags_to_resource(&self) -> AddTagsToResource<C, M, R>

Constructs a fluent builder for the AddTagsToResource operation.

See AddTagsToResource for more information about the operation and its arguments.

pub fn apply_pending_maintenance_action(
&self
) -> ApplyPendingMaintenanceAction<C, M, R>

Constructs a fluent builder for the ApplyPendingMaintenanceAction operation.

See ApplyPendingMaintenanceAction for more information about the operation and its arguments.

pub fn authorize_db_security_group_ingress(
&self
) -> AuthorizeDBSecurityGroupIngress<C, M, R>

Constructs a fluent builder for the AuthorizeDBSecurityGroupIngress operation.

See AuthorizeDBSecurityGroupIngress for more information about the operation and its arguments.

pub fn backtrack_db_cluster(&self) -> BacktrackDBCluster<C, M, R>

Constructs a fluent builder for the BacktrackDBCluster operation.

See BacktrackDBCluster for more information about the operation and its arguments.

pub fn cancel_export_task(&self) -> CancelExportTask<C, M, R>

Constructs a fluent builder for the CancelExportTask operation.

See CancelExportTask for more information about the operation and its arguments.

pub fn copy_db_cluster_parameter_group(
&self
) -> CopyDBClusterParameterGroup<C, M, R>

Constructs a fluent builder for the CopyDBClusterParameterGroup operation.

See CopyDBClusterParameterGroup for more information about the operation and its arguments.

pub fn copy_db_cluster_snapshot(&self) -> CopyDBClusterSnapshot<C, M, R>

Constructs a fluent builder for the CopyDBClusterSnapshot operation.

See CopyDBClusterSnapshot for more information about the operation and its arguments.

pub fn copy_db_parameter_group(&self) -> CopyDBParameterGroup<C, M, R>

Constructs a fluent builder for the CopyDBParameterGroup operation.

See CopyDBParameterGroup for more information about the operation and its arguments.

pub fn copy_db_snapshot(&self) -> CopyDBSnapshot<C, M, R>

Constructs a fluent builder for the CopyDBSnapshot operation.

See CopyDBSnapshot for more information about the operation and its arguments.

pub fn copy_option_group(&self) -> CopyOptionGroup<C, M, R>

Constructs a fluent builder for the CopyOptionGroup operation.

See CopyOptionGroup for more information about the operation and its arguments.

pub fn create_custom_availability_zone(
&self
) -> CreateCustomAvailabilityZone<C, M, R>

Constructs a fluent builder for the CreateCustomAvailabilityZone operation.

See CreateCustomAvailabilityZone for more information about the operation and its arguments.

pub fn create_custom_db_engine_version(
&self
) -> CreateCustomDBEngineVersion<C, M, R>

Constructs a fluent builder for the CreateCustomDBEngineVersion operation.

See CreateCustomDBEngineVersion for more information about the operation and its arguments.

pub fn create_db_cluster(&self) -> CreateDBCluster<C, M, R>

Constructs a fluent builder for the CreateDBCluster operation.

See CreateDBCluster for more information about the operation and its arguments.

pub fn create_db_cluster_endpoint(&self) -> CreateDBClusterEndpoint<C, M, R>

Constructs a fluent builder for the CreateDBClusterEndpoint operation.

See CreateDBClusterEndpoint for more information about the operation and its arguments.

pub fn create_db_cluster_parameter_group(
&self
) -> CreateDBClusterParameterGroup<C, M, R>

Constructs a fluent builder for the CreateDBClusterParameterGroup operation.

See CreateDBClusterParameterGroup for more information about the operation and its arguments.

pub fn create_db_cluster_snapshot(&self) -> CreateDBClusterSnapshot<C, M, R>

Constructs a fluent builder for the CreateDBClusterSnapshot operation.

See CreateDBClusterSnapshot for more information about the operation and its arguments.

pub fn create_db_instance(&self) -> CreateDBInstance<C, M, R>

Constructs a fluent builder for the CreateDBInstance operation.

See CreateDBInstance for more information about the operation and its arguments.

pub fn create_db_instance_read_replica(
&self
) -> CreateDBInstanceReadReplica<C, M, R>

Constructs a fluent builder for the CreateDBInstanceReadReplica operation.

See CreateDBInstanceReadReplica for more information about the operation and its arguments.

pub fn create_db_parameter_group(&self) -> CreateDBParameterGroup<C, M, R>

Constructs a fluent builder for the CreateDBParameterGroup operation.

See CreateDBParameterGroup for more information about the operation and its arguments.

pub fn create_db_proxy(&self) -> CreateDBProxy<C, M, R>

Constructs a fluent builder for the CreateDBProxy operation.

See CreateDBProxy for more information about the operation and its arguments.

pub fn create_db_proxy_endpoint(&self) -> CreateDBProxyEndpoint<C, M, R>

Constructs a fluent builder for the CreateDBProxyEndpoint operation.

See CreateDBProxyEndpoint for more information about the operation and its arguments.

pub fn create_db_security_group(&self) -> CreateDBSecurityGroup<C, M, R>

Constructs a fluent builder for the CreateDBSecurityGroup operation.

See CreateDBSecurityGroup for more information about the operation and its arguments.

pub fn create_db_snapshot(&self) -> CreateDBSnapshot<C, M, R>

Constructs a fluent builder for the CreateDBSnapshot operation.

See CreateDBSnapshot for more information about the operation and its arguments.

pub fn create_db_subnet_group(&self) -> CreateDBSubnetGroup<C, M, R>

Constructs a fluent builder for the CreateDBSubnetGroup operation.

See CreateDBSubnetGroup for more information about the operation and its arguments.

pub fn create_event_subscription(&self) -> CreateEventSubscription<C, M, R>

Constructs a fluent builder for the CreateEventSubscription operation.

See CreateEventSubscription for more information about the operation and its arguments.

pub fn create_global_cluster(&self) -> CreateGlobalCluster<C, M, R>

Constructs a fluent builder for the CreateGlobalCluster operation.

See CreateGlobalCluster for more information about the operation and its arguments.

pub fn create_option_group(&self) -> CreateOptionGroup<C, M, R>

Constructs a fluent builder for the CreateOptionGroup operation.

See CreateOptionGroup for more information about the operation and its arguments.

pub fn delete_custom_availability_zone(
&self
) -> DeleteCustomAvailabilityZone<C, M, R>

Constructs a fluent builder for the DeleteCustomAvailabilityZone operation.

See DeleteCustomAvailabilityZone for more information about the operation and its arguments.

pub fn delete_custom_db_engine_version(
&self
) -> DeleteCustomDBEngineVersion<C, M, R>

Constructs a fluent builder for the DeleteCustomDBEngineVersion operation.

See DeleteCustomDBEngineVersion for more information about the operation and its arguments.

pub fn delete_db_cluster(&self) -> DeleteDBCluster<C, M, R>

Constructs a fluent builder for the DeleteDBCluster operation.

See DeleteDBCluster for more information about the operation and its arguments.

pub fn delete_db_cluster_endpoint(&self) -> DeleteDBClusterEndpoint<C, M, R>

Constructs a fluent builder for the DeleteDBClusterEndpoint operation.

See DeleteDBClusterEndpoint for more information about the operation and its arguments.

pub fn delete_db_cluster_parameter_group(
&self
) -> DeleteDBClusterParameterGroup<C, M, R>

Constructs a fluent builder for the DeleteDBClusterParameterGroup operation.

See DeleteDBClusterParameterGroup for more information about the operation and its arguments.

pub fn delete_db_cluster_snapshot(&self) -> DeleteDBClusterSnapshot<C, M, R>

Constructs a fluent builder for the DeleteDBClusterSnapshot operation.

See DeleteDBClusterSnapshot for more information about the operation and its arguments.

pub fn delete_db_instance(&self) -> DeleteDBInstance<C, M, R>

Constructs a fluent builder for the DeleteDBInstance operation.

See DeleteDBInstance for more information about the operation and its arguments.

pub fn delete_db_instance_automated_backup(
&self
) -> DeleteDBInstanceAutomatedBackup<C, M, R>

Constructs a fluent builder for the DeleteDBInstanceAutomatedBackup operation.

See DeleteDBInstanceAutomatedBackup for more information about the operation and its arguments.

pub fn delete_db_parameter_group(&self) -> DeleteDBParameterGroup<C, M, R>

Constructs a fluent builder for the DeleteDBParameterGroup operation.

See DeleteDBParameterGroup for more information about the operation and its arguments.

pub fn delete_db_proxy(&self) -> DeleteDBProxy<C, M, R>

Constructs a fluent builder for the DeleteDBProxy operation.

See DeleteDBProxy for more information about the operation and its arguments.

pub fn delete_db_proxy_endpoint(&self) -> DeleteDBProxyEndpoint<C, M, R>

Constructs a fluent builder for the DeleteDBProxyEndpoint operation.

See DeleteDBProxyEndpoint for more information about the operation and its arguments.

pub fn delete_db_security_group(&self) -> DeleteDBSecurityGroup<C, M, R>

Constructs a fluent builder for the DeleteDBSecurityGroup operation.

See DeleteDBSecurityGroup for more information about the operation and its arguments.

pub fn delete_db_snapshot(&self) -> DeleteDBSnapshot<C, M, R>

Constructs a fluent builder for the DeleteDBSnapshot operation.

See DeleteDBSnapshot for more information about the operation and its arguments.

pub fn delete_db_subnet_group(&self) -> DeleteDBSubnetGroup<C, M, R>

Constructs a fluent builder for the DeleteDBSubnetGroup operation.

See DeleteDBSubnetGroup for more information about the operation and its arguments.

pub fn delete_event_subscription(&self) -> DeleteEventSubscription<C, M, R>

Constructs a fluent builder for the DeleteEventSubscription operation.

See DeleteEventSubscription for more information about the operation and its arguments.

pub fn delete_global_cluster(&self) -> DeleteGlobalCluster<C, M, R>

Constructs a fluent builder for the DeleteGlobalCluster operation.

See DeleteGlobalCluster for more information about the operation and its arguments.

pub fn delete_installation_media(&self) -> DeleteInstallationMedia<C, M, R>

Constructs a fluent builder for the DeleteInstallationMedia operation.

See DeleteInstallationMedia for more information about the operation and its arguments.

pub fn delete_option_group(&self) -> DeleteOptionGroup<C, M, R>

Constructs a fluent builder for the DeleteOptionGroup operation.

See DeleteOptionGroup for more information about the operation and its arguments.

pub fn deregister_db_proxy_targets(&self) -> DeregisterDBProxyTargets<C, M, R>

Constructs a fluent builder for the DeregisterDBProxyTargets operation.

See DeregisterDBProxyTargets for more information about the operation and its arguments.

pub fn describe_account_attributes(&self) -> DescribeAccountAttributes<C, M, R>

Constructs a fluent builder for the DescribeAccountAttributes operation.

See DescribeAccountAttributes for more information about the operation and its arguments.

pub fn describe_certificates(&self) -> DescribeCertificates<C, M, R>

Constructs a fluent builder for the DescribeCertificates operation.

See DescribeCertificates for more information about the operation and its arguments.

pub fn describe_custom_availability_zones(
&self
) -> DescribeCustomAvailabilityZones<C, M, R>

Constructs a fluent builder for the DescribeCustomAvailabilityZones operation.

See DescribeCustomAvailabilityZones for more information about the operation and its arguments.

pub fn describe_db_cluster_backtracks(
&self
) -> DescribeDBClusterBacktracks<C, M, R>

Constructs a fluent builder for the DescribeDBClusterBacktracks operation.

See DescribeDBClusterBacktracks for more information about the operation and its arguments.

pub fn describe_db_cluster_endpoints(
&self
) -> DescribeDBClusterEndpoints<C, M, R>

Constructs a fluent builder for the DescribeDBClusterEndpoints operation.

See DescribeDBClusterEndpoints for more information about the operation and its arguments.

pub fn describe_db_cluster_parameter_groups(
&self
) -> DescribeDBClusterParameterGroups<C, M, R>

Constructs a fluent builder for the DescribeDBClusterParameterGroups operation.

See DescribeDBClusterParameterGroups for more information about the operation and its arguments.

pub fn describe_db_cluster_parameters(
&self
) -> DescribeDBClusterParameters<C, M, R>

Constructs a fluent builder for the DescribeDBClusterParameters operation.

See DescribeDBClusterParameters for more information about the operation and its arguments.

pub fn describe_db_clusters(&self) -> DescribeDBClusters<C, M, R>

Constructs a fluent builder for the DescribeDBClusters operation.

See DescribeDBClusters for more information about the operation and its arguments.

pub fn describe_db_cluster_snapshot_attributes(
&self
) -> DescribeDBClusterSnapshotAttributes<C, M, R>

Constructs a fluent builder for the DescribeDBClusterSnapshotAttributes operation.

See DescribeDBClusterSnapshotAttributes for more information about the operation and its arguments.

pub fn describe_db_cluster_snapshots(
&self
) -> DescribeDBClusterSnapshots<C, M, R>

Constructs a fluent builder for the DescribeDBClusterSnapshots operation.

See DescribeDBClusterSnapshots for more information about the operation and its arguments.

pub fn describe_db_engine_versions(&self) -> DescribeDBEngineVersions<C, M, R>

Constructs a fluent builder for the DescribeDBEngineVersions operation.

See DescribeDBEngineVersions for more information about the operation and its arguments.

pub fn describe_db_instance_automated_backups(
&self
) -> DescribeDBInstanceAutomatedBackups<C, M, R>

Constructs a fluent builder for the DescribeDBInstanceAutomatedBackups operation.

See DescribeDBInstanceAutomatedBackups for more information about the operation and its arguments.

pub fn describe_db_instances(&self) -> DescribeDBInstances<C, M, R>

Constructs a fluent builder for the DescribeDBInstances operation.

See DescribeDBInstances for more information about the operation and its arguments.

pub fn describe_db_log_files(&self) -> DescribeDBLogFiles<C, M, R>

Constructs a fluent builder for the DescribeDBLogFiles operation.

See DescribeDBLogFiles for more information about the operation and its arguments.

pub fn describe_db_parameter_groups(&self) -> DescribeDBParameterGroups<C, M, R>

Constructs a fluent builder for the DescribeDBParameterGroups operation.

See DescribeDBParameterGroups for more information about the operation and its arguments.

pub fn describe_db_parameters(&self) -> DescribeDBParameters<C, M, R>

Constructs a fluent builder for the DescribeDBParameters operation.

See DescribeDBParameters for more information about the operation and its arguments.

pub fn describe_db_proxies(&self) -> DescribeDBProxies<C, M, R>

Constructs a fluent builder for the DescribeDBProxies operation.

See DescribeDBProxies for more information about the operation and its arguments.

pub fn describe_db_proxy_endpoints(&self) -> DescribeDBProxyEndpoints<C, M, R>

Constructs a fluent builder for the DescribeDBProxyEndpoints operation.

See DescribeDBProxyEndpoints for more information about the operation and its arguments.

pub fn describe_db_proxy_target_groups(
&self
) -> DescribeDBProxyTargetGroups<C, M, R>

Constructs a fluent builder for the DescribeDBProxyTargetGroups operation.

See DescribeDBProxyTargetGroups for more information about the operation and its arguments.

pub fn describe_db_proxy_targets(&self) -> DescribeDBProxyTargets<C, M, R>

Constructs a fluent builder for the DescribeDBProxyTargets operation.

See DescribeDBProxyTargets for more information about the operation and its arguments.

pub fn describe_db_security_groups(&self) -> DescribeDBSecurityGroups<C, M, R>

Constructs a fluent builder for the DescribeDBSecurityGroups operation.

See DescribeDBSecurityGroups for more information about the operation and its arguments.

pub fn describe_db_snapshot_attributes(
&self
) -> DescribeDBSnapshotAttributes<C, M, R>

Constructs a fluent builder for the DescribeDBSnapshotAttributes operation.

See DescribeDBSnapshotAttributes for more information about the operation and its arguments.

pub fn describe_db_snapshots(&self) -> DescribeDBSnapshots<C, M, R>

Constructs a fluent builder for the DescribeDBSnapshots operation.

See DescribeDBSnapshots for more information about the operation and its arguments.

pub fn describe_db_subnet_groups(&self) -> DescribeDBSubnetGroups<C, M, R>

Constructs a fluent builder for the DescribeDBSubnetGroups operation.

See DescribeDBSubnetGroups for more information about the operation and its arguments.

pub fn describe_engine_default_cluster_parameters(
&self
) -> DescribeEngineDefaultClusterParameters<C, M, R>

Constructs a fluent builder for the DescribeEngineDefaultClusterParameters operation.

See DescribeEngineDefaultClusterParameters for more information about the operation and its arguments.

pub fn describe_engine_default_parameters(
&self
) -> DescribeEngineDefaultParameters<C, M, R>

Constructs a fluent builder for the DescribeEngineDefaultParameters operation.

See DescribeEngineDefaultParameters for more information about the operation and its arguments.

pub fn describe_event_categories(&self) -> DescribeEventCategories<C, M, R>

Constructs a fluent builder for the DescribeEventCategories operation.

See DescribeEventCategories for more information about the operation and its arguments.

pub fn describe_events(&self) -> DescribeEvents<C, M, R>

Constructs a fluent builder for the DescribeEvents operation.

See DescribeEvents for more information about the operation and its arguments.

pub fn describe_event_subscriptions(
&self
) -> DescribeEventSubscriptions<C, M, R>

Constructs a fluent builder for the DescribeEventSubscriptions operation.

See DescribeEventSubscriptions for more information about the operation and its arguments.

pub fn describe_export_tasks(&self) -> DescribeExportTasks<C, M, R>

Constructs a fluent builder for the DescribeExportTasks operation.

See DescribeExportTasks for more information about the operation and its arguments.

pub fn describe_global_clusters(&self) -> DescribeGlobalClusters<C, M, R>

Constructs a fluent builder for the DescribeGlobalClusters operation.

See DescribeGlobalClusters for more information about the operation and its arguments.

pub fn describe_installation_media(&self) -> DescribeInstallationMedia<C, M, R>

Constructs a fluent builder for the DescribeInstallationMedia operation.

See DescribeInstallationMedia for more information about the operation and its arguments.

pub fn describe_option_group_options(
&self
) -> DescribeOptionGroupOptions<C, M, R>

Constructs a fluent builder for the DescribeOptionGroupOptions operation.

See DescribeOptionGroupOptions for more information about the operation and its arguments.

pub fn describe_option_groups(&self) -> DescribeOptionGroups<C, M, R>

Constructs a fluent builder for the DescribeOptionGroups operation.

See DescribeOptionGroups for more information about the operation and its arguments.

pub fn describe_orderable_db_instance_options(
&self
) -> DescribeOrderableDBInstanceOptions<C, M, R>

Constructs a fluent builder for the DescribeOrderableDBInstanceOptions operation.

See DescribeOrderableDBInstanceOptions for more information about the operation and its arguments.

pub fn describe_pending_maintenance_actions(
&self
) -> DescribePendingMaintenanceActions<C, M, R>

Constructs a fluent builder for the DescribePendingMaintenanceActions operation.

See DescribePendingMaintenanceActions for more information about the operation and its arguments.

pub fn describe_reserved_db_instances(
&self
) -> DescribeReservedDBInstances<C, M, R>

Constructs a fluent builder for the DescribeReservedDBInstances operation.

See DescribeReservedDBInstances for more information about the operation and its arguments.

pub fn describe_reserved_db_instances_offerings(
&self
) -> DescribeReservedDBInstancesOfferings<C, M, R>

Constructs a fluent builder for the DescribeReservedDBInstancesOfferings operation.

See DescribeReservedDBInstancesOfferings for more information about the operation and its arguments.

pub fn describe_source_regions(&self) -> DescribeSourceRegions<C, M, R>

Constructs a fluent builder for the DescribeSourceRegions operation.

See DescribeSourceRegions for more information about the operation and its arguments.

pub fn describe_valid_db_instance_modifications(
&self
) -> DescribeValidDBInstanceModifications<C, M, R>

Constructs a fluent builder for the DescribeValidDBInstanceModifications operation.

See DescribeValidDBInstanceModifications for more information about the operation and its arguments.

pub fn download_db_log_file_portion(&self) -> DownloadDBLogFilePortion<C, M, R>

Constructs a fluent builder for the DownloadDBLogFilePortion operation.

See DownloadDBLogFilePortion for more information about the operation and its arguments.

pub fn failover_db_cluster(&self) -> FailoverDBCluster<C, M, R>

Constructs a fluent builder for the FailoverDBCluster operation.

See FailoverDBCluster for more information about the operation and its arguments.

pub fn failover_global_cluster(&self) -> FailoverGlobalCluster<C, M, R>

Constructs a fluent builder for the FailoverGlobalCluster operation.

See FailoverGlobalCluster for more information about the operation and its arguments.

pub fn import_installation_media(&self) -> ImportInstallationMedia<C, M, R>

Constructs a fluent builder for the ImportInstallationMedia operation.

See ImportInstallationMedia for more information about the operation and its arguments.

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

Constructs a fluent builder for the ListTagsForResource operation.

See ListTagsForResource for more information about the operation and its arguments.

pub fn modify_certificates(&self) -> ModifyCertificates<C, M, R>

Constructs a fluent builder for the ModifyCertificates operation.

See ModifyCertificates for more information about the operation and its arguments.

pub fn modify_current_db_cluster_capacity(
&self
) -> ModifyCurrentDBClusterCapacity<C, M, R>

Constructs a fluent builder for the ModifyCurrentDBClusterCapacity operation.

See ModifyCurrentDBClusterCapacity for more information about the operation and its arguments.

pub fn modify_custom_db_engine_version(
&self
) -> ModifyCustomDBEngineVersion<C, M, R>

Constructs a fluent builder for the ModifyCustomDBEngineVersion operation.

See ModifyCustomDBEngineVersion for more information about the operation and its arguments.

pub fn modify_db_cluster(&self) -> ModifyDBCluster<C, M, R>

Constructs a fluent builder for the ModifyDBCluster operation.

See ModifyDBCluster for more information about the operation and its arguments.

pub fn modify_db_cluster_endpoint(&self) -> ModifyDBClusterEndpoint<C, M, R>

Constructs a fluent builder for the ModifyDBClusterEndpoint operation.

See ModifyDBClusterEndpoint for more information about the operation and its arguments.

pub fn modify_db_cluster_parameter_group(
&self
) -> ModifyDBClusterParameterGroup<C, M, R>

Constructs a fluent builder for the ModifyDBClusterParameterGroup operation.

See ModifyDBClusterParameterGroup for more information about the operation and its arguments.

pub fn modify_db_cluster_snapshot_attribute(
&self
) -> ModifyDBClusterSnapshotAttribute<C, M, R>

Constructs a fluent builder for the ModifyDBClusterSnapshotAttribute operation.

See ModifyDBClusterSnapshotAttribute for more information about the operation and its arguments.

pub fn modify_db_instance(&self) -> ModifyDBInstance<C, M, R>

Constructs a fluent builder for the ModifyDBInstance operation.

See ModifyDBInstance for more information about the operation and its arguments.

pub fn modify_db_parameter_group(&self) -> ModifyDBParameterGroup<C, M, R>

Constructs a fluent builder for the ModifyDBParameterGroup operation.

See ModifyDBParameterGroup for more information about the operation and its arguments.

pub fn modify_db_proxy(&self) -> ModifyDBProxy<C, M, R>

Constructs a fluent builder for the ModifyDBProxy operation.

See ModifyDBProxy for more information about the operation and its arguments.

pub fn modify_db_proxy_endpoint(&self) -> ModifyDBProxyEndpoint<C, M, R>

Constructs a fluent builder for the ModifyDBProxyEndpoint operation.

See ModifyDBProxyEndpoint for more information about the operation and its arguments.

pub fn modify_db_proxy_target_group(&self) -> ModifyDBProxyTargetGroup<C, M, R>

Constructs a fluent builder for the ModifyDBProxyTargetGroup operation.

See ModifyDBProxyTargetGroup for more information about the operation and its arguments.

pub fn modify_db_snapshot(&self) -> ModifyDBSnapshot<C, M, R>

Constructs a fluent builder for the ModifyDBSnapshot operation.

See ModifyDBSnapshot for more information about the operation and its arguments.

pub fn modify_db_snapshot_attribute(&self) -> ModifyDBSnapshotAttribute<C, M, R>

Constructs a fluent builder for the ModifyDBSnapshotAttribute operation.

See ModifyDBSnapshotAttribute for more information about the operation and its arguments.

pub fn modify_db_subnet_group(&self) -> ModifyDBSubnetGroup<C, M, R>

Constructs a fluent builder for the ModifyDBSubnetGroup operation.

See ModifyDBSubnetGroup for more information about the operation and its arguments.

pub fn modify_event_subscription(&self) -> ModifyEventSubscription<C, M, R>

Constructs a fluent builder for the ModifyEventSubscription operation.

See ModifyEventSubscription for more information about the operation and its arguments.

pub fn modify_global_cluster(&self) -> ModifyGlobalCluster<C, M, R>

Constructs a fluent builder for the ModifyGlobalCluster operation.

See ModifyGlobalCluster for more information about the operation and its arguments.

pub fn modify_option_group(&self) -> ModifyOptionGroup<C, M, R>

Constructs a fluent builder for the ModifyOptionGroup operation.

See ModifyOptionGroup for more information about the operation and its arguments.

pub fn promote_read_replica(&self) -> PromoteReadReplica<C, M, R>

Constructs a fluent builder for the PromoteReadReplica operation.

See PromoteReadReplica for more information about the operation and its arguments.

pub fn promote_read_replica_db_cluster(
&self
) -> PromoteReadReplicaDBCluster<C, M, R>

Constructs a fluent builder for the PromoteReadReplicaDBCluster operation.

See PromoteReadReplicaDBCluster for more information about the operation and its arguments.

pub fn purchase_reserved_db_instances_offering(
&self
) -> PurchaseReservedDBInstancesOffering<C, M, R>

Constructs a fluent builder for the PurchaseReservedDBInstancesOffering operation.

See PurchaseReservedDBInstancesOffering for more information about the operation and its arguments.

pub fn reboot_db_instance(&self) -> RebootDBInstance<C, M, R>

Constructs a fluent builder for the RebootDBInstance operation.

See RebootDBInstance for more information about the operation and its arguments.

pub fn register_db_proxy_targets(&self) -> RegisterDBProxyTargets<C, M, R>

Constructs a fluent builder for the RegisterDBProxyTargets operation.

See RegisterDBProxyTargets for more information about the operation and its arguments.

pub fn remove_from_global_cluster(&self) -> RemoveFromGlobalCluster<C, M, R>

Constructs a fluent builder for the RemoveFromGlobalCluster operation.

See RemoveFromGlobalCluster for more information about the operation and its arguments.

pub fn remove_role_from_db_cluster(&self) -> RemoveRoleFromDBCluster<C, M, R>

Constructs a fluent builder for the RemoveRoleFromDBCluster operation.

See RemoveRoleFromDBCluster for more information about the operation and its arguments.

pub fn remove_role_from_db_instance(&self) -> RemoveRoleFromDBInstance<C, M, R>

Constructs a fluent builder for the RemoveRoleFromDBInstance operation.

See RemoveRoleFromDBInstance for more information about the operation and its arguments.

pub fn remove_source_identifier_from_subscription(
&self
) -> RemoveSourceIdentifierFromSubscription<C, M, R>

Constructs a fluent builder for the RemoveSourceIdentifierFromSubscription operation.

See RemoveSourceIdentifierFromSubscription for more information about the operation and its arguments.

pub fn remove_tags_from_resource(&self) -> RemoveTagsFromResource<C, M, R>

Constructs a fluent builder for the RemoveTagsFromResource operation.

See RemoveTagsFromResource for more information about the operation and its arguments.

pub fn reset_db_cluster_parameter_group(
&self
) -> ResetDBClusterParameterGroup<C, M, R>

Constructs a fluent builder for the ResetDBClusterParameterGroup operation.

See ResetDBClusterParameterGroup for more information about the operation and its arguments.

pub fn reset_db_parameter_group(&self) -> ResetDBParameterGroup<C, M, R>

Constructs a fluent builder for the ResetDBParameterGroup operation.

See ResetDBParameterGroup for more information about the operation and its arguments.

pub fn restore_db_cluster_from_s3(&self) -> RestoreDBClusterFromS3<C, M, R>

Constructs a fluent builder for the RestoreDBClusterFromS3 operation.

See RestoreDBClusterFromS3 for more information about the operation and its arguments.

pub fn restore_db_cluster_from_snapshot(
&self
) -> RestoreDBClusterFromSnapshot<C, M, R>

Constructs a fluent builder for the RestoreDBClusterFromSnapshot operation.

See RestoreDBClusterFromSnapshot for more information about the operation and its arguments.

pub fn restore_db_cluster_to_point_in_time(
&self
) -> RestoreDBClusterToPointInTime<C, M, R>

Constructs a fluent builder for the RestoreDBClusterToPointInTime operation.

See RestoreDBClusterToPointInTime for more information about the operation and its arguments.

pub fn restore_db_instance_from_db_snapshot(
&self
) -> RestoreDBInstanceFromDBSnapshot<C, M, R>

Constructs a fluent builder for the RestoreDBInstanceFromDBSnapshot operation.

See RestoreDBInstanceFromDBSnapshot for more information about the operation and its arguments.

pub fn restore_db_instance_from_s3(&self) -> RestoreDBInstanceFromS3<C, M, R>

Constructs a fluent builder for the RestoreDBInstanceFromS3 operation.

See RestoreDBInstanceFromS3 for more information about the operation and its arguments.

pub fn restore_db_instance_to_point_in_time(
&self
) -> RestoreDBInstanceToPointInTime<C, M, R>

Constructs a fluent builder for the RestoreDBInstanceToPointInTime operation.

See RestoreDBInstanceToPointInTime for more information about the operation and its arguments.

pub fn revoke_db_security_group_ingress(
&self
) -> RevokeDBSecurityGroupIngress<C, M, R>

Constructs a fluent builder for the RevokeDBSecurityGroupIngress operation.

See RevokeDBSecurityGroupIngress for more information about the operation and its arguments.

pub fn start_activity_stream(&self) -> StartActivityStream<C, M, R>

Constructs a fluent builder for the StartActivityStream operation.

See StartActivityStream for more information about the operation and its arguments.

pub fn start_db_cluster(&self) -> StartDBCluster<C, M, R>

Constructs a fluent builder for the StartDBCluster operation.

See StartDBCluster for more information about the operation and its arguments.

pub fn start_db_instance(&self) -> StartDBInstance<C, M, R>

Constructs a fluent builder for the StartDBInstance operation.

See StartDBInstance for more information about the operation and its arguments.

pub fn start_db_instance_automated_backups_replication(
&self
) -> StartDBInstanceAutomatedBackupsReplication<C, M, R>

Constructs a fluent builder for the StartDBInstanceAutomatedBackupsReplication operation.

See StartDBInstanceAutomatedBackupsReplication for more information about the operation and its arguments.

pub fn start_export_task(&self) -> StartExportTask<C, M, R>

Constructs a fluent builder for the StartExportTask operation.

See StartExportTask for more information about the operation and its arguments.

pub fn stop_activity_stream(&self) -> StopActivityStream<C, M, R>

Constructs a fluent builder for the StopActivityStream operation.

See StopActivityStream for more information about the operation and its arguments.

pub fn stop_db_cluster(&self) -> StopDBCluster<C, M, R>

Constructs a fluent builder for the StopDBCluster operation.

See StopDBCluster for more information about the operation and its arguments.

pub fn stop_db_instance(&self) -> StopDBInstance<C, M, R>

Constructs a fluent builder for the StopDBInstance operation.

See StopDBInstance for more information about the operation and its arguments.

pub fn stop_db_instance_automated_backups_replication(
&self
) -> StopDBInstanceAutomatedBackupsReplication<C, M, R>

Constructs a fluent builder for the StopDBInstanceAutomatedBackupsReplication operation.

See StopDBInstanceAutomatedBackupsReplication for more information about the operation and its arguments.

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

Creates a client with the given service config and connector override.

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

Creates a new client from a shared config.

pub fn from_conf(conf: Config) -> Self

Creates a new client from the service Config.