Struct aws_sdk_rds::model::db_cluster::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for DbCluster

Implementations

impl Builder

pub fn allocated_storage(self, input: i32) -> Self

For all database engines except Amazon Aurora, AllocatedStorage specifies the allocated storage size in gibibytes (GiB). For Aurora, AllocatedStorage always returns 1, because Aurora DB cluster storage size isn't fixed, but instead automatically adjusts as needed.

pub fn set_allocated_storage(self, input: Option<i32>) -> Self

For all database engines except Amazon Aurora, AllocatedStorage specifies the allocated storage size in gibibytes (GiB). For Aurora, AllocatedStorage always returns 1, because Aurora DB cluster storage size isn't fixed, but instead automatically adjusts as needed.

pub fn availability_zones(self, input: impl Into<String>) -> Self

Appends an item to availability_zones.

To override the contents of this collection use set_availability_zones.

Provides the list of Availability Zones (AZs) where instances in the DB cluster can be created.

pub fn set_availability_zones(self, input: Option<Vec<String>>) -> Self

Provides the list of Availability Zones (AZs) where instances in the DB cluster can be created.

pub fn backup_retention_period(self, input: i32) -> Self

Specifies the number of days for which automatic DB snapshots are retained.

pub fn set_backup_retention_period(self, input: Option<i32>) -> Self

Specifies the number of days for which automatic DB snapshots are retained.

pub fn character_set_name(self, input: impl Into<String>) -> Self

If present, specifies the name of the character set that this cluster is associated with.

pub fn set_character_set_name(self, input: Option<String>) -> Self

If present, specifies the name of the character set that this cluster is associated with.

pub fn database_name(self, input: impl Into<String>) -> Self

Contains the name of the initial database of this DB cluster that was provided at create time, if one was specified when the DB cluster was created. This same name is returned for the life of the DB cluster.

pub fn set_database_name(self, input: Option<String>) -> Self

Contains the name of the initial database of this DB cluster that was provided at create time, if one was specified when the DB cluster was created. This same name is returned for the life of the DB cluster.

pub fn db_cluster_identifier(self, input: impl Into<String>) -> Self

Contains a user-supplied DB cluster identifier. This identifier is the unique key that identifies a DB cluster.

pub fn set_db_cluster_identifier(self, input: Option<String>) -> Self

Contains a user-supplied DB cluster identifier. This identifier is the unique key that identifies a DB cluster.

pub fn db_cluster_parameter_group(self, input: impl Into<String>) -> Self

Specifies the name of the DB cluster parameter group for the DB cluster.

pub fn set_db_cluster_parameter_group(self, input: Option<String>) -> Self

Specifies the name of the DB cluster parameter group for the DB cluster.

pub fn db_subnet_group(self, input: impl Into<String>) -> Self

Specifies information on the subnet group associated with the DB cluster, including the name, description, and subnets in the subnet group.

pub fn set_db_subnet_group(self, input: Option<String>) -> Self

Specifies information on the subnet group associated with the DB cluster, including the name, description, and subnets in the subnet group.

pub fn status(self, input: impl Into<String>) -> Self

Specifies the current state of this DB cluster.

pub fn set_status(self, input: Option<String>) -> Self

Specifies the current state of this DB cluster.

pub fn automatic_restart_time(self, input: Instant) -> Self

The time when a stopped DB cluster is restarted automatically.

pub fn set_automatic_restart_time(self, input: Option<Instant>) -> Self

The time when a stopped DB cluster is restarted automatically.

pub fn percent_progress(self, input: impl Into<String>) -> Self

Specifies the progress of the operation as a percentage.

pub fn set_percent_progress(self, input: Option<String>) -> Self

Specifies the progress of the operation as a percentage.

pub fn earliest_restorable_time(self, input: Instant) -> Self

The earliest time to which a database can be restored with point-in-time restore.

pub fn set_earliest_restorable_time(self, input: Option<Instant>) -> Self

The earliest time to which a database can be restored with point-in-time restore.

pub fn endpoint(self, input: impl Into<String>) -> Self

Specifies the connection endpoint for the primary instance of the DB cluster.

pub fn set_endpoint(self, input: Option<String>) -> Self

Specifies the connection endpoint for the primary instance of the DB cluster.

pub fn reader_endpoint(self, input: impl Into<String>) -> Self

The reader endpoint for the DB cluster. The reader endpoint for a DB cluster load-balances connections across the Aurora Replicas that are available in a DB cluster. As clients request new connections to the reader endpoint, Aurora distributes the connection requests among the Aurora Replicas in the DB cluster. This functionality can help balance your read workload across multiple Aurora Replicas in your DB cluster.

If a failover occurs, and the Aurora Replica that you are connected to is promoted to be the primary instance, your connection is dropped. To continue sending your read workload to other Aurora Replicas in the cluster, you can then reconnect to the reader endpoint.

pub fn set_reader_endpoint(self, input: Option<String>) -> Self

The reader endpoint for the DB cluster. The reader endpoint for a DB cluster load-balances connections across the Aurora Replicas that are available in a DB cluster. As clients request new connections to the reader endpoint, Aurora distributes the connection requests among the Aurora Replicas in the DB cluster. This functionality can help balance your read workload across multiple Aurora Replicas in your DB cluster.

If a failover occurs, and the Aurora Replica that you are connected to is promoted to be the primary instance, your connection is dropped. To continue sending your read workload to other Aurora Replicas in the cluster, you can then reconnect to the reader endpoint.

pub fn custom_endpoints(self, input: impl Into<String>) -> Self

Appends an item to custom_endpoints.

To override the contents of this collection use set_custom_endpoints.

Identifies all custom endpoints associated with the cluster.

pub fn set_custom_endpoints(self, input: Option<Vec<String>>) -> Self

Identifies all custom endpoints associated with the cluster.

pub fn multi_az(self, input: bool) -> Self

Specifies whether the DB cluster has instances in multiple Availability Zones.

pub fn set_multi_az(self, input: Option<bool>) -> Self

Specifies whether the DB cluster has instances in multiple Availability Zones.

pub fn engine(self, input: impl Into<String>) -> Self

The name of the database engine to be used for this DB cluster.

pub fn set_engine(self, input: Option<String>) -> Self

The name of the database engine to be used for this DB cluster.

pub fn engine_version(self, input: impl Into<String>) -> Self

Indicates the database engine version.

pub fn set_engine_version(self, input: Option<String>) -> Self

Indicates the database engine version.

pub fn latest_restorable_time(self, input: Instant) -> Self

Specifies the latest time to which a database can be restored with point-in-time restore.

pub fn set_latest_restorable_time(self, input: Option<Instant>) -> Self

Specifies the latest time to which a database can be restored with point-in-time restore.

pub fn port(self, input: i32) -> Self

Specifies the port that the database engine is listening on.

pub fn set_port(self, input: Option<i32>) -> Self

Specifies the port that the database engine is listening on.

pub fn master_username(self, input: impl Into<String>) -> Self

Contains the master username for the DB cluster.

pub fn set_master_username(self, input: Option<String>) -> Self

Contains the master username for the DB cluster.

pub fn db_cluster_option_group_memberships(
self,
input: impl Into<DbClusterOptionGroupStatus>
) -> Self

Appends an item to db_cluster_option_group_memberships.

To override the contents of this collection use set_db_cluster_option_group_memberships.

Provides the list of option group memberships for this DB cluster.

pub fn set_db_cluster_option_group_memberships(
self,
input: Option<Vec<DbClusterOptionGroupStatus>>
) -> Self

Provides the list of option group memberships for this DB cluster.

pub fn preferred_backup_window(self, input: impl Into<String>) -> Self

Specifies the daily time range during which automated backups are created if automated backups are enabled, as determined by the BackupRetentionPeriod.

pub fn set_preferred_backup_window(self, input: Option<String>) -> Self

Specifies the daily time range during which automated backups are created if automated backups are enabled, as determined by the BackupRetentionPeriod.

pub fn preferred_maintenance_window(self, input: impl Into<String>) -> Self

Specifies the weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

pub fn set_preferred_maintenance_window(self, input: Option<String>) -> Self

Specifies the weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

pub fn replication_source_identifier(self, input: impl Into<String>) -> Self

Contains the identifier of the source DB cluster if this DB cluster is a read replica.

pub fn set_replication_source_identifier(self, input: Option<String>) -> Self

Contains the identifier of the source DB cluster if this DB cluster is a read replica.

pub fn read_replica_identifiers(self, input: impl Into<String>) -> Self

Appends an item to read_replica_identifiers.

To override the contents of this collection use set_read_replica_identifiers.

Contains one or more identifiers of the read replicas associated with this DB cluster.

pub fn set_read_replica_identifiers(self, input: Option<Vec<String>>) -> Self

Contains one or more identifiers of the read replicas associated with this DB cluster.

pub fn db_cluster_members(self, input: impl Into<DbClusterMember>) -> Self

Appends an item to db_cluster_members.

To override the contents of this collection use set_db_cluster_members.

Provides the list of instances that make up the DB cluster.

pub fn set_db_cluster_members(self, input: Option<Vec<DbClusterMember>>) -> Self

Provides the list of instances that make up the DB cluster.

pub fn vpc_security_groups(
self,
input: impl Into<VpcSecurityGroupMembership>
) -> Self

Appends an item to vpc_security_groups.

To override the contents of this collection use set_vpc_security_groups.

Provides a list of VPC security groups that the DB cluster belongs to.

pub fn set_vpc_security_groups(
self,
input: Option<Vec<VpcSecurityGroupMembership>>
) -> Self

Provides a list of VPC security groups that the DB cluster belongs to.

pub fn hosted_zone_id(self, input: impl Into<String>) -> Self

Specifies the ID that Amazon Route 53 assigns when you create a hosted zone.

pub fn set_hosted_zone_id(self, input: Option<String>) -> Self

Specifies the ID that Amazon Route 53 assigns when you create a hosted zone.

pub fn storage_encrypted(self, input: bool) -> Self

Specifies whether the DB cluster is encrypted.

pub fn set_storage_encrypted(self, input: Option<bool>) -> Self

Specifies whether the DB cluster is encrypted.

pub fn kms_key_id(self, input: impl Into<String>) -> Self

If StorageEncrypted is enabled, the Amazon Web Services KMS key identifier for the encrypted DB cluster.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the KMS key.

pub fn set_kms_key_id(self, input: Option<String>) -> Self

If StorageEncrypted is enabled, the Amazon Web Services KMS key identifier for the encrypted DB cluster.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the KMS key.

pub fn db_cluster_resource_id(self, input: impl Into<String>) -> Self

The Amazon Web Services Region-unique, immutable identifier for the DB cluster. This identifier is found in Amazon Web Services CloudTrail log entries whenever the KMS key for the DB cluster is accessed.

pub fn set_db_cluster_resource_id(self, input: Option<String>) -> Self

The Amazon Web Services Region-unique, immutable identifier for the DB cluster. This identifier is found in Amazon Web Services CloudTrail log entries whenever the KMS key for the DB cluster is accessed.

pub fn db_cluster_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) for the DB cluster.

pub fn set_db_cluster_arn(self, input: Option<String>) -> Self

The Amazon Resource Name (ARN) for the DB cluster.

pub fn associated_roles(self, input: impl Into<DbClusterRole>) -> Self

Appends an item to associated_roles.

To override the contents of this collection use set_associated_roles.

Provides a list of the Amazon Web Services Identity and Access Management (IAM) roles that are associated with the DB cluster. IAM roles that are associated with a DB cluster grant permission for the DB cluster to access other Amazon Web Services on your behalf.

pub fn set_associated_roles(self, input: Option<Vec<DbClusterRole>>) -> Self

Provides a list of the Amazon Web Services Identity and Access Management (IAM) roles that are associated with the DB cluster. IAM roles that are associated with a DB cluster grant permission for the DB cluster to access other Amazon Web Services on your behalf.

pub fn iam_database_authentication_enabled(self, input: bool) -> Self

A value that indicates whether the mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts is enabled.

pub fn set_iam_database_authentication_enabled(
self,
input: Option<bool>
) -> Self

A value that indicates whether the mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts is enabled.

pub fn clone_group_id(self, input: impl Into<String>) -> Self

Identifies the clone group to which the DB cluster is associated.

pub fn set_clone_group_id(self, input: Option<String>) -> Self

Identifies the clone group to which the DB cluster is associated.

pub fn cluster_create_time(self, input: Instant) -> Self

Specifies the time when the DB cluster was created, in Universal Coordinated Time (UTC).

pub fn set_cluster_create_time(self, input: Option<Instant>) -> Self

Specifies the time when the DB cluster was created, in Universal Coordinated Time (UTC).

pub fn earliest_backtrack_time(self, input: Instant) -> Self

The earliest time to which a DB cluster can be backtracked.

pub fn set_earliest_backtrack_time(self, input: Option<Instant>) -> Self

The earliest time to which a DB cluster can be backtracked.

pub fn backtrack_window(self, input: i64) -> Self

The target backtrack window, in seconds. If this value is set to 0, backtracking is disabled for the DB cluster. Otherwise, backtracking is enabled.

pub fn set_backtrack_window(self, input: Option<i64>) -> Self

The target backtrack window, in seconds. If this value is set to 0, backtracking is disabled for the DB cluster. Otherwise, backtracking is enabled.

pub fn backtrack_consumed_change_records(self, input: i64) -> Self

The number of change records stored for Backtrack.

pub fn set_backtrack_consumed_change_records(self, input: Option<i64>) -> Self

The number of change records stored for Backtrack.

pub fn enabled_cloudwatch_logs_exports(self, input: impl Into<String>) -> Self

Appends an item to enabled_cloudwatch_logs_exports.

To override the contents of this collection use set_enabled_cloudwatch_logs_exports.

A list of log types that this DB cluster is configured to export to CloudWatch Logs.

Log types vary by DB engine. For information about the log types for each DB engine, see Amazon RDS Database Log Files in the Amazon Aurora User Guide.

pub fn set_enabled_cloudwatch_logs_exports(
self,
input: Option<Vec<String>>
) -> Self

A list of log types that this DB cluster is configured to export to CloudWatch Logs.

Log types vary by DB engine. For information about the log types for each DB engine, see Amazon RDS Database Log Files in the Amazon Aurora User Guide.

pub fn capacity(self, input: i32) -> Self

The current capacity of an Aurora Serverless DB cluster. The capacity is 0 (zero) when the cluster is paused.

For more information about Aurora Serverless, see Using Amazon Aurora Serverless in the Amazon Aurora User Guide.

pub fn set_capacity(self, input: Option<i32>) -> Self

The current capacity of an Aurora Serverless DB cluster. The capacity is 0 (zero) when the cluster is paused.

For more information about Aurora Serverless, see Using Amazon Aurora Serverless in the Amazon Aurora User Guide.

pub fn engine_mode(self, input: impl Into<String>) -> Self

The DB engine mode of the DB cluster, either provisioned, serverless, parallelquery, global, or multimaster.

For more information, see CreateDBCluster.

pub fn set_engine_mode(self, input: Option<String>) -> Self

The DB engine mode of the DB cluster, either provisioned, serverless, parallelquery, global, or multimaster.

For more information, see CreateDBCluster.

pub fn scaling_configuration_info(self, input: ScalingConfigurationInfo) -> Self

Shows the scaling configuration for an Aurora DB cluster in serverless DB engine mode.

For more information, see Using Amazon Aurora Serverless in the Amazon Aurora User Guide.

pub fn set_scaling_configuration_info(
self,
input: Option<ScalingConfigurationInfo>
) -> Self

Shows the scaling configuration for an Aurora DB cluster in serverless DB engine mode.

For more information, see Using Amazon Aurora Serverless in the Amazon Aurora User Guide.

pub fn deletion_protection(self, input: bool) -> Self

Indicates if the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled.

pub fn set_deletion_protection(self, input: Option<bool>) -> Self

Indicates if the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled.

pub fn http_endpoint_enabled(self, input: bool) -> Self

A value that indicates whether the HTTP endpoint for an Aurora Serverless DB cluster is enabled.

When enabled, the HTTP endpoint provides a connectionless web service API for running SQL queries on the Aurora Serverless DB cluster. You can also query your database from inside the RDS console with the query editor.

For more information, see Using the Data API for Aurora Serverless in the Amazon Aurora User Guide.

pub fn set_http_endpoint_enabled(self, input: Option<bool>) -> Self

A value that indicates whether the HTTP endpoint for an Aurora Serverless DB cluster is enabled.

When enabled, the HTTP endpoint provides a connectionless web service API for running SQL queries on the Aurora Serverless DB cluster. You can also query your database from inside the RDS console with the query editor.

For more information, see Using the Data API for Aurora Serverless in the Amazon Aurora User Guide.

pub fn activity_stream_mode(self, input: ActivityStreamMode) -> Self

The mode of the database activity stream. Database events such as a change or access generate an activity stream event. The database session can handle these events either synchronously or asynchronously.

pub fn set_activity_stream_mode(self, input: Option<ActivityStreamMode>) -> Self

The mode of the database activity stream. Database events such as a change or access generate an activity stream event. The database session can handle these events either synchronously or asynchronously.

pub fn activity_stream_status(self, input: ActivityStreamStatus) -> Self

The status of the database activity stream.

pub fn set_activity_stream_status(
self,
input: Option<ActivityStreamStatus>
) -> Self

The status of the database activity stream.

pub fn activity_stream_kms_key_id(self, input: impl Into<String>) -> Self

The Amazon Web Services KMS key identifier used for encrypting messages in the database activity stream.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the KMS key.

pub fn set_activity_stream_kms_key_id(self, input: Option<String>) -> Self

The Amazon Web Services KMS key identifier used for encrypting messages in the database activity stream.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the KMS key.

pub fn activity_stream_kinesis_stream_name(
self,
input: impl Into<String>
) -> Self

The name of the Amazon Kinesis data stream used for the database activity stream.

pub fn set_activity_stream_kinesis_stream_name(
self,
input: Option<String>
) -> Self

The name of the Amazon Kinesis data stream used for the database activity stream.

pub fn copy_tags_to_snapshot(self, input: bool) -> Self

Specifies whether tags are copied from the DB cluster to snapshots of the DB cluster.

pub fn set_copy_tags_to_snapshot(self, input: Option<bool>) -> Self

Specifies whether tags are copied from the DB cluster to snapshots of the DB cluster.

pub fn cross_account_clone(self, input: bool) -> Self

Specifies whether the DB cluster is a clone of a DB cluster owned by a different Amazon Web Services account.

pub fn set_cross_account_clone(self, input: Option<bool>) -> Self

Specifies whether the DB cluster is a clone of a DB cluster owned by a different Amazon Web Services account.

pub fn domain_memberships(self, input: impl Into<DomainMembership>) -> Self

Appends an item to domain_memberships.

To override the contents of this collection use set_domain_memberships.

The Active Directory Domain membership records associated with the DB cluster.

pub fn set_domain_memberships(
self,
input: Option<Vec<DomainMembership>>
) -> Self

The Active Directory Domain membership records associated with the DB cluster.

pub fn tag_list(self, input: impl Into<Tag>) -> Self

Appends an item to tag_list.

To override the contents of this collection use set_tag_list.

A list of tags. For more information, see Tagging Amazon RDS Resources in the Amazon RDS User Guide.

pub fn set_tag_list(self, input: Option<Vec<Tag>>) -> Self

A list of tags. For more information, see Tagging Amazon RDS Resources in the Amazon RDS User Guide.

pub fn global_write_forwarding_status(
self,
input: WriteForwardingStatus
) -> Self

Specifies whether a secondary cluster in an Aurora global database has write forwarding enabled, not enabled, or is in the process of enabling it.

pub fn set_global_write_forwarding_status(
self,
input: Option<WriteForwardingStatus>
) -> Self

Specifies whether a secondary cluster in an Aurora global database has write forwarding enabled, not enabled, or is in the process of enabling it.

pub fn global_write_forwarding_requested(self, input: bool) -> Self

Specifies whether you have requested to enable write forwarding for a secondary cluster in an Aurora global database. Because write forwarding takes time to enable, check the value of GlobalWriteForwardingStatus to confirm that the request has completed before using the write forwarding feature for this cluster.

pub fn set_global_write_forwarding_requested(self, input: Option<bool>) -> Self

Specifies whether you have requested to enable write forwarding for a secondary cluster in an Aurora global database. Because write forwarding takes time to enable, check the value of GlobalWriteForwardingStatus to confirm that the request has completed before using the write forwarding feature for this cluster.

pub fn pending_modified_values(
self,
input: ClusterPendingModifiedValues
) -> Self

A value that specifies that changes to the DB cluster are pending. This element is only included when changes are pending. Specific changes are identified by subelements.

pub fn set_pending_modified_values(
self,
input: Option<ClusterPendingModifiedValues>
) -> Self

A value that specifies that changes to the DB cluster are pending. This element is only included when changes are pending. Specific changes are identified by subelements.

pub fn build(self) -> DbCluster

Consumes the builder and constructs a DbCluster