Struct aws_sdk_rds::model::db_instance::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for DbInstance

Implementations

impl Builder

pub fn db_instance_identifier(self, input: impl Into<String>) -> Self

Contains a user-supplied database identifier. This identifier is the unique key that identifies a DB instance.

pub fn set_db_instance_identifier(self, input: Option<String>) -> Self

Contains a user-supplied database identifier. This identifier is the unique key that identifies a DB instance.

pub fn db_instance_class(self, input: impl Into<String>) -> Self

Contains the name of the compute and memory capacity class of the DB instance.

pub fn set_db_instance_class(self, input: Option<String>) -> Self

Contains the name of the compute and memory capacity class of the DB instance.

pub fn engine(self, input: impl Into<String>) -> Self

The name of the database engine to be used for this DB instance.

pub fn set_engine(self, input: Option<String>) -> Self

The name of the database engine to be used for this DB instance.

pub fn db_instance_status(self, input: impl Into<String>) -> Self

Specifies the current state of this database.

For information about DB instance statuses, see Viewing DB instance status in the Amazon RDS User Guide.

pub fn set_db_instance_status(self, input: Option<String>) -> Self

Specifies the current state of this database.

For information about DB instance statuses, see Viewing DB instance status in the Amazon RDS User Guide.

pub fn automatic_restart_time(self, input: Instant) -> Self

The time when a stopped DB instance is restarted automatically.

pub fn set_automatic_restart_time(self, input: Option<Instant>) -> Self

The time when a stopped DB instance is restarted automatically.

pub fn master_username(self, input: impl Into<String>) -> Self

Contains the master username for the DB instance.

pub fn set_master_username(self, input: Option<String>) -> Self

Contains the master username for the DB instance.

pub fn db_name(self, input: impl Into<String>) -> Self

The meaning of this parameter differs according to the database engine you use.

MySQL, MariaDB, SQL Server, PostgreSQL

Contains the name of the initial database of this instance that was provided at create time, if one was specified when the DB instance was created. This same name is returned for the life of the DB instance.

Type: String

Oracle

Contains the Oracle System ID (SID) of the created DB instance. Not shown when the returned parameters do not apply to an Oracle DB instance.

pub fn set_db_name(self, input: Option<String>) -> Self

The meaning of this parameter differs according to the database engine you use.

MySQL, MariaDB, SQL Server, PostgreSQL

Contains the name of the initial database of this instance that was provided at create time, if one was specified when the DB instance was created. This same name is returned for the life of the DB instance.

Type: String

Oracle

Contains the Oracle System ID (SID) of the created DB instance. Not shown when the returned parameters do not apply to an Oracle DB instance.

pub fn endpoint(self, input: Endpoint) -> Self

Specifies the connection endpoint.

The endpoint might not be shown for instances whose status is creating.

pub fn set_endpoint(self, input: Option<Endpoint>) -> Self

Specifies the connection endpoint.

The endpoint might not be shown for instances whose status is creating.

pub fn allocated_storage(self, input: i32) -> Self

Specifies the allocated storage size specified in gibibytes (GiB).

pub fn set_allocated_storage(self, input: Option<i32>) -> Self

Specifies the allocated storage size specified in gibibytes (GiB).

pub fn instance_create_time(self, input: Instant) -> Self

Provides the date and time the DB instance was created.

pub fn set_instance_create_time(self, input: Option<Instant>) -> Self

Provides the date and time the DB instance was created.

pub fn preferred_backup_window(self, input: impl Into<String>) -> Self

Specifies the daily time range during which automated backups are created if automated backups are enabled, as determined by the BackupRetentionPeriod.

pub fn set_preferred_backup_window(self, input: Option<String>) -> Self

Specifies the daily time range during which automated backups are created if automated backups are enabled, as determined by the BackupRetentionPeriod.

pub fn backup_retention_period(self, input: i32) -> Self

Specifies the number of days for which automatic DB snapshots are retained.

pub fn set_backup_retention_period(self, input: Option<i32>) -> Self

Specifies the number of days for which automatic DB snapshots are retained.

pub fn db_security_groups(
self,
input: impl Into<DbSecurityGroupMembership>
) -> Self

Appends an item to db_security_groups.

To override the contents of this collection use set_db_security_groups.

A list of DB security group elements containing DBSecurityGroup.Name and DBSecurityGroup.Status subelements.

pub fn set_db_security_groups(
self,
input: Option<Vec<DbSecurityGroupMembership>>
) -> Self

A list of DB security group elements containing DBSecurityGroup.Name and DBSecurityGroup.Status subelements.

pub fn vpc_security_groups(
self,
input: impl Into<VpcSecurityGroupMembership>
) -> Self

Appends an item to vpc_security_groups.

To override the contents of this collection use set_vpc_security_groups.

Provides a list of VPC security group elements that the DB instance belongs to.

pub fn set_vpc_security_groups(
self,
input: Option<Vec<VpcSecurityGroupMembership>>
) -> Self

Provides a list of VPC security group elements that the DB instance belongs to.

pub fn db_parameter_groups(
self,
input: impl Into<DbParameterGroupStatus>
) -> Self

Appends an item to db_parameter_groups.

To override the contents of this collection use set_db_parameter_groups.

Provides the list of DB parameter groups applied to this DB instance.

pub fn set_db_parameter_groups(
self,
input: Option<Vec<DbParameterGroupStatus>>
) -> Self

Provides the list of DB parameter groups applied to this DB instance.

pub fn availability_zone(self, input: impl Into<String>) -> Self

Specifies the name of the Availability Zone the DB instance is located in.

pub fn set_availability_zone(self, input: Option<String>) -> Self

Specifies the name of the Availability Zone the DB instance is located in.

pub fn db_subnet_group(self, input: DbSubnetGroup) -> Self

Specifies information on the subnet group associated with the DB instance, including the name, description, and subnets in the subnet group.

pub fn set_db_subnet_group(self, input: Option<DbSubnetGroup>) -> Self

Specifies information on the subnet group associated with the DB instance, including the name, description, and subnets in the subnet group.

pub fn preferred_maintenance_window(self, input: impl Into<String>) -> Self

Specifies the weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

pub fn set_preferred_maintenance_window(self, input: Option<String>) -> Self

Specifies the weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

pub fn pending_modified_values(self, input: PendingModifiedValues) -> Self

A value that specifies that changes to the DB instance are pending. This element is only included when changes are pending. Specific changes are identified by subelements.

pub fn set_pending_modified_values(
self,
input: Option<PendingModifiedValues>
) -> Self

A value that specifies that changes to the DB instance are pending. This element is only included when changes are pending. Specific changes are identified by subelements.

pub fn latest_restorable_time(self, input: Instant) -> Self

Specifies the latest time to which a database can be restored with point-in-time restore.

pub fn set_latest_restorable_time(self, input: Option<Instant>) -> Self

Specifies the latest time to which a database can be restored with point-in-time restore.

pub fn multi_az(self, input: bool) -> Self

Specifies if the DB instance is a Multi-AZ deployment. This setting doesn't apply to RDS Custom.

pub fn set_multi_az(self, input: Option<bool>) -> Self

Specifies if the DB instance is a Multi-AZ deployment. This setting doesn't apply to RDS Custom.

pub fn engine_version(self, input: impl Into<String>) -> Self

Indicates the database engine version.

pub fn set_engine_version(self, input: Option<String>) -> Self

Indicates the database engine version.

pub fn auto_minor_version_upgrade(self, input: bool) -> Self

A value that indicates that minor version patches are applied automatically.

pub fn set_auto_minor_version_upgrade(self, input: Option<bool>) -> Self

A value that indicates that minor version patches are applied automatically.

pub fn read_replica_source_db_instance_identifier(
self,
input: impl Into<String>
) -> Self

Contains the identifier of the source DB instance if this DB instance is a read replica.

pub fn set_read_replica_source_db_instance_identifier(
self,
input: Option<String>
) -> Self

Contains the identifier of the source DB instance if this DB instance is a read replica.

pub fn read_replica_db_instance_identifiers(
self,
input: impl Into<String>
) -> Self

Appends an item to read_replica_db_instance_identifiers.

To override the contents of this collection use set_read_replica_db_instance_identifiers.

Contains one or more identifiers of the read replicas associated with this DB instance.

pub fn set_read_replica_db_instance_identifiers(
self,
input: Option<Vec<String>>
) -> Self

Contains one or more identifiers of the read replicas associated with this DB instance.

pub fn read_replica_db_cluster_identifiers(
self,
input: impl Into<String>
) -> Self

Appends an item to read_replica_db_cluster_identifiers.

To override the contents of this collection use set_read_replica_db_cluster_identifiers.

Contains one or more identifiers of Aurora DB clusters to which the RDS DB instance is replicated as a read replica. For example, when you create an Aurora read replica of an RDS MySQL DB instance, the Aurora MySQL DB cluster for the Aurora read replica is shown. This output does not contain information about cross region Aurora read replicas.

Currently, each RDS DB instance can have only one Aurora read replica.

pub fn set_read_replica_db_cluster_identifiers(
self,
input: Option<Vec<String>>
) -> Self

Contains one or more identifiers of Aurora DB clusters to which the RDS DB instance is replicated as a read replica. For example, when you create an Aurora read replica of an RDS MySQL DB instance, the Aurora MySQL DB cluster for the Aurora read replica is shown. This output does not contain information about cross region Aurora read replicas.

Currently, each RDS DB instance can have only one Aurora read replica.

pub fn replica_mode(self, input: ReplicaMode) -> Self

The open mode of an Oracle read replica. The default is open-read-only. For more information, see Working with Oracle Read Replicas for Amazon RDS in the Amazon RDS User Guide.

This attribute is only supported in RDS for Oracle.

pub fn set_replica_mode(self, input: Option<ReplicaMode>) -> Self

The open mode of an Oracle read replica. The default is open-read-only. For more information, see Working with Oracle Read Replicas for Amazon RDS in the Amazon RDS User Guide.

This attribute is only supported in RDS for Oracle.

pub fn license_model(self, input: impl Into<String>) -> Self

License model information for this DB instance. This setting doesn't apply to RDS Custom.

pub fn set_license_model(self, input: Option<String>) -> Self

License model information for this DB instance. This setting doesn't apply to RDS Custom.

pub fn iops(self, input: i32) -> Self

Specifies the Provisioned IOPS (I/O operations per second) value.

pub fn set_iops(self, input: Option<i32>) -> Self

Specifies the Provisioned IOPS (I/O operations per second) value.

pub fn option_group_memberships(
self,
input: impl Into<OptionGroupMembership>
) -> Self

Appends an item to option_group_memberships.

To override the contents of this collection use set_option_group_memberships.

Provides the list of option group memberships for this DB instance.

pub fn set_option_group_memberships(
self,
input: Option<Vec<OptionGroupMembership>>
) -> Self

Provides the list of option group memberships for this DB instance.

pub fn character_set_name(self, input: impl Into<String>) -> Self

If present, specifies the name of the character set that this instance is associated with.

pub fn set_character_set_name(self, input: Option<String>) -> Self

If present, specifies the name of the character set that this instance is associated with.

pub fn nchar_character_set_name(self, input: impl Into<String>) -> Self

The name of the NCHAR character set for the Oracle DB instance. This character set specifies the Unicode encoding for data stored in table columns of type NCHAR, NCLOB, or NVARCHAR2.

pub fn set_nchar_character_set_name(self, input: Option<String>) -> Self

The name of the NCHAR character set for the Oracle DB instance. This character set specifies the Unicode encoding for data stored in table columns of type NCHAR, NCLOB, or NVARCHAR2.

pub fn secondary_availability_zone(self, input: impl Into<String>) -> Self

If present, specifies the name of the secondary Availability Zone for a DB instance with multi-AZ support.

pub fn set_secondary_availability_zone(self, input: Option<String>) -> Self

If present, specifies the name of the secondary Availability Zone for a DB instance with multi-AZ support.

pub fn publicly_accessible(self, input: bool) -> Self

Specifies the accessibility options for the DB instance.

When the DB instance is publicly accessible, its DNS endpoint resolves to the private IP address from within the DB instance's VPC, and to the public IP address from outside of the DB instance's VPC. Access to the DB instance is ultimately controlled by the security group it uses, and that public access is not permitted if the security group assigned to the DB instance doesn't permit it.

When the DB instance isn't publicly accessible, it is an internal DB instance with a DNS name that resolves to a private IP address.

For more information, see CreateDBInstance.

pub fn set_publicly_accessible(self, input: Option<bool>) -> Self

Specifies the accessibility options for the DB instance.

When the DB instance is publicly accessible, its DNS endpoint resolves to the private IP address from within the DB instance's VPC, and to the public IP address from outside of the DB instance's VPC. Access to the DB instance is ultimately controlled by the security group it uses, and that public access is not permitted if the security group assigned to the DB instance doesn't permit it.

When the DB instance isn't publicly accessible, it is an internal DB instance with a DNS name that resolves to a private IP address.

For more information, see CreateDBInstance.

pub fn status_infos(self, input: impl Into<DbInstanceStatusInfo>) -> Self

Appends an item to status_infos.

To override the contents of this collection use set_status_infos.

The status of a read replica. If the instance isn't a read replica, this is blank.

pub fn set_status_infos(self, input: Option<Vec<DbInstanceStatusInfo>>) -> Self

The status of a read replica. If the instance isn't a read replica, this is blank.

pub fn storage_type(self, input: impl Into<String>) -> Self

Specifies the storage type associated with DB instance.

pub fn set_storage_type(self, input: Option<String>) -> Self

Specifies the storage type associated with DB instance.

pub fn tde_credential_arn(self, input: impl Into<String>) -> Self

The ARN from the key store with which the instance is associated for TDE encryption.

pub fn set_tde_credential_arn(self, input: Option<String>) -> Self

The ARN from the key store with which the instance is associated for TDE encryption.

pub fn db_instance_port(self, input: i32) -> Self

Specifies the port that the DB instance listens on. If the DB instance is part of a DB cluster, this can be a different port than the DB cluster port.

pub fn set_db_instance_port(self, input: Option<i32>) -> Self

Specifies the port that the DB instance listens on. If the DB instance is part of a DB cluster, this can be a different port than the DB cluster port.

pub fn db_cluster_identifier(self, input: impl Into<String>) -> Self

If the DB instance is a member of a DB cluster, contains the name of the DB cluster that the DB instance is a member of.

pub fn set_db_cluster_identifier(self, input: Option<String>) -> Self

If the DB instance is a member of a DB cluster, contains the name of the DB cluster that the DB instance is a member of.

pub fn storage_encrypted(self, input: bool) -> Self

Specifies whether the DB instance is encrypted.

pub fn set_storage_encrypted(self, input: Option<bool>) -> Self

Specifies whether the DB instance is encrypted.

pub fn kms_key_id(self, input: impl Into<String>) -> Self

If StorageEncrypted is true, the Amazon Web Services KMS key identifier for the encrypted DB instance.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the KMS key.

pub fn set_kms_key_id(self, input: Option<String>) -> Self

If StorageEncrypted is true, the Amazon Web Services KMS key identifier for the encrypted DB instance.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the KMS key.

pub fn dbi_resource_id(self, input: impl Into<String>) -> Self

The Amazon Web Services Region-unique, immutable identifier for the DB instance. This identifier is found in Amazon Web Services CloudTrail log entries whenever the Amazon Web Services KMS key for the DB instance is accessed.

pub fn set_dbi_resource_id(self, input: Option<String>) -> Self

The Amazon Web Services Region-unique, immutable identifier for the DB instance. This identifier is found in Amazon Web Services CloudTrail log entries whenever the Amazon Web Services KMS key for the DB instance is accessed.

pub fn ca_certificate_identifier(self, input: impl Into<String>) -> Self

The identifier of the CA certificate for this DB instance.

pub fn set_ca_certificate_identifier(self, input: Option<String>) -> Self

The identifier of the CA certificate for this DB instance.

pub fn domain_memberships(self, input: impl Into<DomainMembership>) -> Self

Appends an item to domain_memberships.

To override the contents of this collection use set_domain_memberships.

The Active Directory Domain membership records associated with the DB instance.

pub fn set_domain_memberships(
self,
input: Option<Vec<DomainMembership>>
) -> Self

The Active Directory Domain membership records associated with the DB instance.

pub fn copy_tags_to_snapshot(self, input: bool) -> Self

Specifies whether tags are copied from the DB instance to snapshots of the DB instance.

Amazon Aurora

Not applicable. Copying tags to snapshots is managed by the DB cluster. Setting this value for an Aurora DB instance has no effect on the DB cluster setting. For more information, see DBCluster.

pub fn set_copy_tags_to_snapshot(self, input: Option<bool>) -> Self

Specifies whether tags are copied from the DB instance to snapshots of the DB instance.

Amazon Aurora

Not applicable. Copying tags to snapshots is managed by the DB cluster. Setting this value for an Aurora DB instance has no effect on the DB cluster setting. For more information, see DBCluster.

pub fn monitoring_interval(self, input: i32) -> Self

The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the DB instance.

pub fn set_monitoring_interval(self, input: Option<i32>) -> Self

The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the DB instance.

pub fn enhanced_monitoring_resource_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the Amazon CloudWatch Logs log stream that receives the Enhanced Monitoring metrics data for the DB instance.

pub fn set_enhanced_monitoring_resource_arn(self, input: Option<String>) -> Self

The Amazon Resource Name (ARN) of the Amazon CloudWatch Logs log stream that receives the Enhanced Monitoring metrics data for the DB instance.

pub fn monitoring_role_arn(self, input: impl Into<String>) -> Self

The ARN for the IAM role that permits RDS to send Enhanced Monitoring metrics to Amazon CloudWatch Logs.

pub fn set_monitoring_role_arn(self, input: Option<String>) -> Self

The ARN for the IAM role that permits RDS to send Enhanced Monitoring metrics to Amazon CloudWatch Logs.

pub fn promotion_tier(self, input: i32) -> Self

A value that specifies the order in which an Aurora Replica is promoted to the primary instance after a failure of the existing primary instance. For more information, see Fault Tolerance for an Aurora DB Cluster in the Amazon Aurora User Guide.

pub fn set_promotion_tier(self, input: Option<i32>) -> Self

A value that specifies the order in which an Aurora Replica is promoted to the primary instance after a failure of the existing primary instance. For more information, see Fault Tolerance for an Aurora DB Cluster in the Amazon Aurora User Guide.

pub fn db_instance_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) for the DB instance.

pub fn set_db_instance_arn(self, input: Option<String>) -> Self

The Amazon Resource Name (ARN) for the DB instance.

pub fn timezone(self, input: impl Into<String>) -> Self

The time zone of the DB instance. In most cases, the Timezone element is empty. Timezone content appears only for Microsoft SQL Server DB instances that were created with a time zone specified.

pub fn set_timezone(self, input: Option<String>) -> Self

The time zone of the DB instance. In most cases, the Timezone element is empty. Timezone content appears only for Microsoft SQL Server DB instances that were created with a time zone specified.

pub fn iam_database_authentication_enabled(self, input: bool) -> Self

True if mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts is enabled, and otherwise false.

IAM database authentication can be enabled for the following database engines

For MySQL 5.6, minor version 5.6.34 or higher

For MySQL 5.7, minor version 5.7.16 or higher

Aurora 5.6 or higher. To enable IAM database authentication for Aurora, see DBCluster Type.

pub fn set_iam_database_authentication_enabled(
self,
input: Option<bool>
) -> Self

True if mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts is enabled, and otherwise false.

IAM database authentication can be enabled for the following database engines

For MySQL 5.6, minor version 5.6.34 or higher

For MySQL 5.7, minor version 5.7.16 or higher

Aurora 5.6 or higher. To enable IAM database authentication for Aurora, see DBCluster Type.

pub fn performance_insights_enabled(self, input: bool) -> Self

True if Performance Insights is enabled for the DB instance, and otherwise false.

pub fn set_performance_insights_enabled(self, input: Option<bool>) -> Self

True if Performance Insights is enabled for the DB instance, and otherwise false.

pub fn performance_insights_kms_key_id(self, input: impl Into<String>) -> Self

The Amazon Web Services KMS key identifier for encryption of Performance Insights data.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the KMS key.

pub fn set_performance_insights_kms_key_id(self, input: Option<String>) -> Self

The Amazon Web Services KMS key identifier for encryption of Performance Insights data.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the KMS key.

pub fn performance_insights_retention_period(self, input: i32) -> Self

The amount of time, in days, to retain Performance Insights data. Valid values are 7 or 731 (2 years).

pub fn set_performance_insights_retention_period(
self,
input: Option<i32>
) -> Self

The amount of time, in days, to retain Performance Insights data. Valid values are 7 or 731 (2 years).

pub fn enabled_cloudwatch_logs_exports(self, input: impl Into<String>) -> Self

Appends an item to enabled_cloudwatch_logs_exports.

To override the contents of this collection use set_enabled_cloudwatch_logs_exports.

A list of log types that this DB instance is configured to export to CloudWatch Logs.

Log types vary by DB engine. For information about the log types for each DB engine, see Amazon RDS Database Log Files in the Amazon RDS User Guide.

pub fn set_enabled_cloudwatch_logs_exports(
self,
input: Option<Vec<String>>
) -> Self

A list of log types that this DB instance is configured to export to CloudWatch Logs.

Log types vary by DB engine. For information about the log types for each DB engine, see Amazon RDS Database Log Files in the Amazon RDS User Guide.

pub fn processor_features(self, input: impl Into<ProcessorFeature>) -> Self

Appends an item to processor_features.

To override the contents of this collection use set_processor_features.

The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

pub fn set_processor_features(
self,
input: Option<Vec<ProcessorFeature>>
) -> Self

The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

pub fn deletion_protection(self, input: bool) -> Self

Indicates if the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. For more information, see Deleting a DB Instance.

pub fn set_deletion_protection(self, input: Option<bool>) -> Self

Indicates if the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. For more information, see Deleting a DB Instance.

pub fn associated_roles(self, input: impl Into<DbInstanceRole>) -> Self

Appends an item to associated_roles.

To override the contents of this collection use set_associated_roles.

The Amazon Web Services Identity and Access Management (IAM) roles associated with the DB instance.

pub fn set_associated_roles(self, input: Option<Vec<DbInstanceRole>>) -> Self

The Amazon Web Services Identity and Access Management (IAM) roles associated with the DB instance.

pub fn listener_endpoint(self, input: Endpoint) -> Self

Specifies the listener connection endpoint for SQL Server Always On.

pub fn set_listener_endpoint(self, input: Option<Endpoint>) -> Self

Specifies the listener connection endpoint for SQL Server Always On.

pub fn max_allocated_storage(self, input: i32) -> Self

The upper limit in gibibytes (GiB) to which Amazon RDS can automatically scale the storage of the DB instance.

pub fn set_max_allocated_storage(self, input: Option<i32>) -> Self

The upper limit in gibibytes (GiB) to which Amazon RDS can automatically scale the storage of the DB instance.

pub fn tag_list(self, input: impl Into<Tag>) -> Self

Appends an item to tag_list.

To override the contents of this collection use set_tag_list.

A list of tags. For more information, see Tagging Amazon RDS Resources in the Amazon RDS User Guide.

pub fn set_tag_list(self, input: Option<Vec<Tag>>) -> Self

A list of tags. For more information, see Tagging Amazon RDS Resources in the Amazon RDS User Guide.

pub fn db_instance_automated_backups_replications(
self,
input: impl Into<DbInstanceAutomatedBackupsReplication>
) -> Self

Appends an item to db_instance_automated_backups_replications.

To override the contents of this collection use set_db_instance_automated_backups_replications.

The list of replicated automated backups associated with the DB instance.

pub fn set_db_instance_automated_backups_replications(
self,
input: Option<Vec<DbInstanceAutomatedBackupsReplication>>
) -> Self

The list of replicated automated backups associated with the DB instance.

pub fn customer_owned_ip_enabled(self, input: bool) -> Self

Specifies whether a customer-owned IP address (CoIP) is enabled for an RDS on Outposts DB instance.

A CoIP provides local or external connectivity to resources in your Outpost subnets through your on-premises network. For some use cases, a CoIP can provide lower latency for connections to the DB instance from outside of its virtual private cloud (VPC) on your local network.

For more information about RDS on Outposts, see Working with Amazon RDS on Amazon Web Services Outposts in the Amazon RDS User Guide.

For more information about CoIPs, see Customer-owned IP addresses in the Amazon Web Services Outposts User Guide.

pub fn set_customer_owned_ip_enabled(self, input: Option<bool>) -> Self

Specifies whether a customer-owned IP address (CoIP) is enabled for an RDS on Outposts DB instance.

A CoIP provides local or external connectivity to resources in your Outpost subnets through your on-premises network. For some use cases, a CoIP can provide lower latency for connections to the DB instance from outside of its virtual private cloud (VPC) on your local network.

For more information about RDS on Outposts, see Working with Amazon RDS on Amazon Web Services Outposts in the Amazon RDS User Guide.

For more information about CoIPs, see Customer-owned IP addresses in the Amazon Web Services Outposts User Guide.

pub fn aws_backup_recovery_point_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the recovery point in Amazon Web Services Backup.

pub fn set_aws_backup_recovery_point_arn(self, input: Option<String>) -> Self

The Amazon Resource Name (ARN) of the recovery point in Amazon Web Services Backup.

pub fn activity_stream_status(self, input: ActivityStreamStatus) -> Self

The status of the database activity stream.

pub fn set_activity_stream_status(
self,
input: Option<ActivityStreamStatus>
) -> Self

The status of the database activity stream.

pub fn activity_stream_kms_key_id(self, input: impl Into<String>) -> Self

The Amazon Web Services KMS key identifier used for encrypting messages in the database activity stream. The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the KMS key.

pub fn set_activity_stream_kms_key_id(self, input: Option<String>) -> Self

The Amazon Web Services KMS key identifier used for encrypting messages in the database activity stream. The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the KMS key.

pub fn activity_stream_kinesis_stream_name(
self,
input: impl Into<String>
) -> Self

The name of the Amazon Kinesis data stream used for the database activity stream.

pub fn set_activity_stream_kinesis_stream_name(
self,
input: Option<String>
) -> Self

The name of the Amazon Kinesis data stream used for the database activity stream.

pub fn activity_stream_mode(self, input: ActivityStreamMode) -> Self

The mode of the database activity stream. Database events such as a change or access generate an activity stream event. RDS for Oracle always handles these events asynchronously.

pub fn set_activity_stream_mode(self, input: Option<ActivityStreamMode>) -> Self

The mode of the database activity stream. Database events such as a change or access generate an activity stream event. RDS for Oracle always handles these events asynchronously.

pub fn activity_stream_engine_native_audit_fields_included(
self,
input: bool
) -> Self

Indicates whether engine-native audit fields are included in the database activity stream.

pub fn set_activity_stream_engine_native_audit_fields_included(
self,
input: Option<bool>
) -> Self

Indicates whether engine-native audit fields are included in the database activity stream.

pub fn automation_mode(self, input: AutomationMode) -> Self

The automation mode of the RDS Custom DB instance: full or all paused. If full, the DB instance automates monitoring and instance recovery. If all paused, the instance pauses automation for the duration set by --resume-full-automation-mode-minutes.

pub fn set_automation_mode(self, input: Option<AutomationMode>) -> Self

The automation mode of the RDS Custom DB instance: full or all paused. If full, the DB instance automates monitoring and instance recovery. If all paused, the instance pauses automation for the duration set by --resume-full-automation-mode-minutes.

pub fn resume_full_automation_mode_time(self, input: Instant) -> Self

The number of minutes to pause the automation. When the time period ends, RDS Custom resumes full automation. The minimum value is 60 (default). The maximum value is 1,440.

pub fn set_resume_full_automation_mode_time(
self,
input: Option<Instant>
) -> Self

The number of minutes to pause the automation. When the time period ends, RDS Custom resumes full automation. The minimum value is 60 (default). The maximum value is 1,440.

pub fn custom_iam_instance_profile(self, input: impl Into<String>) -> Self

The instance profile associated with the underlying Amazon EC2 instance of an RDS Custom DB instance. The instance profile must meet the following requirements:

The profile must exist in your account.

The profile must have an IAM role that Amazon EC2 has permissions to assume.

The instance profile name and the associated IAM role name must start with the prefix AWSRDSCustom.

For the list of permissions required for the IAM role, see Configure IAM and your VPC in the Amazon Relational Database Service User Guide.

pub fn set_custom_iam_instance_profile(self, input: Option<String>) -> Self

The instance profile associated with the underlying Amazon EC2 instance of an RDS Custom DB instance. The instance profile must meet the following requirements:

The profile must exist in your account.

The profile must have an IAM role that Amazon EC2 has permissions to assume.

The instance profile name and the associated IAM role name must start with the prefix AWSRDSCustom.

For the list of permissions required for the IAM role, see Configure IAM and your VPC in the Amazon Relational Database Service User Guide.

pub fn build(self) -> DbInstance

Consumes the builder and constructs a DbInstance