Struct aws_sdk_rds::model::db_instance::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for DBInstance

Implementations

impl Builder

pub fn db_instance_identifier(self, input: impl Into<String>) -> Self

Contains a user-supplied database identifier. This identifier is the unique key that identifies a DB instance.

pub fn set_db_instance_identifier(self, input: Option<String>) -> Self

pub fn db_instance_class(self, input: impl Into<String>) -> Self

Contains the name of the compute and memory capacity class of the DB instance.

pub fn set_db_instance_class(self, input: Option<String>) -> Self

pub fn engine(self, input: impl Into<String>) -> Self

The name of the database engine to be used for this DB instance.

pub fn set_engine(self, input: Option<String>) -> Self

pub fn db_instance_status(self, input: impl Into<String>) -> Self

Specifies the current state of this database.

For information about DB instance statuses, see Viewing DB instance status in the Amazon RDS User Guide.

pub fn set_db_instance_status(self, input: Option<String>) -> Self

pub fn master_username(self, input: impl Into<String>) -> Self

Contains the master username for the DB instance.

pub fn set_master_username(self, input: Option<String>) -> Self

pub fn db_name(self, input: impl Into<String>) -> Self

The meaning of this parameter differs according to the database engine you use.

MySQL, MariaDB, SQL Server, PostgreSQL

Contains the name of the initial database of this instance that was provided at create time, if one was specified when the DB instance was created. This same name is returned for the life of the DB instance.

Type: String

Oracle

Contains the Oracle System ID (SID) of the created DB instance. Not shown when the returned parameters do not apply to an Oracle DB instance.

pub fn set_db_name(self, input: Option<String>) -> Self

pub fn endpoint(self, input: Endpoint) -> Self

Specifies the connection endpoint.

pub fn set_endpoint(self, input: Option<Endpoint>) -> Self

pub fn allocated_storage(self, input: i32) -> Self

Specifies the allocated storage size specified in gibibytes.

pub fn set_allocated_storage(self, input: Option<i32>) -> Self

pub fn instance_create_time(self, input: Instant) -> Self

Provides the date and time the DB instance was created.

pub fn set_instance_create_time(self, input: Option<Instant>) -> Self

pub fn preferred_backup_window(self, input: impl Into<String>) -> Self

Specifies the daily time range during which automated backups are created if automated backups are enabled, as determined by the BackupRetentionPeriod.

pub fn set_preferred_backup_window(self, input: Option<String>) -> Self

pub fn backup_retention_period(self, input: i32) -> Self

Specifies the number of days for which automatic DB snapshots are retained.

pub fn set_backup_retention_period(self, input: Option<i32>) -> Self

pub fn db_security_groups(
self,
input: impl Into<DBSecurityGroupMembership>
) -> Self

pub fn set_db_security_groups(
self,
input: Option<Vec<DBSecurityGroupMembership>>
) -> Self

pub fn vpc_security_groups(
self,
input: impl Into<VpcSecurityGroupMembership>
) -> Self

pub fn set_vpc_security_groups(
self,
input: Option<Vec<VpcSecurityGroupMembership>>
) -> Self

pub fn db_parameter_groups(
self,
input: impl Into<DBParameterGroupStatus>
) -> Self

pub fn set_db_parameter_groups(
self,
input: Option<Vec<DBParameterGroupStatus>>
) -> Self

pub fn availability_zone(self, input: impl Into<String>) -> Self

Specifies the name of the Availability Zone the DB instance is located in.

pub fn set_availability_zone(self, input: Option<String>) -> Self

pub fn db_subnet_group(self, input: DBSubnetGroup) -> Self

Specifies information on the subnet group associated with the DB instance, including the name, description, and subnets in the subnet group.

pub fn set_db_subnet_group(self, input: Option<DBSubnetGroup>) -> Self

pub fn preferred_maintenance_window(self, input: impl Into<String>) -> Self

Specifies the weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

pub fn set_preferred_maintenance_window(self, input: Option<String>) -> Self

pub fn pending_modified_values(self, input: PendingModifiedValues) -> Self

A value that specifies that changes to the DB instance are pending. This element is only included when changes are pending. Specific changes are identified by subelements.

pub fn set_pending_modified_values(
self,
input: Option<PendingModifiedValues>
) -> Self

pub fn latest_restorable_time(self, input: Instant) -> Self

Specifies the latest time to which a database can be restored with point-in-time restore.

pub fn set_latest_restorable_time(self, input: Option<Instant>) -> Self

pub fn multi_az(self, input: bool) -> Self

Specifies if the DB instance is a Multi-AZ deployment.

pub fn set_multi_az(self, input: Option<bool>) -> Self

pub fn engine_version(self, input: impl Into<String>) -> Self

Indicates the database engine version.

pub fn set_engine_version(self, input: Option<String>) -> Self

pub fn auto_minor_version_upgrade(self, input: bool) -> Self

A value that indicates that minor version patches are applied automatically.

pub fn set_auto_minor_version_upgrade(self, input: Option<bool>) -> Self

pub fn read_replica_source_db_instance_identifier(
self,
input: impl Into<String>
) -> Self

Contains the identifier of the source DB instance if this DB instance is a read replica.

pub fn set_read_replica_source_db_instance_identifier(
self,
input: Option<String>
) -> Self

pub fn read_replica_db_instance_identifiers(
self,
input: impl Into<String>
) -> Self

pub fn set_read_replica_db_instance_identifiers(
self,
input: Option<Vec<String>>
) -> Self

pub fn read_replica_db_cluster_identifiers(
self,
input: impl Into<String>
) -> Self

pub fn set_read_replica_db_cluster_identifiers(
self,
input: Option<Vec<String>>
) -> Self

pub fn replica_mode(self, input: ReplicaMode) -> Self

The open mode of an Oracle read replica. The default is open-read-only. For more information, see Working with Oracle Read Replicas for Amazon RDS in the Amazon RDS User Guide.

This attribute is only supported in RDS for Oracle.

pub fn set_replica_mode(self, input: Option<ReplicaMode>) -> Self

pub fn license_model(self, input: impl Into<String>) -> Self

License model information for this DB instance.

pub fn set_license_model(self, input: Option<String>) -> Self

pub fn iops(self, input: i32) -> Self

Specifies the Provisioned IOPS (I/O operations per second) value.

pub fn set_iops(self, input: Option<i32>) -> Self

pub fn option_group_memberships(
self,
input: impl Into<OptionGroupMembership>
) -> Self

pub fn set_option_group_memberships(
self,
input: Option<Vec<OptionGroupMembership>>
) -> Self

pub fn character_set_name(self, input: impl Into<String>) -> Self

If present, specifies the name of the character set that this instance is associated with.

pub fn set_character_set_name(self, input: Option<String>) -> Self

pub fn nchar_character_set_name(self, input: impl Into<String>) -> Self

The name of the NCHAR character set for the Oracle DB instance. This character set specifies the Unicode encoding for data stored in table columns of type NCHAR, NCLOB, or NVARCHAR2.

pub fn set_nchar_character_set_name(self, input: Option<String>) -> Self

pub fn secondary_availability_zone(self, input: impl Into<String>) -> Self

If present, specifies the name of the secondary Availability Zone for a DB instance with multi-AZ support.

pub fn set_secondary_availability_zone(self, input: Option<String>) -> Self

pub fn publicly_accessible(self, input: bool) -> Self

Specifies the accessibility options for the DB instance.

When the DB instance is publicly accessible, its DNS endpoint resolves to the private IP address from within the DB instance's VPC, and to the public IP address from outside of the DB instance's VPC. Access to the DB instance is ultimately controlled by the security group it uses, and that public access is not permitted if the security group assigned to the DB instance doesn't permit it.

When the DB instance isn't publicly accessible, it is an internal DB instance with a DNS name that resolves to a private IP address.

For more information, see CreateDBInstance.

pub fn set_publicly_accessible(self, input: Option<bool>) -> Self

pub fn status_infos(self, input: impl Into<DBInstanceStatusInfo>) -> Self

pub fn set_status_infos(self, input: Option<Vec<DBInstanceStatusInfo>>) -> Self

pub fn storage_type(self, input: impl Into<String>) -> Self

Specifies the storage type associated with DB instance.

pub fn set_storage_type(self, input: Option<String>) -> Self

pub fn tde_credential_arn(self, input: impl Into<String>) -> Self

The ARN from the key store with which the instance is associated for TDE encryption.

pub fn set_tde_credential_arn(self, input: Option<String>) -> Self

pub fn db_instance_port(self, input: i32) -> Self

Specifies the port that the DB instance listens on. If the DB instance is part of a DB cluster, this can be a different port than the DB cluster port.

pub fn set_db_instance_port(self, input: Option<i32>) -> Self

pub fn db_cluster_identifier(self, input: impl Into<String>) -> Self

If the DB instance is a member of a DB cluster, contains the name of the DB cluster that the DB instance is a member of.

pub fn set_db_cluster_identifier(self, input: Option<String>) -> Self

pub fn storage_encrypted(self, input: bool) -> Self

Specifies whether the DB instance is encrypted.

pub fn set_storage_encrypted(self, input: Option<bool>) -> Self

pub fn kms_key_id(self, input: impl Into<String>) -> Self

If StorageEncrypted is true, the AWS KMS key identifier for the encrypted DB instance.

The AWS KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the AWS KMS customer master key (CMK).

pub fn set_kms_key_id(self, input: Option<String>) -> Self

pub fn dbi_resource_id(self, input: impl Into<String>) -> Self

The AWS Region-unique, immutable identifier for the DB instance. This identifier is found in AWS CloudTrail log entries whenever the AWS KMS customer master key (CMK) for the DB instance is accessed.

pub fn set_dbi_resource_id(self, input: Option<String>) -> Self

pub fn ca_certificate_identifier(self, input: impl Into<String>) -> Self

The identifier of the CA certificate for this DB instance.

pub fn set_ca_certificate_identifier(self, input: Option<String>) -> Self

pub fn domain_memberships(self, input: impl Into<DomainMembership>) -> Self

pub fn set_domain_memberships(
self,
input: Option<Vec<DomainMembership>>
) -> Self

pub fn copy_tags_to_snapshot(self, input: bool) -> Self

Specifies whether tags are copied from the DB instance to snapshots of the DB instance.

Amazon Aurora

Not applicable. Copying tags to snapshots is managed by the DB cluster. Setting this value for an Aurora DB instance has no effect on the DB cluster setting. For more information, see DBCluster.

pub fn set_copy_tags_to_snapshot(self, input: Option<bool>) -> Self

pub fn monitoring_interval(self, input: i32) -> Self

The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the DB instance.

pub fn set_monitoring_interval(self, input: Option<i32>) -> Self

pub fn enhanced_monitoring_resource_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the Amazon CloudWatch Logs log stream that receives the Enhanced Monitoring metrics data for the DB instance.

pub fn set_enhanced_monitoring_resource_arn(self, input: Option<String>) -> Self

pub fn monitoring_role_arn(self, input: impl Into<String>) -> Self

The ARN for the IAM role that permits RDS to send Enhanced Monitoring metrics to Amazon CloudWatch Logs.

pub fn set_monitoring_role_arn(self, input: Option<String>) -> Self

pub fn promotion_tier(self, input: i32) -> Self

A value that specifies the order in which an Aurora Replica is promoted to the primary instance after a failure of the existing primary instance. For more information, see Fault Tolerance for an Aurora DB Cluster in the Amazon Aurora User Guide.

pub fn set_promotion_tier(self, input: Option<i32>) -> Self

pub fn db_instance_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) for the DB instance.

pub fn set_db_instance_arn(self, input: Option<String>) -> Self

pub fn timezone(self, input: impl Into<String>) -> Self

The time zone of the DB instance. In most cases, the Timezone element is empty. Timezone content appears only for Microsoft SQL Server DB instances that were created with a time zone specified.

pub fn set_timezone(self, input: Option<String>) -> Self

pub fn iam_database_authentication_enabled(self, input: bool) -> Self

True if mapping of AWS Identity and Access Management (IAM) accounts to database accounts is enabled, and otherwise false.

IAM database authentication can be enabled for the following database engines

For MySQL 5.6, minor version 5.6.34 or higher

For MySQL 5.7, minor version 5.7.16 or higher

Aurora 5.6 or higher. To enable IAM database authentication for Aurora, see DBCluster Type.

pub fn set_iam_database_authentication_enabled(
self,
input: Option<bool>
) -> Self

pub fn performance_insights_enabled(self, input: bool) -> Self

True if Performance Insights is enabled for the DB instance, and otherwise false.

pub fn set_performance_insights_enabled(self, input: Option<bool>) -> Self

pub fn performance_insights_kms_key_id(self, input: impl Into<String>) -> Self

The AWS KMS key identifier for encryption of Performance Insights data.

The AWS KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the AWS KMS customer master key (CMK).

pub fn set_performance_insights_kms_key_id(self, input: Option<String>) -> Self

pub fn performance_insights_retention_period(self, input: i32) -> Self

The amount of time, in days, to retain Performance Insights data. Valid values are 7 or 731 (2 years).

pub fn set_performance_insights_retention_period(
self,
input: Option<i32>
) -> Self

pub fn enabled_cloudwatch_logs_exports(self, input: impl Into<String>) -> Self

pub fn set_enabled_cloudwatch_logs_exports(
self,
input: Option<Vec<String>>
) -> Self

pub fn processor_features(self, input: impl Into<ProcessorFeature>) -> Self

pub fn set_processor_features(
self,
input: Option<Vec<ProcessorFeature>>
) -> Self

pub fn deletion_protection(self, input: bool) -> Self

Indicates if the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. For more information, see Deleting a DB Instance.

pub fn set_deletion_protection(self, input: Option<bool>) -> Self

pub fn associated_roles(self, input: impl Into<DBInstanceRole>) -> Self

pub fn set_associated_roles(self, input: Option<Vec<DBInstanceRole>>) -> Self

pub fn listener_endpoint(self, input: Endpoint) -> Self

Specifies the listener connection endpoint for SQL Server Always On.

pub fn set_listener_endpoint(self, input: Option<Endpoint>) -> Self

pub fn max_allocated_storage(self, input: i32) -> Self

The upper limit to which Amazon RDS can automatically scale the storage of the DB instance.

pub fn set_max_allocated_storage(self, input: Option<i32>) -> Self

pub fn tag_list(self, input: impl Into<Tag>) -> Self

pub fn set_tag_list(self, input: Option<Vec<Tag>>) -> Self

pub fn db_instance_automated_backups_replications(
self,
input: impl Into<DBInstanceAutomatedBackupsReplication>
) -> Self

pub fn set_db_instance_automated_backups_replications(
self,
input: Option<Vec<DBInstanceAutomatedBackupsReplication>>
) -> Self

pub fn customer_owned_ip_enabled(self, input: bool) -> Self

Specifies whether a customer-owned IP address (CoIP) is enabled for an RDS on Outposts DB instance.

A CoIP provides local or external connectivity to resources in your Outpost subnets through your on-premises network. For some use cases, a CoIP can provide lower latency for connections to the DB instance from outside of its virtual private cloud (VPC) on your local network.

For more information about RDS on Outposts, see Working with Amazon RDS on AWS Outposts in the Amazon RDS User Guide.

For more information about CoIPs, see Customer-owned IP addresses in the AWS Outposts User Guide.

pub fn set_customer_owned_ip_enabled(self, input: Option<bool>) -> Self

pub fn aws_backup_recovery_point_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the recovery point in AWS Backup.

pub fn set_aws_backup_recovery_point_arn(self, input: Option<String>) -> Self

pub fn build(self) -> DBInstance

Consumes the builder and constructs a DBInstance