Struct aws_sdk_rds::model::DbCluster

#[non_exhaustive]
pub struct DbCluster {

    pub allocated_storage: Option<i32>,
    pub availability_zones: Option<Vec<String>>,
    pub backup_retention_period: Option<i32>,
    pub character_set_name: Option<String>,
    pub database_name: Option<String>,
    pub db_cluster_identifier: Option<String>,
    pub db_cluster_parameter_group: Option<String>,
    pub db_subnet_group: Option<String>,
    pub status: Option<String>,
    pub automatic_restart_time: Option<Instant>,
    pub percent_progress: Option<String>,
    pub earliest_restorable_time: Option<Instant>,
    pub endpoint: Option<String>,
    pub reader_endpoint: Option<String>,
    pub custom_endpoints: Option<Vec<String>>,
    pub multi_az: Option<bool>,
    pub engine: Option<String>,
    pub engine_version: Option<String>,
    pub latest_restorable_time: Option<Instant>,
    pub port: Option<i32>,
    pub master_username: Option<String>,
    pub db_cluster_option_group_memberships: Option<Vec<DbClusterOptionGroupStatus>>,
    pub preferred_backup_window: Option<String>,
    pub preferred_maintenance_window: Option<String>,
    pub replication_source_identifier: Option<String>,
    pub read_replica_identifiers: Option<Vec<String>>,
    pub db_cluster_members: Option<Vec<DbClusterMember>>,
    pub vpc_security_groups: Option<Vec<VpcSecurityGroupMembership>>,
    pub hosted_zone_id: Option<String>,
    pub storage_encrypted: bool,
    pub kms_key_id: Option<String>,
    pub db_cluster_resource_id: Option<String>,
    pub db_cluster_arn: Option<String>,
    pub associated_roles: Option<Vec<DbClusterRole>>,
    pub iam_database_authentication_enabled: Option<bool>,
    pub clone_group_id: Option<String>,
    pub cluster_create_time: Option<Instant>,
    pub earliest_backtrack_time: Option<Instant>,
    pub backtrack_window: Option<i64>,
    pub backtrack_consumed_change_records: Option<i64>,
    pub enabled_cloudwatch_logs_exports: Option<Vec<String>>,
    pub capacity: Option<i32>,
    pub engine_mode: Option<String>,
    pub scaling_configuration_info: Option<ScalingConfigurationInfo>,
    pub deletion_protection: Option<bool>,
    pub http_endpoint_enabled: Option<bool>,
    pub activity_stream_mode: Option<ActivityStreamMode>,
    pub activity_stream_status: Option<ActivityStreamStatus>,
    pub activity_stream_kms_key_id: Option<String>,
    pub activity_stream_kinesis_stream_name: Option<String>,
    pub copy_tags_to_snapshot: Option<bool>,
    pub cross_account_clone: Option<bool>,
    pub domain_memberships: Option<Vec<DomainMembership>>,
    pub tag_list: Option<Vec<Tag>>,
    pub global_write_forwarding_status: Option<WriteForwardingStatus>,
    pub global_write_forwarding_requested: Option<bool>,
    pub pending_modified_values: Option<ClusterPendingModifiedValues>,

}

Contains the details of an Amazon Aurora DB cluster.

This data type is used as a response element in the DescribeDBClusters, StopDBCluster, and StartDBCluster actions.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

allocated_storage: Option<i32>

For all database engines except Amazon Aurora, AllocatedStorage specifies the allocated storage size in gibibytes (GiB). For Aurora, AllocatedStorage always returns 1, because Aurora DB cluster storage size isn't fixed, but instead automatically adjusts as needed.

availability_zones: Option<Vec<String>>

Provides the list of Availability Zones (AZs) where instances in the DB cluster can be created.

backup_retention_period: Option<i32>

Specifies the number of days for which automatic DB snapshots are retained.

character_set_name: Option<String>

If present, specifies the name of the character set that this cluster is associated with.

database_name: Option<String>

Contains the name of the initial database of this DB cluster that was provided at create time, if one was specified when the DB cluster was created. This same name is returned for the life of the DB cluster.

db_cluster_identifier: Option<String>

Contains a user-supplied DB cluster identifier. This identifier is the unique key that identifies a DB cluster.

db_cluster_parameter_group: Option<String>

Specifies the name of the DB cluster parameter group for the DB cluster.

db_subnet_group: Option<String>

Specifies information on the subnet group associated with the DB cluster, including the name, description, and subnets in the subnet group.

status: Option<String>

Specifies the current state of this DB cluster.

automatic_restart_time: Option<Instant>

The time when a stopped DB cluster is restarted automatically.

percent_progress: Option<String>

Specifies the progress of the operation as a percentage.

earliest_restorable_time: Option<Instant>

The earliest time to which a database can be restored with point-in-time restore.

endpoint: Option<String>

Specifies the connection endpoint for the primary instance of the DB cluster.

reader_endpoint: Option<String>

The reader endpoint for the DB cluster. The reader endpoint for a DB cluster load-balances connections across the Aurora Replicas that are available in a DB cluster. As clients request new connections to the reader endpoint, Aurora distributes the connection requests among the Aurora Replicas in the DB cluster. This functionality can help balance your read workload across multiple Aurora Replicas in your DB cluster.

If a failover occurs, and the Aurora Replica that you are connected to is promoted to be the primary instance, your connection is dropped. To continue sending your read workload to other Aurora Replicas in the cluster, you can then reconnect to the reader endpoint.

custom_endpoints: Option<Vec<String>>

Identifies all custom endpoints associated with the cluster.

multi_az: Option<bool>

Specifies whether the DB cluster has instances in multiple Availability Zones.

engine: Option<String>

The name of the database engine to be used for this DB cluster.

engine_version: Option<String>

Indicates the database engine version.

latest_restorable_time: Option<Instant>

Specifies the latest time to which a database can be restored with point-in-time restore.

port: Option<i32>

Specifies the port that the database engine is listening on.

master_username: Option<String>

Contains the master username for the DB cluster.

db_cluster_option_group_memberships: Option<Vec<DbClusterOptionGroupStatus>>

Provides the list of option group memberships for this DB cluster.

preferred_backup_window: Option<String>

Specifies the daily time range during which automated backups are created if automated backups are enabled, as determined by the BackupRetentionPeriod.

preferred_maintenance_window: Option<String>

Specifies the weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

replication_source_identifier: Option<String>

Contains the identifier of the source DB cluster if this DB cluster is a read replica.

read_replica_identifiers: Option<Vec<String>>

Contains one or more identifiers of the read replicas associated with this DB cluster.

db_cluster_members: Option<Vec<DbClusterMember>>

Provides the list of instances that make up the DB cluster.

vpc_security_groups: Option<Vec<VpcSecurityGroupMembership>>

Provides a list of VPC security groups that the DB cluster belongs to.

hosted_zone_id: Option<String>

Specifies the ID that Amazon Route 53 assigns when you create a hosted zone.

storage_encrypted: bool

Specifies whether the DB cluster is encrypted.

kms_key_id: Option<String>

If StorageEncrypted is enabled, the Amazon Web Services KMS key identifier for the encrypted DB cluster.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the KMS key.

db_cluster_resource_id: Option<String>

The Amazon Web Services Region-unique, immutable identifier for the DB cluster. This identifier is found in Amazon Web Services CloudTrail log entries whenever the KMS key for the DB cluster is accessed.

db_cluster_arn: Option<String>

The Amazon Resource Name (ARN) for the DB cluster.

associated_roles: Option<Vec<DbClusterRole>>

Provides a list of the Amazon Web Services Identity and Access Management (IAM) roles that are associated with the DB cluster. IAM roles that are associated with a DB cluster grant permission for the DB cluster to access other Amazon Web Services on your behalf.

iam_database_authentication_enabled: Option<bool>

A value that indicates whether the mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts is enabled.

clone_group_id: Option<String>

Identifies the clone group to which the DB cluster is associated.

cluster_create_time: Option<Instant>

Specifies the time when the DB cluster was created, in Universal Coordinated Time (UTC).

earliest_backtrack_time: Option<Instant>

The earliest time to which a DB cluster can be backtracked.

backtrack_window: Option<i64>

The target backtrack window, in seconds. If this value is set to 0, backtracking is disabled for the DB cluster. Otherwise, backtracking is enabled.

backtrack_consumed_change_records: Option<i64>

The number of change records stored for Backtrack.

enabled_cloudwatch_logs_exports: Option<Vec<String>>

A list of log types that this DB cluster is configured to export to CloudWatch Logs.

Log types vary by DB engine. For information about the log types for each DB engine, see Amazon RDS Database Log Files in the Amazon Aurora User Guide.

capacity: Option<i32>

The current capacity of an Aurora Serverless DB cluster. The capacity is 0 (zero) when the cluster is paused.

For more information about Aurora Serverless, see Using Amazon Aurora Serverless in the Amazon Aurora User Guide.

engine_mode: Option<String>

The DB engine mode of the DB cluster, either provisioned, serverless, parallelquery, global, or multimaster.

For more information, see CreateDBCluster.

scaling_configuration_info: Option<ScalingConfigurationInfo>

Shows the scaling configuration for an Aurora DB cluster in serverless DB engine mode.

For more information, see Using Amazon Aurora Serverless in the Amazon Aurora User Guide.

deletion_protection: Option<bool>

Indicates if the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled.

http_endpoint_enabled: Option<bool>

A value that indicates whether the HTTP endpoint for an Aurora Serverless DB cluster is enabled.

When enabled, the HTTP endpoint provides a connectionless web service API for running SQL queries on the Aurora Serverless DB cluster. You can also query your database from inside the RDS console with the query editor.

For more information, see Using the Data API for Aurora Serverless in the Amazon Aurora User Guide.

activity_stream_mode: Option<ActivityStreamMode>

The mode of the database activity stream. Database events such as a change or access generate an activity stream event. The database session can handle these events either synchronously or asynchronously.

activity_stream_status: Option<ActivityStreamStatus>

The status of the database activity stream.

activity_stream_kms_key_id: Option<String>

The Amazon Web Services KMS key identifier used for encrypting messages in the database activity stream.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the KMS key.

activity_stream_kinesis_stream_name: Option<String>

The name of the Amazon Kinesis data stream used for the database activity stream.

copy_tags_to_snapshot: Option<bool>

Specifies whether tags are copied from the DB cluster to snapshots of the DB cluster.

cross_account_clone: Option<bool>

Specifies whether the DB cluster is a clone of a DB cluster owned by a different Amazon Web Services account.

domain_memberships: Option<Vec<DomainMembership>>

The Active Directory Domain membership records associated with the DB cluster.

tag_list: Option<Vec<Tag>>

A list of tags. For more information, see Tagging Amazon RDS Resources in the Amazon RDS User Guide.

global_write_forwarding_status: Option<WriteForwardingStatus>

Specifies whether a secondary cluster in an Aurora global database has write forwarding enabled, not enabled, or is in the process of enabling it.

global_write_forwarding_requested: Option<bool>

Specifies whether you have requested to enable write forwarding for a secondary cluster in an Aurora global database. Because write forwarding takes time to enable, check the value of GlobalWriteForwardingStatus to confirm that the request has completed before using the write forwarding feature for this cluster.

pending_modified_values: Option<ClusterPendingModifiedValues>

A value that specifies that changes to the DB cluster are pending. This element is only included when changes are pending. Specific changes are identified by subelements.

Implementations

impl DbCluster

pub fn allocated_storage(&self) -> Option<i32>

For all database engines except Amazon Aurora, AllocatedStorage specifies the allocated storage size in gibibytes (GiB). For Aurora, AllocatedStorage always returns 1, because Aurora DB cluster storage size isn't fixed, but instead automatically adjusts as needed.

pub fn availability_zones(&self) -> Option<&[String]>

Provides the list of Availability Zones (AZs) where instances in the DB cluster can be created.

pub fn backup_retention_period(&self) -> Option<i32>

Specifies the number of days for which automatic DB snapshots are retained.

pub fn character_set_name(&self) -> Option<&str>

If present, specifies the name of the character set that this cluster is associated with.

pub fn database_name(&self) -> Option<&str>

Contains the name of the initial database of this DB cluster that was provided at create time, if one was specified when the DB cluster was created. This same name is returned for the life of the DB cluster.

pub fn db_cluster_identifier(&self) -> Option<&str>

Contains a user-supplied DB cluster identifier. This identifier is the unique key that identifies a DB cluster.

pub fn db_cluster_parameter_group(&self) -> Option<&str>

Specifies the name of the DB cluster parameter group for the DB cluster.

pub fn db_subnet_group(&self) -> Option<&str>

Specifies information on the subnet group associated with the DB cluster, including the name, description, and subnets in the subnet group.

pub fn status(&self) -> Option<&str>

Specifies the current state of this DB cluster.

pub fn automatic_restart_time(&self) -> Option<&Instant>

The time when a stopped DB cluster is restarted automatically.

pub fn percent_progress(&self) -> Option<&str>

Specifies the progress of the operation as a percentage.

pub fn earliest_restorable_time(&self) -> Option<&Instant>

The earliest time to which a database can be restored with point-in-time restore.

pub fn endpoint(&self) -> Option<&str>

Specifies the connection endpoint for the primary instance of the DB cluster.

pub fn reader_endpoint(&self) -> Option<&str>

The reader endpoint for the DB cluster. The reader endpoint for a DB cluster load-balances connections across the Aurora Replicas that are available in a DB cluster. As clients request new connections to the reader endpoint, Aurora distributes the connection requests among the Aurora Replicas in the DB cluster. This functionality can help balance your read workload across multiple Aurora Replicas in your DB cluster.

If a failover occurs, and the Aurora Replica that you are connected to is promoted to be the primary instance, your connection is dropped. To continue sending your read workload to other Aurora Replicas in the cluster, you can then reconnect to the reader endpoint.

pub fn custom_endpoints(&self) -> Option<&[String]>

Identifies all custom endpoints associated with the cluster.

pub fn multi_az(&self) -> Option<bool>

Specifies whether the DB cluster has instances in multiple Availability Zones.

pub fn engine(&self) -> Option<&str>

The name of the database engine to be used for this DB cluster.

pub fn engine_version(&self) -> Option<&str>

Indicates the database engine version.

pub fn latest_restorable_time(&self) -> Option<&Instant>

Specifies the latest time to which a database can be restored with point-in-time restore.

pub fn port(&self) -> Option<i32>

Specifies the port that the database engine is listening on.

pub fn master_username(&self) -> Option<&str>

Contains the master username for the DB cluster.

pub fn db_cluster_option_group_memberships(
&self
) -> Option<&[DbClusterOptionGroupStatus]>

Provides the list of option group memberships for this DB cluster.

pub fn preferred_backup_window(&self) -> Option<&str>

Specifies the daily time range during which automated backups are created if automated backups are enabled, as determined by the BackupRetentionPeriod.

pub fn preferred_maintenance_window(&self) -> Option<&str>

Specifies the weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

pub fn replication_source_identifier(&self) -> Option<&str>

Contains the identifier of the source DB cluster if this DB cluster is a read replica.

pub fn read_replica_identifiers(&self) -> Option<&[String]>

Contains one or more identifiers of the read replicas associated with this DB cluster.

pub fn db_cluster_members(&self) -> Option<&[DbClusterMember]>

Provides the list of instances that make up the DB cluster.

pub fn vpc_security_groups(&self) -> Option<&[VpcSecurityGroupMembership]>

Provides a list of VPC security groups that the DB cluster belongs to.

pub fn hosted_zone_id(&self) -> Option<&str>

Specifies the ID that Amazon Route 53 assigns when you create a hosted zone.

pub fn storage_encrypted(&self) -> bool

Specifies whether the DB cluster is encrypted.

pub fn kms_key_id(&self) -> Option<&str>

If StorageEncrypted is enabled, the Amazon Web Services KMS key identifier for the encrypted DB cluster.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the KMS key.

pub fn db_cluster_resource_id(&self) -> Option<&str>

The Amazon Web Services Region-unique, immutable identifier for the DB cluster. This identifier is found in Amazon Web Services CloudTrail log entries whenever the KMS key for the DB cluster is accessed.

pub fn db_cluster_arn(&self) -> Option<&str>

The Amazon Resource Name (ARN) for the DB cluster.

pub fn associated_roles(&self) -> Option<&[DbClusterRole]>

Provides a list of the Amazon Web Services Identity and Access Management (IAM) roles that are associated with the DB cluster. IAM roles that are associated with a DB cluster grant permission for the DB cluster to access other Amazon Web Services on your behalf.

pub fn iam_database_authentication_enabled(&self) -> Option<bool>

A value that indicates whether the mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts is enabled.

pub fn clone_group_id(&self) -> Option<&str>

Identifies the clone group to which the DB cluster is associated.

pub fn cluster_create_time(&self) -> Option<&Instant>

Specifies the time when the DB cluster was created, in Universal Coordinated Time (UTC).

pub fn earliest_backtrack_time(&self) -> Option<&Instant>

The earliest time to which a DB cluster can be backtracked.

pub fn backtrack_window(&self) -> Option<i64>

The target backtrack window, in seconds. If this value is set to 0, backtracking is disabled for the DB cluster. Otherwise, backtracking is enabled.

pub fn backtrack_consumed_change_records(&self) -> Option<i64>

The number of change records stored for Backtrack.

pub fn enabled_cloudwatch_logs_exports(&self) -> Option<&[String]>

A list of log types that this DB cluster is configured to export to CloudWatch Logs.

Log types vary by DB engine. For information about the log types for each DB engine, see Amazon RDS Database Log Files in the Amazon Aurora User Guide.

pub fn capacity(&self) -> Option<i32>

The current capacity of an Aurora Serverless DB cluster. The capacity is 0 (zero) when the cluster is paused.

For more information about Aurora Serverless, see Using Amazon Aurora Serverless in the Amazon Aurora User Guide.

pub fn engine_mode(&self) -> Option<&str>

The DB engine mode of the DB cluster, either provisioned, serverless, parallelquery, global, or multimaster.

For more information, see CreateDBCluster.

pub fn scaling_configuration_info(&self) -> Option<&ScalingConfigurationInfo>

Shows the scaling configuration for an Aurora DB cluster in serverless DB engine mode.

For more information, see Using Amazon Aurora Serverless in the Amazon Aurora User Guide.

pub fn deletion_protection(&self) -> Option<bool>

Indicates if the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled.

pub fn http_endpoint_enabled(&self) -> Option<bool>

A value that indicates whether the HTTP endpoint for an Aurora Serverless DB cluster is enabled.

When enabled, the HTTP endpoint provides a connectionless web service API for running SQL queries on the Aurora Serverless DB cluster. You can also query your database from inside the RDS console with the query editor.

For more information, see Using the Data API for Aurora Serverless in the Amazon Aurora User Guide.

pub fn activity_stream_mode(&self) -> Option<&ActivityStreamMode>

The mode of the database activity stream. Database events such as a change or access generate an activity stream event. The database session can handle these events either synchronously or asynchronously.

pub fn activity_stream_status(&self) -> Option<&ActivityStreamStatus>

The status of the database activity stream.

pub fn activity_stream_kms_key_id(&self) -> Option<&str>

The Amazon Web Services KMS key identifier used for encrypting messages in the database activity stream.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the KMS key.

pub fn activity_stream_kinesis_stream_name(&self) -> Option<&str>

The name of the Amazon Kinesis data stream used for the database activity stream.

pub fn copy_tags_to_snapshot(&self) -> Option<bool>

Specifies whether tags are copied from the DB cluster to snapshots of the DB cluster.

pub fn cross_account_clone(&self) -> Option<bool>

Specifies whether the DB cluster is a clone of a DB cluster owned by a different Amazon Web Services account.

pub fn domain_memberships(&self) -> Option<&[DomainMembership]>

The Active Directory Domain membership records associated with the DB cluster.

pub fn tag_list(&self) -> Option<&[Tag]>

A list of tags. For more information, see Tagging Amazon RDS Resources in the Amazon RDS User Guide.

pub fn global_write_forwarding_status(&self) -> Option<&WriteForwardingStatus>

Specifies whether a secondary cluster in an Aurora global database has write forwarding enabled, not enabled, or is in the process of enabling it.

pub fn global_write_forwarding_requested(&self) -> Option<bool>

Specifies whether you have requested to enable write forwarding for a secondary cluster in an Aurora global database. Because write forwarding takes time to enable, check the value of GlobalWriteForwardingStatus to confirm that the request has completed before using the write forwarding feature for this cluster.

pub fn pending_modified_values(&self) -> Option<&ClusterPendingModifiedValues>

A value that specifies that changes to the DB cluster are pending. This element is only included when changes are pending. Specific changes are identified by subelements.

impl DbCluster

pub fn builder() -> Builder

Creates a new builder-style object to manufacture DbCluster